Data set with 500 isolates - WHONET example

This example data set has the exact same structure as an export file from WHONET. Such files can be used with this package, as this example data set shows. The data itself was based on our example_isolates data set.

WHONET

Format

A data.frame with 500 observations and 53 variables:

Identification number

ID of the sample

Specimen number

ID of the specimen

Organism

Name of the microorganism. Before analysis, you should transform this to a valid microbial class, using as.mo.

Country

Country of origin

Laboratory

Name of laboratory

Last name

Last name of patient

First name

Initial of patient

Sex

Gender of patient

Age

Age of patient

Age category

Age group, can also be looked up using age_groups

Date of admission

Date of hospital admission

Specimen date

Date when specimen was received at laboratory

Specimen type

Specimen type or group

Specimen type (Numeric)

Translation of "Specimen type"

Reason

Reason of request with Differential Diagnosis

Isolate number

ID of isolate

Organism type

Type of microorganism, can also be looked up using mo_type

Serotype

Serotype of microorganism

Beta-lactamase

Microorganism produces beta-lactamase?

ESBL

Microorganism produces extended spectrum beta-lactamase?

Carbapenemase

Microorganism produces carbapenemase?

MRSA screening test

Microorganism is possible MRSA?

Inducible clindamycin resistance

Clindamycin can be induced?

Comment

Other comments

Date of data entry

Date this data was entered in WHONET

AMP_ND10:CIP_EE

27 different antibiotics. You can lookup the abbreviatons in the antibiotics data set, or use e.g. ab_name("AMP") to get the official name immediately. Before analysis, you should transform this to a valid antibiotic class, using as.rsi.

Read more on our website!

On our website https://msberends.gitlab.io/AMR you can find a tutorial about how to conduct AMR analysis, the complete documentation of all functions (which reads a lot easier than here in R) and an example analysis using WHONET data.