Property of an antibiotic

Use these functions to return a specific property of an antibiotic from the antibiotics data set, based on their ATC code. Get such a code with as.atc.

ab_property(x, property = "official")

ab_atc(x)

ab_official(x, language = NULL)

ab_name(x, language = NULL)

ab_trivial_nl(x)

ab_certe(x)

ab_umcg(x)

ab_tradenames(x)

Arguments

x
a (vector of a) valid atc code or any text that can be coerced to a valid atc with as.atc

property
one of the column names of one of the antibiotics data set, like "atc" and "official"

language
language of the returned text, defaults to English ("en") and can be set with getOption("AMR_locale"). Either one of "en" (English) or "nl" (Dutch).

Value

A vector of values. In case of ab_tradenames, if x is of length one, a vector will be returned. Otherwise a list, with x as names.

See also

Examples

ab_atc("amcl")         # J01CR02

#> [1] "J01CR02"

ab_name("amcl")        # Amoxicillin and beta-lactamase inhibitor

#> [1] "Amoxicillin and beta-lactamase inhibitor"

ab_name("amcl", "nl")  # Amoxicilline met enzymremmer

#> [1] "Amoxicilline met enzymremmer"

ab_trivial_nl("amcl")  # Amoxicilline/clavulaanzuur

#> [1] "Amoxicilline/clavulaanzuur"

ab_certe("amcl")       # amcl

#> [1] "amcl"

ab_umcg("amcl")        # AMCL

#> [1] "AMCL"