Name of an antibiotic

Convert antibiotic codes to a (trivial) antibiotic name or ATC code, or vice versa. This uses the data from antibiotics.

abname(abcode, from = c("guess", "atc", "certe", "umcg"),
  to = "official", textbetween = " + ", tolower = FALSE)

Arguments

abcode
a code or name, like "AMOX", "AMCL" or "J01CA04"

from, to
type to transform from and to. See antibiotics for its column names. WIth from = "guess" the from will be guessed from "atc", "certe" and "umcg". When using to = "atc", the ATC code will be searched using as.atc.

textbetween
text to put between multiple returned texts

tolower
return output as lower case with function tolower.

Source

antibiotics

Details

The ab_property functions are faster and more concise, but do not support concatenated strings, like abname("AMCL+GENT".

Read more on our website!

On our website https://msberends.gitlab.io/AMR you can find a omprehensive tutorial about how to conduct AMR analysis and find the complete documentation of all functions, which reads a lot easier than in R.

Examples

# NOT RUN {
abname("AMCL")
# "Amoxicillin and beta-lactamase inhibitor"

# It is quite flexible at default (having `from = "guess"`)
abname(c("amox", "J01CA04", "Trimox", "dispermox", "Amoxil"))
# "Amoxicillin" "Amoxicillin" "Amoxicillin" "Amoxicillin" "Amoxicillin"

# Multiple antibiotics can be combined with "+".
# The second antibiotic will be set to lower case when `tolower` was not set:
abname("AMCL+GENT", textbetween = "/")
# "amoxicillin and enzyme inhibitor/gentamicin"

abname(c("AMCL", "GENT"))
# "Amoxicillin and beta-lactamase inhibitor" "Gentamicin"

abname("AMCL", to = "trivial_nl")
# "Amoxicilline/clavulaanzuur"

abname("AMCL", to = "atc")
# "J01CR02"

# specific codes for University Medical Center Groningen (UMCG):
abname("J01CR02", from = "atc", to = "umcg")
# "AMCL"

# specific codes for Certe:
abname("J01CR02", from = "atc", to = "certe")
# "amcl"
# }