Data set with 423 antibiotics

A data set containing all antibiotics with a J0 code and some other antimicrobial agents, with their DDDs. Except for trade names and abbreviations, all properties were downloaded from the WHO, see Source.

antibiotics

Format

A data.frame with 423 observations and 18 variables:

atc

ATC code, like J01CR02

certe

Certe code, like amcl

umcg

UMCG code, like AMCL

abbr

Abbreviation as used by many countries, used internally by as.atc

official

Official name by the WHO, like "Amoxicillin and beta-lactamase inhibitor"

official_nl

Official name in the Netherlands, like "Amoxicilline met enzymremmer"

trivial_nl

Trivial name in Dutch, like "Amoxicilline/clavulaanzuur"

trade_name

Trade name as used by many countries (a total of 294), used internally by as.atc

oral_ddd

Defined Daily Dose (DDD), oral treatment

oral_units

Units of ddd_units

iv_ddd

Defined Daily Dose (DDD), parenteral treatment

iv_units

Units of iv_ddd

atc_group1

ATC group, like "Macrolides, lincosamides and streptogramins"

atc_group2

Subgroup of atc_group1, like "Macrolides"

atc_group1_nl

ATC group in Dutch, like "Macroliden, lincosamiden en streptograminen"

atc_group2_nl

Subgroup of atc_group1 in Dutch, like "Macroliden"

useful_gramnegative

FALSE if not useful according to EUCAST, NA otherwise (see Source)

useful_grampositive

FALSE if not useful according to EUCAST, NA otherwise (see Source)

Source

- World Health Organization: https://www.whocc.no/atc_ddd_index/
- EUCAST - Expert rules intrinsic exceptional V3.1
- MOLIS (LIS of Certe): https://www.certe.nl
- GLIMS (LIS of UMCG): https://www.umcg.nl

See also