Get ATC properties from WHOCC website

Gets data from the WHO to determine properties of an ATC (e.g. an antibiotic) like name, defined daily dose (DDD) or standard unit.

This function requires an internet connection.

atc_online_property(
  atc_code,
  property,
  administration = "O",
  url = "https://www.whocc.no/atc_ddd_index/?code=%s&showdescription=no"
)

atc_online_groups(atc_code, ...)

atc_online_ddd(atc_code, ...)

Arguments

atc_code
a character or character vector with ATC code(s) of antibiotic(s)

property
property of an ATC code. Valid values are "ATC", "Name", "DDD", "U" ("unit"), "Adm.R", "Note" and groups. For this last option, all hierarchical groups of an ATC code will be returned, see Examples.

administration
type of administration when using property = "Adm.R", see Details

url
url of website of the WHO. The sign %s can be used as a placeholder for ATC codes.

...
parameters to pass on to atc_property

Source

https://www.whocc.no/atc_ddd_alterations__cumulative/ddd_alterations/abbrevations/

Details

Options for parameter administration:

"Implant" = Implant

"Inhal" = Inhalation

"Instill" = Instillation

"N" = nasal

"O" = oral

"P" = parenteral

"R" = rectal

"SL" = sublingual/buccal

"TD" = transdermal

"V" = vaginal

Abbreviations of return values when using property = "U" (unit):

"g" = gram

"mg" = milligram

`"mcg"`` = microgram

"U" = unit

"TU" = thousand units

"MU" = million units

"mmol" = millimole

"ml" = milliliter (e.g. eyedrops)

Questioning lifecycle

The lifecycle of this function is questioning. We are no longer convinced that this function is the optimal approach (but we do not know yet what a better approach would be), or whether this function should be in our AMR package at all.

Read more on our website!

On our website https://msberends.gitlab.io/AMR you can find a comprehensive tutorial about how to conduct AMR analysis, the complete documentation of all functions (which reads a lot easier than here in R) and an example analysis using WHONET data.

Examples

# \donttest{
# oral DDD (Defined Daily Dose) of amoxicillin
atc_online_property("J01CA04", "DDD", "O")
# parenteral DDD (Defined Daily Dose) of amoxicillin
atc_online_property("J01CA04", "DDD", "P")

atc_online_property("J01CA04", property = "groups") # search hierarchical groups of amoxicillin
# [1] "ANTIINFECTIVES FOR SYSTEMIC USE"
# [2] "ANTIBACTERIALS FOR SYSTEMIC USE"
# [3] "BETA-LACTAM ANTIBACTERIALS, PENICILLINS"
# [4] "Penicillins with extended spectrum"
# }