Get Properties of an Antiviral Drug

Use these functions to return a specific property of an antiviral drug from the antivirals data set. All input values will be evaluated internally with as.av().

Usage

av_name(x, language = get_AMR_locale(), tolower = FALSE, ...)

av_cid(x, ...)

av_synonyms(x, ...)

av_tradenames(x, ...)

av_group(x, language = get_AMR_locale(), ...)

av_atc(x, ...)

av_loinc(x, ...)

av_ddd(x, administration = "oral", ...)

av_ddd_units(x, administration = "oral", ...)

av_info(x, language = get_AMR_locale(), ...)

av_url(x, open = FALSE, ...)

av_property(x, property = "name", language = get_AMR_locale(), ...)

Arguments

x

any (vector of) text that can be coerced to a valid antiviral drug code with as.av()

language

language of the returned text, defaults to system language (see get_AMR_locale()) and can also be set with getOption("AMR_locale"). Use language = NULL or language = "" to prevent translation.

tolower

a logical to indicate whether the first character of every output should be transformed to a lower case character.

...

other arguments passed on to as.av()

administration

way of administration, either "oral" or "iv"

open

browse the URL using utils::browseURL()

property

one of the column names of one of the antivirals data set: vector_or(colnames(antivirals), sort = FALSE).

Value

An integer in case of av_cid()

A named list in case of av_info() and multiple av_atc()/av_synonyms()/av_tradenames()

A double in case of av_ddd()

A character in all other cases

Details

All output will be translated where possible.

The function av_url() will return the direct URL to the official WHO website. A warning will be returned if the required ATC code is not available.

Source

World Health Organization (WHO) Collaborating Centre for Drug Statistics Methodology: https://www.whocc.no/atc_ddd_index/

European Commission Public Health PHARMACEUTICALS - COMMUNITY REGISTER: https://ec.europa.eu/health/documents/community-register/html/reg_hum_atc.htm

Reference Data Publicly Available

All data sets in this AMR package (about microorganisms, antibiotics, R/SI interpretation, EUCAST rules, etc.) are publicly and freely available for download in the following formats: R, MS Excel, Apache Feather, Apache Parquet, SPSS, SAS, and Stata. We also provide tab-separated plain text files that are machine-readable and suitable for input in any software program, such as laboratory information systems. Please visit our website for the download links. The actual files are of course available on our GitHub repository.

See also

Examples

# all properties:
av_name("ACI")
#> [1] "Aciclovir"
av_atc("ACI")
#> [1] "J05AB01"
av_cid("ACI")
#> [1] 135398513
av_synonyms("ACI")
#>  [1] "acicloftal"        "aciclovier"        "aciclovirum"      
#>  [4] "activir"           "acyclofoam"        "acycloguanosine"  
#>  [7] "acyclovir"         "acyclovir lauriad" "avaclyr"          
#> [10] "cargosil"          "cyclovir"          "genvir"           
#> [13] "gerpevir"          "hascovir"          "maynar"           
#> [16] "novirus"           "poviral"           "sitavig"          
#> [19] "sitavir"           "vipral"            "viropump"         
#> [22] "virorax"           "zovirax"           "zyclir"           
av_tradenames("ACI")
#>  [1] "acicloftal"        "aciclovier"        "aciclovirum"      
#>  [4] "activir"           "acyclofoam"        "acycloguanosine"  
#>  [7] "acyclovir"         "acyclovir lauriad" "avaclyr"          
#> [10] "cargosil"          "cyclovir"          "genvir"           
#> [13] "gerpevir"          "hascovir"          "maynar"           
#> [16] "novirus"           "poviral"           "sitavig"          
#> [19] "sitavir"           "vipral"            "viropump"         
#> [22] "virorax"           "zovirax"           "zyclir"           
av_group("ACI")
#> [1] "Nucleosides and nucleotides excl. reverse transcriptase inhibitors"
av_url("ACI")
#>                                                             Aciclovir 
#> "https://www.whocc.no/atc_ddd_index/?code=J05AB01&showdescription=no" 

# smart lowercase tranformation
av_name(x = c("ACI", "VALA"))
#> [1] "Aciclovir"    "Valaciclovir"
av_name(x = c("ACI", "VALA"), tolower = TRUE)
#> [1] "aciclovir"    "valaciclovir"

# defined daily doses (DDD)
av_ddd("ACI", "oral")
#> [1] 4
av_ddd_units("ACI", "oral")
#> [1] "g"
av_ddd("ACI", "iv")
#> [1] 4
av_ddd_units("ACI", "iv")
#> [1] "g"

av_info("ACI") # all properties as a list
#> $av
#> [1] "ACI"
#> 
#> $cid
#> [1] 135398513
#> 
#> $name
#> [1] "Aciclovir"
#> 
#> $group
#> [1] "Nucleosides and nucleotides excl. reverse transcriptase inhibitors"
#> 
#> $atc
#> [1] "J05AB01"
#> 
#> $tradenames
#>  [1] "acicloftal"        "aciclovier"        "aciclovirum"      
#>  [4] "activir"           "acyclofoam"        "acycloguanosine"  
#>  [7] "acyclovir"         "acyclovir lauriad" "avaclyr"          
#> [10] "cargosil"          "cyclovir"          "genvir"           
#> [13] "gerpevir"          "hascovir"          "maynar"           
#> [16] "novirus"           "poviral"           "sitavig"          
#> [19] "sitavir"           "vipral"            "viropump"         
#> [22] "virorax"           "zovirax"           "zyclir"           
#> 
#> $loinc
#> [1] ""
#> 
#> $ddd
#> $ddd$oral
#> $ddd$oral$amount
#> [1] 4
#> 
#> $ddd$oral$units
#> [1] "g"
#> 
#> 
#> $ddd$iv
#> $ddd$iv$amount
#> [1] 4
#> 
#> $ddd$iv$units
#> [1] "g"
#> 
#> 
#> 

# all av_* functions use as.av() internally, so you can go from 'any' to 'any':
av_atc("ACI")
#> [1] "J05AB01"
av_group("J05AB01")
#> [1] "Nucleosides and nucleotides excl. reverse transcriptase inhibitors"
av_loinc("abacavir")
#> [1] "29113-8" "78772-1" "78773-9" "79134-3" "80118-3"
av_name("29113-8")
#> [1] "Abacavir"
av_name(135398513)
#> [1] "Aciclovir"
av_name("J05AB01")
#> [1] "Aciclovir"