Data Set with Clinical Breakpoints for SIR Interpretation

Data set containing clinical breakpoints to interpret MIC and disk diffusion to SIR values, according to international guidelines. Currently implemented guidelines are EUCAST (2011-2023) and CLSI (2011-2023). Use as.sir() to transform MICs or disks measurements to SIR values.

Usage

clinical_breakpoints

Format

A tibble with 29 747 observations and 12 variables:

guideline
Name of the guideline

type
Breakpoint type, either "ECOFF", "animal", or "human"

method
Testing method, either "DISK" or "MIC"

site
Body site for which the breakpoint must be applied, e.g. "Oral" or "Respiratory"

mo
Microbial ID, see as.mo()

rank_index
Taxonomic rank index of mo from 1 (subspecies/infraspecies) to 5 (unknown microorganism)

ab
Antibiotic code as used by this package, EARS-Net and WHONET, see as.ab()

ref_tbl
Info about where the guideline rule can be found

disk_dose
Dose of the used disk diffusion method

breakpoint_S
Lowest MIC value or highest number of millimetres that leads to "S"

breakpoint_R
Highest MIC value or lowest number of millimetres that leads to "R"

uti
A logical value (TRUE/FALSE) to indicate whether the rule applies to a urinary tract infection (UTI)

Details

Different types of breakpoints

Supported types of breakpoints are ECOFF, animal, and human. ECOFF (Epidemiological cut-off) values are used in antimicrobial susceptibility testing to differentiate between wild-type and non-wild-type strains of bacteria or fungi.

The default is "human", which can also be set with the package option AMR_breakpoint_type. Use as.sir(..., breakpoint_type = ...) to interpret raw data using a specific breakpoint type, e.g. as.sir(..., breakpoint_type = "ECOFF") to use ECOFFs.

Imported from WHONET

Clinical breakpoints in this package were validated through and imported from WHONET, a free desktop Windows application developed and supported by the WHO Collaborating Centre for Surveillance of Antimicrobial Resistance. More can be read on their website. The developers of WHONET and this AMR package have been in contact about sharing their work. We highly appreciate their development on the WHONET software.

Response from CLSI and EUCAST

The CEO of CLSI and the chairman of EUCAST have endorsed the work and public use of this AMR package (and consequently the use of their breakpoints) in June 2023, when future development of distributing clinical breakpoints was discussed in a meeting between CLSI, EUCAST, the WHO, and developers of WHONET and the AMR package.

Download

Like all data sets in this package, this data set is publicly available for download in the following formats: R, MS Excel, Apache Feather, Apache Parquet, SPSS, SAS, and Stata. Please visit our website for the download links. The actual files are of course available on our GitHub repository. They allow for machine reading EUCAST and CLSI guidelines, which is almost impossible with the MS Excel and PDF files distributed by EUCAST and CLSI, though initiatives have started to overcome these burdens.

NOTE: this AMR package (and the WHONET software as well) contains internal methods to apply the guidelines, which is rather complex. For example, some breakpoints must be applied on certain species groups (which are in case of this package available through the microorganisms.groups data set). It is important that this is considered when using the breakpoints for own use.

See also

Examples

clinical_breakpoints
#> # A tibble: 29,747 × 12
#>    guideline type  method site  mo            rank_index ab   ref_tbl  disk_dose
#>    <chr>     <chr> <chr>  <chr> <mo>               <dbl> <ab> <chr>    <chr>    
#>  1 EUCAST 2… human DISK   NA    B_ACHRMB_XYLS          2 MEM  A. xylo… 10ug     
#>  2 EUCAST 2… human MIC    NA    B_ACHRMB_XYLS          2 MEM  A. xylo… NA       
#>  3 EUCAST 2… human DISK   NA    B_ACHRMB_XYLS          2 SXT  A. xylo… 1.25ug/2…
#>  4 EUCAST 2… human MIC    NA    B_ACHRMB_XYLS          2 SXT  A. xylo… NA       
#>  5 EUCAST 2… human DISK   NA    B_ACHRMB_XYLS          2 TZP  A. xylo… 30ug/6ug 
#>  6 EUCAST 2… human MIC    NA    B_ACHRMB_XYLS          2 TZP  A. xylo… NA       
#>  7 EUCAST 2… human DISK   NA    B_ACNTB                3 AMK  Acineto… 30ug     
#>  8 EUCAST 2… human DISK   Unco… B_ACNTB                3 AMK  Acineto… 30ug     
#>  9 EUCAST 2… human MIC    NA    B_ACNTB                3 AMK  Acineto… NA       
#> 10 EUCAST 2… human MIC    Unco… B_ACNTB                3 AMK  Acineto… NA       
#> # ℹ 29,737 more rows
#> # ℹ 3 more variables: breakpoint_S <dbl>, breakpoint_R <dbl>, uti <lgl>