Create Custom EUCAST Rules

Create Custom EUCAST Rules

custom_eucast_rules(...)

Arguments

...
rules in formula notation, see Examples

Details

This documentation page will be updated shortly. This function is experimental.

How it works

..

It is also possible to define antibiotic groups instead of single antibiotics. The following groups are allowed (case-insensitive): aminoglycosides, aminopenicillins, betalactams, carbapenems, cephalosporins, cephalosporins_1st, cephalosporins_2nd, cephalosporins_3rd, cephalosporins_except_caz, fluoroquinolones, glycopeptides, glycopeptides_except_lipo, lincosamides, lipoglycopeptides, macrolides, oxazolidinones, penicillins, polymyxins, streptogramins, tetracyclines, tetracyclines_except_tgc and ureidopenicillins.

Experimental Lifecycle

The lifecycle of this function is experimental. An experimental function is in early stages of development. The unlying code might be changing frequently. Experimental functions might be removed without deprecation, so you are generally best off waiting until a function is more mature before you use it in production code. Experimental functions are only available in development versions of this AMR package and will thus not be included in releases that are submitted to CRAN, since such functions have not yet matured enough.

Examples

x <- custom_eucast_rules(AMC == "R" & genus == "Klebsiella" ~ aminopenicillins == "R",
                         AMC == "I" & genus == "Klebsiella" ~ aminopenicillins == "I")
eucast_rules(example_isolates,
             rules = "custom",
             custom_rules = x,
             info = FALSE)
             
# combine rule sets
x2 <- c(x,
        custom_eucast_rules(TZP == "R" ~ carbapenems == "R"))
x2