Data Set with Treatment Dosages as Defined by EUCAST

EUCAST breakpoints used in this package are based on the dosages in this data set. They can be retrieved with eucast_dosage().

Usage

dosage

Format

A data.frame with 169 observations and 9 variables:

ab
Antibiotic ID as used in this package (such as AMC), using the official EARS-Net (European Antimicrobial Resistance Surveillance Network) codes where available

name
Official name of the antimicrobial agent as used by WHONET/EARS-Net or the WHO

type
Type of the dosage, either "high_dosage", "standard_dosage" or "uncomplicated_uti"

dose
Dose, such as "2 g" or "25 mg/kg"

dose_times
Number of times a dose must be administered

administration
Route of administration, either "im", "iv" or "oral"

notes
Additional dosage notes

original_txt
Original text in the PDF file of EUCAST

eucast_version
Version number of the EUCAST Clinical Breakpoints guideline to which these dosages apply

Details

'EUCAST Clinical Breakpoint Tables' v11.0 (2021) are based on the dosages in this data set.

Reference Data Publicly Available

All reference data sets (about microorganisms, antibiotics, R/SI interpretation, EUCAST rules, etc.) in this AMR package are publicly and freely available. We continually export our data sets to formats for use in R, SPSS, SAS, Stata and Excel. We also supply flat files that are machine-readable and suitable for input in any software program, such as laboratory information systems. Please find all download links on our website, which is automatically updated with every code change.

Examples

head(dosage)
#>    ab        name              type        dose dose_times administration notes
#> 1 AMK    Amikacin   standard_dosage 25-30 mg/kg          1             iv      
#> 2 AMX Amoxicillin       high_dosage         2 g          6             iv      
#> 3 AMX Amoxicillin   standard_dosage         1 g          3             iv      
#> 4 AMX Amoxicillin       high_dosage    0.75-1 g          3           oral      
#> 5 AMX Amoxicillin   standard_dosage       0.5 g          3           oral      
#> 6 AMX Amoxicillin uncomplicated_uti       0.5 g          3           oral      
#>         original_txt eucast_version
#> 1 25-30 mg/kg x 1 iv             11
#> 2         2 g x 6 iv             11
#> 3       1 g x 3-4 iv             11
#> 4  0.75-1 g x 3 oral             11
#> 5     0.5 g x 3 oral             11
#> 6     0.5 g x 3 oral             11