Data set with 2,000 example isolates

A data set containing 2,000 microbial isolates with their full antibiograms. The data set reflects reality and can be used to practice AMR analysis. For examples, please read the tutorial on our website.

example_isolates

Format

A data.frame with 2,000 observations and 49 variables:

date
date of receipt at the laboratory

hospital_id
ID of the hospital, from A to D

ward_icu
logical to determine if ward is an intensive care unit

ward_clinical
logical to determine if ward is a regular clinical ward

ward_outpatient
logical to determine if ward is an outpatient clinic

age
age of the patient

gender
gender of the patient

patient_id
ID of the patient

mo
ID of microorganism created with as.mo(), see also microorganisms

PEN:RIF
40 different antibiotics with class rsi (see as.rsi()); these column names occur in the antibiotics data set and can be translated with ab_name()

Read more on our website!

On our website https://msberends.github.io/AMR you can find a comprehensive tutorial about how to conduct AMR analysis, the complete documentation of all functions (which reads a lot easier than here in R) and an example analysis using WHONET data.