Data Set with Unclean Data

A data set containing 3,000 microbial isolates that are not cleaned up and consequently not ready for AMR data analysis. This data set can be used for practice.

example_isolates_unclean

Format

A data.frame with 3,000 observations and 8 variables:

patient_id
ID of the patient

date
date of receipt at the laboratory

hospital
ID of the hospital, from A to C

bacteria
info about microorganism that can be transformed with as.mo(), see also microorganisms

AMX:GEN
4 different antibiotics that have to be transformed with as.rsi()

Reference Data Publicly Available

All reference data sets (about microorganisms, antibiotics, R/SI interpretation, EUCAST rules, etc.) in this AMR package are publicly and freely available. We continually export our data sets to formats for use in R, SPSS, SAS, Stata and Excel. We also supply flat files that are machine-readable and suitable for input in any software program, such as laboratory information systems. Please find all download links on our website, which is automatically updated with every code change.

Read more on Our Website!

On our website https://msberends.github.io/AMR/ you can find a comprehensive tutorial about how to conduct AMR data analysis, the complete documentation of all functions and an example analysis using WHONET data. As we would like to better understand the backgrounds and needs of our users, please participate in our survey!