Determine (New) Episodes for Patients

These functions determine which items in a vector can be considered (the start of) a new episode, based on the argument episode_days. This can be used to determine clinical episodes for any epidemiological analysis. The get_episode() function returns the index number of the episode per group, while the is_new_episode() function returns values TRUE/FALSE to indicate whether an item in a vector is the start of a new episode.

Usage

get_episode(x, episode_days, ...)

is_new_episode(x, episode_days, ...)

Arguments

x

vector of dates (class Date or POSIXt), will be sorted internally to determine episodes

episode_days

required episode length in days, can also be less than a day or Inf, see Details

...

ignored, only in place to allow future extensions

Value

get_episode(): a double vector

is_new_episode(): a logical vector

Details

Dates are first sorted from old to new. The oldest date will mark the start of the first episode. After this date, the next date will be marked that is at least episode_days days later than the start of the first episode. From that second marked date on, the next date will be marked that is at least episode_days days later than the start of the second episode which will be the start of the third episode, and so on. Before the vector is being returned, the original order will be restored.

The first_isolate() function is a wrapper around the is_new_episode() function, but is more efficient for data sets containing microorganism codes or names and allows for different isolate selection methods.

The dplyr package is not required for these functions to work, but these functions do support variable grouping and work conveniently inside dplyr verbs such as filter(), mutate() and summarise().

See also

Examples

# `example_isolates` is a data set available in the AMR package.
# See ?example_isolates
df <- example_isolates[sample(seq_len(2000), size = 200), ]

get_episode(df$date, episode_days = 60) # indices
#>   [1] 19 20 54 49 35 60 26  3 18  4 43 56 55 48 60 58 47 55 59  3  6 32 27 25 56
#>  [26] 10 24  9  5 60 57 62 15  1 52  7 23  5  2 46  9 14 46 56 18 38 34 20 44 41
#>  [51] 21 45  1 16 28 52  2 12 31 48  7 30  3 62 54 44 36 47 14 54  4 29 13 54 11
#>  [76] 53  8  3 54 14 48 37 57 16 21 36  6 23 61 45  9 46 38 26 42 27  1 44  9 21
#> [101] 10  2 31 54  7 20 61 40 57 35 62  4  1 46  5  8 49 14 37 38 44 52 18 39 54
#> [126] 22 25 20 59 57  6 54 44  3  7 50 26 17 38 26 33 33 52 53  7 54 54 48 59 17
#> [151] 45 30 43 42 19 41  5 17 15 32  7 20 62 45 14  6 13 10  1 59 28 18 14 34 60
#> [176] 54  6 13 23 55 11 19 34 26 51 15 11 22 54 10 22 28  8 26 38 39 28 37 63 56
is_new_episode(df$date, episode_days = 60) # TRUE/FALSE
#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE
#>  [13]  TRUE FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE
#>  [25]  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
#>  [49] FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE  TRUE
#>  [61] FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE
#>  [73]  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE  TRUE  TRUE FALSE
#>  [85]  TRUE  TRUE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE  TRUE
#> [109] FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE  TRUE FALSE FALSE
#> [121] FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [133] FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE  TRUE FALSE  TRUE  TRUE
#> [145] FALSE FALSE FALSE FALSE FALSE  TRUE FALSE  TRUE  TRUE  TRUE FALSE  TRUE
#> [157]  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#> [169]  TRUE  TRUE FALSE  TRUE FALSE FALSE  TRUE  TRUE FALSE FALSE  TRUE FALSE
#> [181] FALSE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE
#> [193] FALSE FALSE  TRUE FALSE  TRUE FALSE  TRUE FALSE

# filter on results from the third 60-day episode only, using base R
df[which(get_episode(df$date, 60) == 3), ]
#> # A tibble: 5 × 46
#>   date       patient   age gender ward     mo           PEN   OXA   FLC   AMX  
#>   <date>     <chr>   <dbl> <chr>  <chr>    <mo>         <rsi> <rsi> <rsi> <rsi>
#> 1 2002-11-18 956065     89 F      Clinical B_ESCHR_COLI R     NA    NA    NA   
#> 2 2002-11-28 705451     56 M      Clinical B_STPHY_CONS R     NA    S     NA   
#> 3 2002-11-27 496896     47 F      ICU      B_STPHY_CONS R     NA    R     R    
#> 4 2002-10-20 F35553     51 M      ICU      B_STPHY_AURS S     NA    S     NA   
#> 5 2002-10-11 871360     78 M      Clinical B_STPHY_EPDR R     NA    S     NA   
#> # … with 36 more variables: AMC <rsi>, AMP <rsi>, TZP <rsi>, CZO <rsi>,
#> #   FEP <rsi>, CXM <rsi>, FOX <rsi>, CTX <rsi>, CAZ <rsi>, CRO <rsi>,
#> #   GEN <rsi>, TOB <rsi>, AMK <rsi>, KAN <rsi>, TMP <rsi>, SXT <rsi>,
#> #   NIT <rsi>, FOS <rsi>, LNZ <rsi>, CIP <rsi>, MFX <rsi>, VAN <rsi>,
#> #   TEC <rsi>, TCY <rsi>, TGC <rsi>, DOX <rsi>, ERY <rsi>, CLI <rsi>,
#> #   AZM <rsi>, IPM <rsi>, MEM <rsi>, MTR <rsi>, CHL <rsi>, COL <rsi>,
#> #   MUP <rsi>, RIF <rsi>

# the functions also work for less than a day, e.g. to include one per hour:
get_episode(c(
  Sys.time(),
  Sys.time() + 60 * 60
),
episode_days = 1 / 24
)
#> [1] 1 2

# \donttest{
if (require("dplyr")) {
  # is_new_episode() can also be used in dplyr verbs to determine patient
  # episodes based on any (combination of) grouping variables:
  df %>%
    mutate(condition = sample(
      x = c("A", "B", "C"),
      size = 200,
      replace = TRUE
    )) %>%
    group_by(condition) %>%
    mutate(new_episode = is_new_episode(date, 365)) %>%
    select(patient, date, condition, new_episode)
}
#> # A tibble: 200 × 4
#> # Groups:   condition [3]
#>    patient date       condition new_episode
#>    <chr>   <date>     <chr>     <lgl>      
#>  1 E59875  2006-03-25 C         FALSE      
#>  2 C34072  2006-07-05 C         TRUE       
#>  3 0DBF93  2015-10-12 A         FALSE      
#>  4 2F9056  2014-05-06 A         TRUE       
#>  5 953526  2010-04-23 B         FALSE      
#>  6 976997  2017-03-02 C         FALSE      
#>  7 080086  2007-10-26 B         FALSE      
#>  8 956065  2002-11-18 B         FALSE      
#>  9 612575  2006-01-31 B         FALSE      
#> 10 285137  2002-12-13 C         FALSE      
#> # … with 190 more rows
if (require("dplyr")) {
  df %>%
    group_by(ward, patient) %>%
    transmute(date,
      patient,
      new_index = get_episode(date, 60),
      new_logical = is_new_episode(date, 60)
    )
}
#> # A tibble: 200 × 5
#> # Groups:   ward, patient [178]
#>    ward     date       patient new_index new_logical
#>    <chr>    <date>     <chr>       <dbl> <lgl>      
#>  1 ICU      2006-03-25 E59875          1 TRUE       
#>  2 Clinical 2006-07-05 C34072          1 TRUE       
#>  3 Clinical 2015-10-12 0DBF93          1 TRUE       
#>  4 ICU      2014-05-06 2F9056          1 TRUE       
#>  5 ICU      2010-04-23 953526          1 TRUE       
#>  6 Clinical 2017-03-02 976997          1 TRUE       
#>  7 Clinical 2007-10-26 080086          1 TRUE       
#>  8 Clinical 2002-11-18 956065          1 TRUE       
#>  9 Clinical 2006-01-31 612575          1 TRUE       
#> 10 ICU      2002-12-13 285137          1 TRUE       
#> # … with 190 more rows
if (require("dplyr")) {
  df %>%
    group_by(ward) %>%
    summarise(
      n_patients = n_distinct(patient),
      n_episodes_365 = sum(is_new_episode(date, episode_days = 365)),
      n_episodes_60 = sum(is_new_episode(date, episode_days = 60)),
      n_episodes_30 = sum(is_new_episode(date, episode_days = 30))
    )
}
#> # A tibble: 3 × 5
#>   ward       n_patients n_episodes_365 n_episodes_60 n_episodes_30
#>   <chr>           <int>          <int>         <int>         <int>
#> 1 Clinical          115             13            49            69
#> 2 ICU                55             12            34            39
#> 3 Outpatient          8              7             8             8
if (require("dplyr")) {

  # grouping on patients and microorganisms leads to the same
  # results as first_isolate() when using 'episode-based':
  x <- df %>%
    filter_first_isolate(
      include_unknown = TRUE,
      method = "episode-based"
    )

  y <- df %>%
    group_by(patient, mo) %>%
    filter(is_new_episode(date, 365)) %>%
    ungroup()

  identical(x, y)
}
#> Including isolates from ICU.
#> [1] TRUE
if (require("dplyr")) {

  # but is_new_episode() has a lot more flexibility than first_isolate(),
  # since you can now group on anything that seems relevant:
  df %>%
    group_by(patient, mo, ward) %>%
    mutate(flag_episode = is_new_episode(date, 365)) %>%
    select(group_vars(.), flag_episode)
}
#> # A tibble: 200 × 4
#> # Groups:   patient, mo, ward [190]
#>    patient mo            ward     flag_episode
#>    <chr>   <mo>          <chr>    <lgl>       
#>  1 E59875  B_STPHY_EPDR  ICU      TRUE        
#>  2 C34072  B_STPHY_CONS  Clinical TRUE        
#>  3 0DBF93  B_STPHY_AURS  Clinical TRUE        
#>  4 2F9056  B_HAFNI_ALVE  ICU      TRUE        
#>  5 953526  B_STPHY_CONS  ICU      TRUE        
#>  6 976997  B_STRPT_PYGN  Clinical TRUE        
#>  7 080086  B_STRPT_GRPB  Clinical TRUE        
#>  8 956065  B_ESCHR_COLI  Clinical TRUE        
#>  9 612575  B_ENTRBC_CLOC Clinical TRUE        
#> 10 285137  B_ESCHR_COLI  ICU      TRUE        
#> # … with 190 more rows
# }