Determine (new) episodes for patients

This function determines which items in a vector can be considered (the start of) a new episode, based on the parameter episode_days. This can be used to determine clinical episodes for any epidemiological analysis.

is_new_episode(x, episode_days = 365, ...)

Arguments

x
vector of dates (class Date or POSIXt)

episode_days
length of the required episode in days, defaults to 365. Every element in the input will return TRUE after this number of days has passed since the last included date, independent of calendar years. Please see Details.

...
arguments passed on to as.Date()

Value

a logical vector

Details

Dates are first sorted from old to new. The oldest date will mark the start of the first episode. After this date, the next date will be marked that is at least episode_days days later than the start of the first episode. From that second marked date on, the next date will be marked that is at least episode_days days later than the start of the second episode which will be the start of the third episode, and so on. Before the vector is being returned, the original order will be restored.

The dplyr package is not required for this function to work, but this function works conveniently inside dplyr verbs such as filter(), mutate() and summarise().

Experimental lifecycle

The lifecycle of this function is experimental. An experimental function is in early stages of development. The unlying code might be changing frequently. Experimental functions might be removed without deprecation, so you are generally best off waiting until a function is more mature before you use it in production code. Experimental functions are only available in development versions of this AMR package and will thus not be included in releases that are submitted to CRAN, since such functions have not yet matured enough.

Read more on our website!

On our website https://msberends.github.io/AMR/ you can find a comprehensive tutorial about how to conduct AMR analysis, the complete documentation of all functions and an example analysis using WHONET data. As we would like to better understand the backgrounds and needs of our users, please participate in our survey!

Examples