Kurtosis of the Sample

Kurtosis is a measure of the "tailedness" of the probability distribution of a real-valued random variable. A normal distribution has a kurtosis of 3 and a excess kurtosis of 0.

kurtosis(x, na.rm = FALSE, excess = FALSE)

# S3 method for default
kurtosis(x, na.rm = FALSE, excess = FALSE)

# S3 method for matrix
kurtosis(x, na.rm = FALSE, excess = FALSE)

# S3 method for data.frame
kurtosis(x, na.rm = FALSE, excess = FALSE)

Arguments

x

a vector of values, a matrix or a data.frame

na.rm

a logical to indicate whether NA values should be stripped before the computation proceeds

excess

a logical to indicate whether the excess kurtosis should be returned, defined as the kurtosis minus 3.

Stable Lifecycle

The lifecycle of this function is stable. In a stable function, major changes are unlikely. This means that the unlying code will generally evolve by adding new arguments; removing arguments or changing the meaning of existing arguments will be avoided.

If the unlying code needs breaking changes, they will occur gradually. For example, an argument will be deprecated and first continue to work, but will emit an message informing you of the change. Next, typically after at least one newly released version on CRAN, the message will be transformed to an error.

Read more on Our Website!

On our website https://msberends.github.io/AMR/ you can find a comprehensive tutorial about how to conduct AMR data analysis, the complete documentation of all functions and an example analysis using WHONET data.

See also