Lifecycles of functions in the AMR package

Our functions are categorised using the lifecycle circle of the tidyverse as found on www.tidyverse.org/lifecycle.

This page contains a section for every lifecycle (with text borrowed from the aforementioned tidyverse website), so they can be used in the manual pages of our functions.

Experimental lifecycle

The lifecycle of this function is experimental. An experimental function is in the very early stages of development. The unlying code might be changing frequently as we rapidly iterate and explore variations in search of the best fit. Experimental functions might be removed without deprecation, so you are generally best off waiting until a function is more mature before you use it in production code. Experimental functions will not be included in releases we submit to CRAN, since they have not yet matured enough.

Maturing lifecycle

The lifecycle of this function is maturing. The unlying code of a maturing function has been roughed out, but finer details might still change. We will strive to maintain backward compatibility, but the function needs wider usage and more extensive testing in order to optimise the unlying code.

Stable lifecycle

The lifecycle of this function is stable. In a stable function, we are largely happy with the unlying code, and major changes are unlikely. This means that the unlying code will generally evolve by adding new arguments; we will avoid removing arguments or changing the meaning of existing arguments.

If the unlying code needs breaking changes, they will occur gradually. To begin with, the function or argument will be deprecated; it will continue to work but will emit an message informing you of the change. Next, typically after at least one newly released version on CRAN, the message will be transformed to an error.

Retired lifecycle

The lifecycle of this function is retired. A retired function is no longer under active development, and (if appropiate) a better alternative is available. We will only make the necessary changes to ensure that retired functions remain available. No new arguments will be added, and only the most critical bugs will be fixed.

Archived lifecycle

The lifecycle of this function is archived. The development of an archived function has ended, and it is no longer available in future package versions.

Dormant lifecycle

The lifecycle of this function is dormant. A dormant function is currently not under active development and has not reached a stable phase. We might return to it in the future. As with experimental functions, you are best off waiting until a function is more mature before you use it in production code.

Questioning lifecycle

The lifecycle of this function is questioning. We are no longer convinced that this function is the optimal approach (but we do not know yet what a better approach would be), or whether this function should be in our AMR package at all.