Calculate the Mean AMR Distance

Calculates a normalised mean for antimicrobial resistance between multiple observations, to help to identify similar isolates without comparing antibiograms by hand.

Usage

mean_amr_distance(x, ...)

# S3 method for class 'sir'
mean_amr_distance(x, ..., combine_SI = TRUE)

# S3 method for class 'data.frame'
mean_amr_distance(x, ..., combine_SI = TRUE)

amr_distance_from_row(amr_distance, row)

Arguments

x

a vector of class sir, mic or disk, or a data.frame containing columns of any of these classes

...

variables to select (supports tidyselect language such as column1:column4 and where(is.mic), and can thus also be antimicrobial selectors

combine_SI

a logical to indicate whether all values of S, SDD, and I must be merged into one, so the input only consists of S+I vs. R (susceptible vs. resistant) - the default is TRUE

amr_distance

the outcome of mean_amr_distance()

row

an index, such as a row number

Details

The mean AMR distance is effectively the Z-score; a normalised numeric value to compare AMR test results which can help to identify similar isolates, without comparing antibiograms by hand.

MIC values (see as.mic()) are transformed with log2() first; their distance is thus calculated as (log2(x) - mean(log2(x))) / sd(log2(x)).

SIR values (see as.sir()) are transformed using "S" = 1, "I" = 2, and "R" = 3. If combine_SI is TRUE (default), the "I" will be considered to be 1.

For data sets, the mean AMR distance will be calculated per column, after which the mean per row will be returned, see Examples.

Use amr_distance_from_row() to subtract distances from the distance of one row, see Examples.

Interpretation

Isolates with distances less than 0.01 difference from each other should be considered similar. Differences lower than 0.025 should be considered suspicious.

Examples

sir <- random_sir(10)
sir
#> Class 'sir'
#>  [1] S I S I I I R R I R
mean_amr_distance(sir)
#>  [1] -0.621059 -0.621059 -0.621059 -0.621059 -0.621059 -0.621059  1.449138
#>  [8]  1.449138 -0.621059  1.449138

mic <- random_mic(10)
mic
#> Class 'mic'
#>  [1] 64      16      8       64      256     0.5     128     128     2      
#> [10] <=0.002
mean_amr_distance(mic)
#>  [1]  0.54295251  0.15465479 -0.03949407  0.54295251  0.93125022 -0.81608950
#>  [7]  0.73710137  0.73710137 -0.42779178 -2.36263742
# equal to the Z-score of their log2:
(log2(mic) - mean(log2(mic))) / sd(log2(mic))
#>  [1]  0.54295251  0.15465479 -0.03949407  0.54295251  0.93125022 -0.81608950
#>  [7]  0.73710137  0.73710137 -0.42779178 -2.36263742

disk <- random_disk(10)
disk
#> Class 'disk'
#>  [1] 28 27 30 48  9 45 12 26 42 25
mean_amr_distance(disk)
#>  [1] -0.09277857 -0.17009405  0.06185238  1.45353101 -1.56177268  1.22158457
#>  [7] -1.32982624 -0.24740953  0.98963813 -0.32472501

y <- data.frame(
  id = LETTERS[1:10],
  amox = random_sir(10, ab = "amox", mo = "Escherichia coli"),
  cipr = random_disk(10, ab = "cipr", mo = "Escherichia coli"),
  gent = random_mic(10, ab = "gent", mo = "Escherichia coli"),
  tobr = random_mic(10, ab = "tobr", mo = "Escherichia coli")
)
y
#>    id amox cipr gent   tobr
#> 1   A    I   28    2    0.5
#> 2   B    R   29    1      4
#> 3   C    S   23    4      4
#> 4   D    I   27    4      2
#> 5   E    S   19    1      4
#> 6   F    S   22  0.5 <=0.25
#> 7   G    I   26    1      4
#> 8   H    R   17    2    0.5
#> 9   I    S   19    1      1
#> 10  J    I   30    2      2
mean_amr_distance(y)
#> ℹ Calculating mean AMR distance based on columns "amox", "cipr", "gent",
#>   and "tobr"
#>  [1] -0.06537260  0.82078512  0.42232605  0.47166071 -0.31061057 -1.07205573
#>  [7]  0.06631747 -0.06476102 -0.64271615  0.37442671
y$amr_distance <- mean_amr_distance(y, where(is.mic))
#> Error in .subset(x, j): invalid subscript type 'list'
y[order(y$amr_distance), ]
#> Error in order(y$amr_distance): argument 1 is not a vector

if (require("dplyr")) {
  y %>%
    mutate(
      amr_distance = mean_amr_distance(y),
      check_id_C = amr_distance_from_row(amr_distance, id == "C")
    ) %>%
    arrange(check_id_C)
}
#> ℹ Calculating mean AMR distance based on columns "amox", "cipr", "gent",
#>   and "tobr"
#>    id amox cipr gent   tobr amr_distance check_id_C
#> 1   C    S   23    4      4   0.42232605 0.00000000
#> 2   J    I   30    2      2   0.37442671 0.04789934
#> 3   D    I   27    4      2   0.47166071 0.04933466
#> 4   G    I   26    1      4   0.06631747 0.35600858
#> 5   B    R   29    1      4   0.82078512 0.39845906
#> 6   H    R   17    2    0.5  -0.06476102 0.48708707
#> 7   A    I   28    2    0.5  -0.06537260 0.48769865
#> 8   E    S   19    1      4  -0.31061057 0.73293662
#> 9   I    S   19    1      1  -0.64271615 1.06504220
#> 10  F    S   22  0.5 <=0.25  -1.07205573 1.49438178
if (require("dplyr")) {
  # support for groups
  example_isolates %>%
    filter(mo_genus() == "Enterococcus" & mo_species() != "") %>%
    select(mo, TCY, carbapenems()) %>%
    group_by(mo) %>%
    mutate(dist = mean_amr_distance(.)) %>%
    arrange(mo, dist)
}
#> ℹ Using column 'mo' as input for mo_genus()
#> ℹ Using column 'mo' as input for mo_species()
#> ℹ For carbapenems() using columns 'IPM' (imipenem) and 'MEM' (meropenem)
#> ℹ Calculating mean AMR distance based on columns "TCY", "IPM", and "MEM"
#> # A tibble: 63 × 5
#> # Groups:   mo [4]
#>    mo           TCY   IPM   MEM     dist
#>    <mo>         <sir> <sir> <sir>  <dbl>
#>  1 B_ENTRC_AVIM S     S     NA     0    
#>  2 B_ENTRC_AVIM S     S     NA     0    
#>  3 B_ENTRC_CSSL NA    S     NA    NA    
#>  4 B_ENTRC_FACM S     S     NA    -2.66 
#>  5 B_ENTRC_FACM S     R     R     -0.423
#>  6 B_ENTRC_FACM S     R     R     -0.423
#>  7 B_ENTRC_FACM NA    R     R      0.224
#>  8 B_ENTRC_FACM NA    R     R      0.224
#>  9 B_ENTRC_FACM NA    R     R      0.224
#> 10 B_ENTRC_FACM NA    R     R      0.224
#> # ℹ 53 more rows