Data set with old taxonomic data from ITIS

A data set containing old (previously valid or accepted) taxonomic names according to ITIS. This data set is used internally by as.mo.

microorganisms.old

Format

A data.frame with 2,383 observations and 4 variables:

tsn

Old Taxonomic Serial Number (TSN), as defined by ITIS

name

Old taxonomic name of the microorganism as found in ITIS, see Source

tsn_new

New Taxonomic Serial Number (TSN), as defined by ITIS

ref

Author(s) and year of concerning publication as found in ITIS, see Source

Source

[3] Integrated Taxonomic Information System (ITIS) on-line database, https://www.itis.gov.

ITIS

This package contains the complete microbial taxonomic data (with all nine taxonomic ranks - from kingdom to subspecies) from the publicly available Integrated Taxonomic Information System (ITIS, https://www.itis.gov).

All (sub)species from the taxonomic kingdoms Bacteria, Fungi and Protozoa are included in this package, as well as all previously accepted names known to ITIS. Furthermore, the responsible authors and year of publication are available. This allows users to use authoritative taxonomic information for their data analysis on any microorganism, not only human pathogens. It also helps to quickly determine the Gram stain of bacteria, since all bacteria are classified into subkingdom Negibacteria or Posibacteria.

ITIS is a partnership of U.S., Canadian, and Mexican agencies and taxonomic specialists [3].

See also