Translation table for UMCG

A data set containing all bacteria codes of UMCG MMB. These codes can be joined to data with an ID from microorganisms$mo (using left_join_microorganisms). GLIMS codes can also be translated to valid MOs with guess_mo.

microorganisms.umcg

Format

A data.frame with 1,095 observations and 2 variables:

umcg

Code of microorganism according to UMCG MMB

certe

Code of microorganism according to Certe MMB

See also