Calculate resistance of isolates

These functions can be used to calculate the (co-)resistance of microbial isolates (i.e. percentage of S, SI, I, IR or R). All functions support quasiquotation with pipes, can be used in dplyrs summarise and support grouped variables, see Examples.

portion_R and portion_IR can be used to calculate resistance, portion_S and portion_SI can be used to calculate susceptibility.

portion_R(..., minimum = 30, as_percent = FALSE,
  only_all_tested = FALSE)

portion_IR(..., minimum = 30, as_percent = FALSE,
  only_all_tested = FALSE)

portion_I(..., minimum = 30, as_percent = FALSE,
  only_all_tested = FALSE)

portion_SI(..., minimum = 30, as_percent = FALSE,
  only_all_tested = FALSE)

portion_S(..., minimum = 30, as_percent = FALSE,
  only_all_tested = FALSE)

portion_df(data, translate_ab = "name", language = get_locale(),
  minimum = 30, as_percent = FALSE, combine_SI = TRUE,
  combine_IR = FALSE)

rsi_df(data, translate_ab = "name", language = get_locale(),
  minimum = 30, as_percent = FALSE, combine_SI = TRUE,
  combine_IR = FALSE)

Arguments

...
one or more vectors (or columns) with antibiotic interpretations. They will be transformed internally with as.rsi if needed. Use multiple columns to calculate (the lack of) co-resistance: the probability where one of two drugs have a resistant or susceptible result. See Examples.

minimum
the minimum allowed number of available (tested) isolates. Any isolate count lower than minimum will return NA with a warning. The default number of 30 isolates is advised by the Clinical and Laboratory Standards Institute (CLSI) as best practice, see Source.

as_percent
a logical to indicate whether the output must be returned as a hundred fold with % sign (a character). A value of 0.123456 will then be returned as "12.3%".

only_all_tested
(for combination therapies, i.e. using more than one variable for ...) a logical to indicate that isolates must be tested for all antibiotics, see section Combination therapy below

data
a data.frame containing columns with class rsi (see as.rsi)

translate_ab
a column name of the antibiotics data set to translate the antibiotic abbreviations to, using ab_property

language
language of the returned text, defaults to system language (see get_locale) and can also be set with getOption("AMR_locale"). Use language = NULL or language = "" to prevent translation.

combine_SI
a logical to indicate whether all values of S and I must be merged into one, so the output only consists of S+I vs. R (susceptible vs. resistant). This used to be the parameter combine_IR, but this now follows the redefinition by EUCAST about the interpretion of I (increased exposure) in 2019, see section 'Interpretation of S, I and R' below. Default is TRUE.

combine_IR
a logical to indicate whether all values of I and R must be merged into one, so the output only consists of S vs. I+R (susceptible vs. non-susceptible). This is outdated, see parameter combine_SI.

Source

M39 Analysis and Presentation of Cumulative Antimicrobial Susceptibility Test Data, 4th Edition, 2014, Clinical and Laboratory Standards Institute (CLSI). https://clsi.org/standards/products/microbiology/documents/m39/.

Wickham H. Tidy Data. The Journal of Statistical Software, vol. 59, 2014. http://vita.had.co.nz/papers/tidy-data.html

Value

Double or, when as_percent = TRUE, a character.

Details

Remember that you should filter your table to let it contain only first isolates! This is needed to exclude duplicates and to reduce selection bias. Use first_isolate to determine them in your data set.

These functions are not meant to count isolates, but to calculate the portion of resistance/susceptibility. Use the count functions to count isolates. The function portion_SI() is essentially equal to count_SI() / count_all(). Low counts can infuence the outcome - the portion functions may camouflage this, since they only return the portion (albeit being dependent on the minimum parameter).

The function portion_df takes any variable from data that has an "rsi" class (created with as.rsi) and calculates the portions R, I and S. The resulting tidy data (see Source) data.frame will have three rows (S/I/R) and a column for each group and each variable with class "rsi".

The function rsi_df works exactly like portion_df, but adds the number of isolates.

Combination therapy

When using more than one variable for ... (= combination therapy)), use only_all_tested to only count isolates that are tested for all antibiotics/variables that you test them for. See this example for two antibiotics, Antibiotic A and Antibiotic B, about how portion_SI works to calculate the %SI:

-------------------------------------------------------------------------
                        only_all_tested = FALSE   only_all_tested = TRUE
Antibiotic  Antibiotic  -----------------------   -----------------------
    A           B       include as  include as    include as  include as
                        numerator   denominator   numerator   denominator
----------  ----------  ----------  -----------   ----------  -----------
    S           S           X            X             X            X
    I           S           X            X             X            X
    R           S           X            X             X            X
not tested      S           X            X             -            -
    S           I           X            X             X            X
    I           I           X            X             X            X
    R           I           X            X             X            X
not tested      I           X            X             -            -
    S           R           X            X             X            X
    I           R           X            X             X            X
    R           R           -            X             -            X
not tested      R           -            -             -            -
    S       not tested      X            X             -            -
    I       not tested      X            X             -            -
    R       not tested      -            -             -            -
not tested  not tested      -            -             -            -
-------------------------------------------------------------------------

Please note that, in combination therapies, for only_all_tested = TRUE applies that:

   count_S()  +  count_I()  +  count_R()  == count_all()
  portion_S() + portion_I() + portion_R() == 1

and that, in combination therapies, for only_all_tested = FALSE applies that:

   count_S()  +  count_I()  +  count_R()  >= count_all()
  portion_S() + portion_I() + portion_R() >= 1

Using only_all_tested has no impact when only using one antibiotic as input.

Interpretation of S, I and R

In 2019, EUCAST has decided to change the definitions of susceptibility testing categories S, I and R as shown below (http://www.eucast.org/newsiandr/). Results of several consultations on the new definitions are available on the EUCAST website under "Consultations".

S - Susceptible, standard dosing regimen: A microorganism is categorised as "Susceptible, standard dosing regimen", when there is a high likelihood of therapeutic success using a standard dosing regimen of the agent.

I - Susceptible, increased exposure: A microorganism is categorised as "Susceptible, Increased exposure" when there is a high likelihood of therapeutic success because exposure to the agent is increased by adjusting the dosing regimen or by its concentration at the site of infection.

R - Resistant: A microorganism is categorised as "Resistant" when there is a high likelihood of therapeutic failure even when there is increased exposure.

Exposure is a function of how the mode of administration, dose, dosing interval, infusion time, as well as distribution and excretion of the antimicrobial agent will influence the infecting organism at the site of infection.

This AMR package honours this new insight. Use portion_SI to determine antimicrobial susceptibility and count_SI to count susceptible isolates.

Read more on our website!

On our website https://msberends.gitlab.io/AMR you can find a tutorial about how to conduct AMR analysis, the complete documentation of all functions (which reads a lot easier than here in R) and an example analysis using WHONET data.

See also

Examples

# NOT RUN {
# septic_patients is a data set available in the AMR package. It is true, genuine data.
?septic_patients

# Calculate resistance
portion_R(septic_patients$AMX)
portion_IR(septic_patients$AMX)

# Or susceptibility
portion_S(septic_patients$AMX)
portion_SI(septic_patients$AMX)

# Do the above with pipes:
library(dplyr)
septic_patients %>% portion_R(AMX)
septic_patients %>% portion_IR(AMX)
septic_patients %>% portion_S(AMX)
septic_patients %>% portion_SI(AMX)

septic_patients %>%
  group_by(hospital_id) %>%
  summarise(p = portion_SI(CIP),
            n = n_rsi(CIP)) # n_rsi works like n_distinct in dplyr

septic_patients %>%
  group_by(hospital_id) %>%
  summarise(R = portion_R(CIP, as_percent = TRUE),
            I = portion_I(CIP, as_percent = TRUE),
            S = portion_S(CIP, as_percent = TRUE),
            n1 = count_all(CIP),  # the actual total; sum of all three
            n2 = n_rsi(CIP),      # same - analogous to n_distinct
            total = n())          # NOT the number of tested isolates!

# Calculate co-resistance between amoxicillin/clav acid and gentamicin,
# so we can see that combination therapy does a lot more than mono therapy:
septic_patients %>% portion_SI(AMC)      # %SI = 76.3%
septic_patients %>% count_all(AMC)       #   n = 1879

septic_patients %>% portion_SI(GEN)      # %SI = 75.4%
septic_patients %>% count_all(GEN)       #   n = 1855

septic_patients %>% portion_SI(AMC, GEN) # %SI = 94.1%
septic_patients %>% count_all(AMC, GEN)  #   n = 1939

# See Details on how `only_all_tested` works. Example:
septic_patients %>%
  summarise(numerator = count_SI(AMC, GEN),
            denominator = count_all(AMC, GEN),
            portion = portion_SI(AMC, GEN))
#   numerator denominator portion
#        1764        1936  0.9408
septic_patients %>%
  summarise(numerator = count_SI(AMC, GEN, only_all_tested = TRUE),
            denominator = count_all(AMC, GEN, only_all_tested = TRUE),
            portion = portion_SI(AMC, GEN, only_all_tested = TRUE))
#   numerator denominator portion
#       1687        1798   0.9383

septic_patients %>%
  group_by(hospital_id) %>%
  summarise(cipro_p = portion_SI(CIP, as_percent = TRUE),
            cipro_n = count_all(CIP),
            genta_p = portion_SI(GEN, as_percent = TRUE),
            genta_n = count_all(GEN),
            combination_p = portion_SI(CIP, GEN, as_percent = TRUE),
            combination_n = count_all(CIP, GEN))

# Get portions S/I/R immediately of all rsi columns
septic_patients %>%
  select(AMX, CIP) %>%
  portion_df(translate = FALSE)

# It also supports grouping variables
septic_patients %>%
  select(hospital_id, AMX, CIP) %>%
  group_by(hospital_id) %>%
  portion_df(translate = FALSE)

# }# NOT RUN {
# calculate current empiric combination therapy of Helicobacter gastritis:
my_table %>%
  filter(first_isolate == TRUE,
         genus == "Helicobacter") %>%
  summarise(p = portion_S(AMX, MTR),  # amoxicillin with metronidazole
            n = count_all(AMX, MTR))
# }