Calculate microbial resistance

These functions can be used to calculate the (co-)resistance or susceptibility of microbial isolates (i.e. percentage of S, SI, I, IR or R). All functions support quasiquotation with pipes, can be used in dplyrs summarise and support grouped variables, see Examples.

resistance should be used to calculate resistance, susceptibility should be used to calculate susceptibility.

resistance(..., minimum = 30, as_percent = FALSE,
  only_all_tested = FALSE)

susceptibility(..., minimum = 30, as_percent = FALSE,
  only_all_tested = FALSE)

proportion_R(..., minimum = 30, as_percent = FALSE,
  only_all_tested = FALSE)

proportion_IR(..., minimum = 30, as_percent = FALSE,
  only_all_tested = FALSE)

proportion_I(..., minimum = 30, as_percent = FALSE,
  only_all_tested = FALSE)

proportion_SI(..., minimum = 30, as_percent = FALSE,
  only_all_tested = FALSE)

proportion_S(..., minimum = 30, as_percent = FALSE,
  only_all_tested = FALSE)

proportion_df(data, translate_ab = "name", language = get_locale(),
  minimum = 30, as_percent = FALSE, combine_SI = TRUE,
  combine_IR = FALSE)

rsi_df(data, translate_ab = "name", language = get_locale(),
  minimum = 30, as_percent = FALSE, combine_SI = TRUE,
  combine_IR = FALSE)

Arguments

...
one or more vectors (or columns) with antibiotic interpretations. They will be transformed internally with as.rsi if needed. Use multiple columns to calculate (the lack of) co-resistance: the probability where one of two drugs have a resistant or susceptible result. See Examples.

minimum
the minimum allowed number of available (tested) isolates. Any isolate count lower than minimum will return NA with a warning. The default number of 30 isolates is advised by the Clinical and Laboratory Standards Institute (CLSI) as best practice, see Source.

as_percent
a logical to indicate whether the output must be returned as a hundred fold with % sign (a character). A value of 0.123456 will then be returned as "12.3%".

only_all_tested
(for combination therapies, i.e. using more than one variable for ...) a logical to indicate that isolates must be tested for all antibiotics, see section Combination therapy below

data
a data.frame containing columns with class rsi (see as.rsi)

translate_ab
a column name of the antibiotics data set to translate the antibiotic abbreviations to, using ab_property

language
language of the returned text, defaults to system language (see get_locale) and can also be set with getOption("AMR_locale"). Use language = NULL or language = "" to prevent translation.

combine_SI
a logical to indicate whether all values of S and I must be merged into one, so the output only consists of S+I vs. R (susceptible vs. resistant). This used to be the parameter combine_IR, but this now follows the redefinition by EUCAST about the interpretion of I (increased exposure) in 2019, see section 'Interpretation of S, I and R' below. Default is TRUE.

combine_IR
a logical to indicate whether all values of I and R must be merged into one, so the output only consists of S vs. I+R (susceptible vs. non-susceptible). This is outdated, see parameter combine_SI.

Source

M39 Analysis and Presentation of Cumulative Antimicrobial Susceptibility Test Data, 4th Edition, 2014, Clinical and Laboratory Standards Institute (CLSI). https://clsi.org/standards/products/microbiology/documents/m39/.

Value

Double or, when as_percent = TRUE, a character.

Details

The function resistance is equal to the function proportion_R. The function susceptibility is equal to the function proportion_SI.

Remember that you should filter your table to let it contain only first isolates! This is needed to exclude duplicates and to reduce selection bias. Use first_isolate to determine them in your data set.

These functions are not meant to count isolates, but to calculate the proportion of resistance/susceptibility. Use the count functions to count isolates. The function susceptibility() is essentially equal to count_susceptible() / count_all(). Low counts can infuence the outcome - the proportion functions may camouflage this, since they only return the proportion (albeit being dependent on the minimum parameter).

The function proportion_df takes any variable from data that has an "rsi" class (created with as.rsi) and calculates the proportions R, I and S. The function rsi_df works exactly like proportion_df, but adds the number of isolates.

Combination therapy

When using more than one variable for ... (= combination therapy)), use only_all_tested to only count isolates that are tested for all antibiotics/variables that you test them for. See this example for two antibiotics, Antibiotic A and Antibiotic B, about how susceptibility works to calculate the %SI:

--------------------------------------------------------------------
                    only_all_tested = FALSE  only_all_tested = TRUE
                    -----------------------  -----------------------
 Drug A    Drug B   include as  include as   include as  include as
                    numerator   denominator  numerator   denominator
--------  --------  ----------  -----------  ----------  -----------
 S or I    S or I       X            X            X            X
   R       S or I       X            X            X            X
  <NA>     S or I       X            X            -            -
 S or I      R          X            X            X            X
   R         R          -            X            -            X
  <NA>       R          -            -            -            -
 S or I     <NA>        X            X            -            -
   R        <NA>        -            -            -            -
  <NA>      <NA>        -            -            -            -
--------------------------------------------------------------------

Please note that, in combination therapies, for only_all_tested = TRUE applies that:

    count_S()    +   count_I()    +   count_R()    = count_all()
  proportion_S() + proportion_I() + proportion_R() = 1

and that, in combination therapies, for only_all_tested = FALSE applies that:

    count_S()    +   count_I()    +   count_R()    >= count_all()
  proportion_S() + proportion_I() + proportion_R() >= 1

Using only_all_tested has no impact when only using one antibiotic as input.

Interpretation of S, I and R

In 2019, the European Committee on Antimicrobial Susceptibility Testing (EUCAST) has decided to change the definitions of susceptibility testing categories S, I and R as shown below (http://www.eucast.org/newsiandr/). Results of several consultations on the new definitions are available on the EUCAST website under "Consultations".

S - Susceptible, standard dosing regimen: A microorganism is categorised as "Susceptible, standard dosing regimen", when there is a high likelihood of therapeutic success using a standard dosing regimen of the agent.

I - Susceptible, increased exposure: A microorganism is categorised as "Susceptible, Increased exposure" when there is a high likelihood of therapeutic success because exposure to the agent is increased by adjusting the dosing regimen or by its concentration at the site of infection.

R - Resistant: A microorganism is categorised as "Resistant" when there is a high likelihood of therapeutic failure even when there is increased exposure.

Exposure is a function of how the mode of administration, dose, dosing interval, infusion time, as well as distribution and excretion of the antimicrobial agent will influence the infecting organism at the site of infection.

This AMR package honours this new insight. Use susceptibility (equal to proportion_SI) to determine antimicrobial susceptibility and count_susceptible (equal to count_SI) to count susceptible isolates.

Read more on our website!

On our website https://msberends.gitlab.io/AMR you can find a tutorial about how to conduct AMR analysis, the complete documentation of all functions (which reads a lot easier than here in R) and an example analysis using WHONET data.

See also

Examples

# example_isolates is a data set available in the AMR package.
?example_isolates

resistance(example_isolates$AMX)     # determines %R
susceptibility(example_isolates$AMX) # determines %S+I

# be more specific
proportion_S(example_isolates$AMX)
proportion_SI(example_isolates$AMX)
proportion_I(example_isolates$AMX)
proportion_IR(example_isolates$AMX)
proportion_R(example_isolates$AMX)

library(dplyr)
example_isolates %>%
  group_by(hospital_id) %>%
  summarise(r = resistance(CIP),
            n = n_rsi(CIP)) # n_rsi works like n_distinct in dplyr, see ?n_rsi

example_isolates %>%
  group_by(hospital_id) %>%
  summarise(R  = resistance(CIP, as_percent = TRUE),
            SI = susceptibility(CIP, as_percent = TRUE),
            n1 = count_all(CIP),  # the actual total; sum of all three
            n2 = n_rsi(CIP),      # same - analogous to n_distinct
            total = n())          # NOT the number of tested isolates!

# Calculate co-resistance between amoxicillin/clav acid and gentamicin,
# so we can see that combination therapy does a lot more than mono therapy:
example_isolates %>% susceptibility(AMC)  # %SI = 76.3%
example_isolates %>% count_all(AMC)       #   n = 1879

example_isolates %>% susceptibility(GEN)  # %SI = 75.4%
example_isolates %>% count_all(GEN)       #   n = 1855

example_isolates %>% susceptibility(AMC, GEN) # %SI = 94.1%
example_isolates %>% count_all(AMC, GEN)      #   n = 1939

# See Details on how `only_all_tested` works. Example:
example_isolates %>%
  summarise(numerator = count_susceptible(AMC, GEN),
            denominator = count_all(AMC, GEN),
            proportion = susceptibility(AMC, GEN))
example_isolates %>%
  summarise(numerator = count_susceptible(AMC, GEN, only_all_tested = TRUE),
            denominator = count_all(AMC, GEN, only_all_tested = TRUE),
            proportion = susceptibility(AMC, GEN, only_all_tested = TRUE))

example_isolates %>%
  group_by(hospital_id) %>%
  summarise(cipro_p = susceptibility(CIP, as_percent = TRUE),
            cipro_n = count_all(CIP),
            genta_p = susceptibility(GEN, as_percent = TRUE),
            genta_n = count_all(GEN),
            combination_p = susceptibility(CIP, GEN, as_percent = TRUE),
            combination_n = count_all(CIP, GEN))

# Get proportions S/I/R immediately of all rsi columns
example_isolates %>%
  select(AMX, CIP) %>%
  proportion_df(translate = FALSE)

# It also supports grouping variables
example_isolates %>%
  select(hospital_id, AMX, CIP) %>%
  group_by(hospital_id) %>%
  proportion_df(translate = FALSE)

if (FALSE) {

# calculate current empiric combination therapy of Helicobacter gastritis:
my_table %>%
  filter(first_isolate == TRUE,
         genus == "Helicobacter") %>%
  summarise(p = susceptibility(AMX, MTR),  # amoxicillin with metronidazole
            n = count_all(AMX, MTR))
}