Random MIC Values/Disk Zones/SIR Generation

These functions can be used for generating random MIC values and disk diffusion diameters, for AMR data analysis practice. By providing a microorganism and antimicrobial drug, the generated results will reflect reality as much as possible.

Usage

random_mic(size = NULL, mo = NULL, ab = NULL, ...)

random_disk(size = NULL, mo = NULL, ab = NULL, ...)

random_sir(size = NULL, prob_SIR = c(0.33, 0.33, 0.33), ...)

Arguments

size

desired size of the returned vector. If used in a data.frame call or dplyr verb, will get the current (group) size if left blank.

mo

any character that can be coerced to a valid microorganism code with as.mo()

ab

any character that can be coerced to a valid antimicrobial drug code with as.ab()

...

ignored, only in place to allow future extensions

prob_SIR

a vector of length 3: the probabilities for "S" (1st value), "I" (2nd value) and "R" (3rd value)

Value

class mic for random_mic() (see as.mic()) and class disk for random_disk() (see as.disk())

Details

The base R function sample() is used for generating values.

Generated values are based on the EUCAST 2023 guideline as implemented in the clinical_breakpoints data set. To create specific generated values per bug or drug, set the mo and/or ab argument.

Examples

random_mic(25)
#> Class 'mic'
#>  [1] 0.025 0.125 2     0.001 16    >=128 2     32    64    2     0.5   0.5  
#> [13] 1     8     0.002 0.25  64    2     32    0.002 0.25  0.025 0.5   16   
#> [25] 16   
random_disk(25)
#> Class 'disk'
#>  [1] 37 43 42 43 44 38 14 25 37  9 10 49 34 17 50 43  9 19 45 42 18 17 43 26 31
random_sir(25)
#> Class 'sir'
#>  [1] S I S I S S I S R S S I S S R I R S R I S I S R S

# \donttest{
# make the random generation more realistic by setting a bug and/or drug:
random_mic(25, "Klebsiella pneumoniae") # range 0.0625-64
#> Class 'mic'
#>  [1] 1       32      2       64      0.5     0.025   32      0.01    64     
#> [10] 8       0.002   0.5     0.01    0.01    <=0.001 0.125   8       16     
#> [19] 0.025   16      4       32      0.01    0.25    0.25   
random_mic(25, "Klebsiella pneumoniae", "meropenem") # range 0.0625-16
#> Class 'mic'
#>  [1] 1    1    8    >=16 4    4    8    1    4    8    1    >=16 2    4    4   
#> [16] >=16 8    8    >=16 2    4    2    4    8    1   
random_mic(25, "Streptococcus pneumoniae", "meropenem") # range 0.0625-4
#> Class 'mic'
#>  [1] 0.25  2     0.5   1     0.125 4     0.25  4     4     0.5   0.25  4    
#> [13] 8     0.25  0.5   0.125 8     4     0.5   0.125 0.125 2     0.25  0.5  
#> [25] 0.5  

random_disk(25, "Klebsiella pneumoniae") # range 8-50
#> Class 'disk'
#>  [1] 17 11 43 16 49 47 41 41 23 28 26 35 44 19 50 43 36 20 45 46 14 50 40  8 17
random_disk(25, "Klebsiella pneumoniae", "ampicillin") # range 11-17
#> Class 'disk'
#>  [1] 15 14 14 13 14 11 11 17 16 15 13 14 11 16 11 13 12 13 17 16 17 12 17 11 12
random_disk(25, "Streptococcus pneumoniae", "ampicillin") # range 12-27
#> Class 'disk'
#>  [1] 26 26 27 21 18 19 22 19 19 15 23 23 24 19 17 19 20 21 19 21 22 23 16 19 23
# }