Calculate resistance of isolates

This function is deprecated. Use the portion functions instead.

rsi(ab1, ab2 = NULL, interpretation = "IR", minimum = 30,
  as_percent = FALSE, ...)

Arguments

ab1, ab2
vector (or column) with antibiotic interpretations. It will be transformed internally with as.rsi if needed.

interpretation
antimicrobial interpretation to check for

minimum
the minimal amount of available isolates. Any number lower than minimum will return NA with a warning. The default number of 30 isolates is advised by the Clinical and Laboratory Standards Institute (CLSI) as best practice, see Source.

as_percent
a logical to indicate whether the output must be returned as a hundred fold with % sign (a character). A value of 0.123456 will then be returned as "12.3%".

...
deprecated parameters to support usage on older versions