Data set for R/SI interpretation

Data set to interpret MIC and disk diffusion to R/SI values. Included guidelines are CLSI (2011-2019) and EUCAST (2011-2020). Use as.rsi() to transform MICs or disks measurements to R/SI values.

rsi_translation

Format

A data.frame with 18,964 observations and 10 variables:

guideline
Name of the guideline

method
Either "MIC" or "DISK"

site
Body site, e.g. "Oral" or "Respiratory"

mo
Microbial ID, see as.mo()

ab
Antibiotic ID, see as.ab()

ref_tbl
Info about where the guideline rule can be found

disk_dose
Dose of the used disk diffusion method

breakpoint_S
Lowest MIC value or highest number of millimetres that leads to "S"

breakpoint_R
Highest MIC value or lowest number of millimetres that leads to "R"

uti
A logical value (TRUE/FALSE) to indicate whether the rule applies to a urinary tract infection (UTI)

Details

The repository of this AMR package contains a file comprising this exact data set: https://gitlab.com/msberends/AMR/blob/master/data-raw/rsi_translation.txt. This file allows for machine reading EUCAST and CLSI guidelines, which is almost impossible with the Excel and PDF files distributed by EUCAST and CLSI. This file is updated automatically.

Read more on our website!

On our website https://msberends.gitlab.io/AMR you can find a comprehensive tutorial about how to conduct AMR analysis, the complete documentation of all functions (which reads a lot easier than here in R) and an example analysis using WHONET data.