Skewness of the sample

Skewness is a measure of the asymmetry of the probability distribution of a real-valued random variable about its mean.

When negative: the left tail is longer; the mass of the distribution is concentrated on the right of the figure. When positive: the right tail is longer; the mass of the distribution is concentrated on the left of the figure.

skewness(x, na.rm = FALSE)

# S3 method for default
skewness(x, na.rm = FALSE)

# S3 method for matrix
skewness(x, na.rm = FALSE)

# S3 method for data.frame
skewness(x, na.rm = FALSE)

Arguments

x
a vector of values, a matrix or a data.frame

na.rm
a logical value indicating whether NA values should be stripped before the computation proceeds.

Questioning lifecycle

The lifecycle of this function is questioning. We are no longer convinced that this function is the optimal approach (but we do not know yet what a better approach would be), or whether this function should be in our AMR package at all.

Read more on our website!

On our website https://msberends.gitlab.io/AMR you can find a comprehensive tutorial about how to conduct AMR analysis, the complete documentation of all functions (which reads a lot easier than here in R) and an example analysis using WHONET data.

See also