The AMR Package

Welcome to the AMR package.

The AMR package is a peer-reviewed, free and open-source R package with zero dependencies to simplify the analysis and prediction of Antimicrobial Resistance (AMR) and to work with microbial and antimicrobial data and properties, by using evidence-based methods. Our aim is to provide a standard for clean and reproducible AMR data analysis, that can therefore empower epidemiological analyses to continuously enable surveillance and treatment evaluation in any setting. We are a team of many different researchers from around the globe to make this a successful and durable project!

This work was published in the Journal of Statistical Software (Volume 104(3); doi:10.18637/jss.v104.i03 ) and formed the basis of two PhD theses (doi:10.33612/diss.177417131 and doi:10.33612/diss.192486375 ).

After installing this package, R knows ~79 000 distinct microbial species (updated June 2024) and all ~620 antimicrobial and antiviral drugs by name and code (including ATC, EARS-Net, ASIARS-Net, PubChem, LOINC and SNOMED CT), and knows all about valid SIR and MIC values. The integral clinical breakpoint guidelines from CLSI 2011-2025 and EUCAST 2011-2025 are included, even with epidemiological cut-off (ECOFF) values. It supports and can read any data format, including WHONET data. This package works on Windows, macOS and Linux with all versions of R since R-3.0 (April 2013). It was designed to work in any setting, including those with very limited resources. It was created for both routine data analysis and academic research at the Faculty of Medical Sciences of the University of Groningen and the University Medical Center Groningen.

The AMR package is available in English, Arabic, Bengali, Chinese, Czech, Danish, Dutch, Finnish, French, German, Greek, Hindi, Indonesian, Italian, Japanese, Korean, Norwegian, Polish, Portuguese, Romanian, Russian, Spanish, Swahili, Swedish, Turkish, Ukrainian, Urdu, and Vietnamese. Antimicrobial drug (group) names and colloquial microorganism names are provided in these languages.

Source

To cite AMR in publications use:

Berends MS, Luz CF, Friedrich AW, Sinha BNM, Albers CJ, Glasner C (2022). "AMR: An R Package for Working with Antimicrobial Resistance Data." Journal of Statistical Software, 104(3), 1-31. doi:10.18637/jss.v104.i03

A BibTeX entry for LaTeX users is:

@Article{,
  title = {{AMR}: An {R} Package for Working with Antimicrobial Resistance Data},
  author = {Matthijs S. Berends and Christian F. Luz and Alexander W. Friedrich and Bhanu N. M. Sinha and Casper J. Albers and Corinna Glasner},
  journal = {Journal of Statistical Software},
  year = {2022},
  volume = {104},
  number = {3},
  pages = {1--31},
  doi = {10.18637/jss.v104.i03},
}

Download Our Reference Data

All reference data sets in the AMR package - including information on microorganisms, antimicrobials, and clinical breakpoints - are freely available for download in multiple formats: R, MS Excel, Apache Feather, Apache Parquet, SPSS, and Stata.

For maximum compatibility, we also provide machine-readable, tab-separated plain text files suitable for use in any software, including laboratory information systems.

Visit our website for direct download links, or explore the actual files in our GitHub repository.

See also

Author

Maintainer: Matthijs S. Berends m.s.berends@umcg.nl (ORCID)

Authors:

Dennis Souverein (ORCID) [contributor]

Erwin E. A. Hassing [contributor]

Other contributors:

Aislinn Cook (ORCID) [contributor]

Andrew P. Norgan (ORCID) [contributor]

Anita Williams (ORCID) [contributor]

Annick Lenglet (ORCID) [contributor]

Anthony Underwood (ORCID) [contributor]

Anton Mymrikov [contributor]

Bart C. Meijer [contributor]

Christian F. Luz (ORCID) [contributor]

Dmytro Mykhailenko [contributor]

Eric H. L. C. M. Hazenberg [contributor]

Gwen Knight (ORCID) [contributor]

Jane Hawkey (ORCID) [contributor]

Jason Stull (ORCID) [contributor]

Javier Sanchez (ORCID) [contributor]

Jonas Salm [contributor]

Judith M. Fonville [contributor]

Kathryn Holt (ORCID) [contributor]

Larisse Bolton (ORCID) [contributor]

Matthew Saab [contributor]

Natacha Couto (ORCID) [contributor]

Peter Dutey-Magni (ORCID) [contributor]

Rogier P. Schade [contributor]

Sofia Ny (ORCID) [contributor]

Alex W. Friedrich (ORCID) [thesis advisor]

Bhanu N. M. Sinha (ORCID) [thesis advisor]

Casper J. Albers (ORCID) [thesis advisor]

Corinna Glasner (ORCID) [thesis advisor]