The AMR Package

Welcome to the AMR package.

The AMR package is a free and open-source R package with zero dependencies to simplify the analysis and prediction of Antimicrobial Resistance (AMR) and to work with microbial and antimicrobial data and properties, by using evidence-based methods. Our aim is to provide a standard for clean and reproducible AMR data analysis, that can therefore empower epidemiological analyses to continuously enable surveillance and treatment evaluation in any setting. Many different researchers from around the globe are continually helping us to make this a successful and durable project!

This work was published in the Journal of Statistical Software (Volume 104(3); doi:10.18637/jss.v104.i03 ) and formed the basis of two PhD theses (doi:10.33612/diss.177417131 and doi:10.33612/diss.192486375 ).

After installing this package, R knows ~71 000 microorganisms (updated June 2024) and all ~600 antibiotic, antimycotic and antiviral drugs by name and code (including ATC, EARS-Net, ASIARS-Net, PubChem, LOINC and SNOMED CT), and knows all about valid SIR and MIC values. The integral clinical breakpoint guidelines from CLSI and EUCAST are included, even with epidemiological cut-off (ECOFF) values. It supports and can read any data format, including WHONET data. This package works on Windows, macOS and Linux with all versions of R since R-3.0 (April 2013). It was designed to work in any setting, including those with very limited resources. It was created for both routine data analysis and academic research at the Faculty of Medical Sciences of the public University of Groningen, in collaboration with non-profit organisations Certe Medical Diagnostics and Advice Foundation and University Medical Center Groningen.

The AMR package is available in English, Chinese, Czech, Danish, Dutch, Finnish, French, German, Greek, Italian, Japanese, Norwegian, Polish, Portuguese, Romanian, Russian, Spanish, Swedish, Turkish, and Ukrainian. Antimicrobial drug (group) names and colloquial microorganism names are provided in these languages.

Source

To cite AMR in publications use:

Berends MS, Luz CF, Friedrich AW, Sinha BNM, Albers CJ, Glasner C (2022). "AMR: An R Package for Working with Antimicrobial Resistance Data." Journal of Statistical Software, 104(3), 1-31. doi:10.18637/jss.v104.i03

A BibTeX entry for LaTeX users is:

@Article{,
  title = {{AMR}: An {R} Package for Working with Antimicrobial Resistance Data},
  author = {Matthijs S. Berends and Christian F. Luz and Alexander W. Friedrich and Bhanu N. M. Sinha and Casper J. Albers and Corinna Glasner},
  journal = {Journal of Statistical Software},
  year = {2022},
  volume = {104},
  number = {3},
  pages = {1--31},
  doi = {10.18637/jss.v104.i03},
}

Reference Data Publicly Available

All data sets in this AMR package (about microorganisms, antibiotics, SIR interpretation, EUCAST rules, etc.) are publicly and freely available for download in the following formats: R, MS Excel, Apache Feather, Apache Parquet, SPSS, SAS, and Stata. We also provide tab-separated plain text files that are machine-readable and suitable for input in any software program, such as laboratory information systems. Please visit our website for the download links. The actual files are of course available on our GitHub repository.

See also

Author

Maintainer: Matthijs S. Berends m.s.berends@umcg.nl (ORCID)

Authors:

Dennis Souverein (ORCID) [contributor]

Erwin E. A. Hassing [contributor]

Other contributors:

Casper J. Albers (ORCID) [thesis advisor]

Peter Dutey-Magni (ORCID) [contributor]

Judith M. Fonville [contributor]

Alex W. Friedrich (ORCID) [thesis advisor]

Corinna Glasner (ORCID) [thesis advisor]

Eric H. L. C. M. Hazenberg [contributor]

Gwen Knight (ORCID) [contributor]

Annick Lenglet (ORCID) [contributor]

Christian F. Luz (ORCID) [contributor]

Bart C. Meijer [contributor]

Dmytro Mykhailenko [contributor]

Anton Mymrikov [contributor]

Andrew P. Norgan (ORCID) [contributor]

Sofia Ny (ORCID) [contributor]

Matthew Saab [contributor]

Jonas Salm [contributor]

Javier Sanchez (ORCID) [contributor]

Rogier P. Schade [contributor]

Bhanu N. M. Sinha (ORCID) [thesis advisor]

Jason Stull (ORCID) [contributor]

Anthony Underwood (ORCID) [contributor]

Anita Williams (ORCID) [contributor]