Data Set with 500 Isolates - WHONET Example

This example data set has the exact same structure as an export file from WHONET. Such files can be used with this package, as this example data set shows. The antibiotic results are from our example_isolates data set. All patient names are created using online surname generators and are only in place for practice purposes.

Usage

WHONET

Format

A data.frame with 500 observations and 53 variables:

Identification number
ID of the sample

Specimen number
ID of the specimen

Organism
Name of the microorganism. Before analysis, you should transform this to a valid microbial class, using as.mo().

Country
Country of origin

Laboratory
Name of laboratory

Last name
Fictitious last name of patient

First name
Fictitious initial of patient

Sex
Fictitious gender of patient

Age
Fictitious age of patient

Age category
Age group, can also be looked up using age_groups()

Date of admission
Date of hospital admission

Specimen date
Date when specimen was received at laboratory

Specimen type
Specimen type or group

Specimen type (Numeric)
Translation of "Specimen type"

Reason
Reason of request with Differential Diagnosis

Isolate number
ID of isolate

Organism type
Type of microorganism, can also be looked up using mo_type()

Serotype
Serotype of microorganism

Beta-lactamase
Microorganism produces beta-lactamase?

ESBL
Microorganism produces extended spectrum beta-lactamase?

Carbapenemase
Microorganism produces carbapenemase?

MRSA screening test
Microorganism is possible MRSA?

Inducible clindamycin resistance
Clindamycin can be induced?

Comment
Other comments

Date of data entry
Date this data was entered in WHONET

AMP_ND10:CIP_EE
28 different antibiotics. You can lookup the abbreviations in the antibiotics data set, or use e.g. ab_name("AMP") to get the official name immediately. Before analysis, you should transform this to a valid antibiotic class, using as.rsi().

Reference Data Publicly Available

All reference data sets (about microorganisms, antibiotics, R/SI interpretation, EUCAST rules, etc.) in this AMR package are publicly and freely available. We continually export our data sets to formats for use in R, SPSS, SAS, Stata and Excel. We also supply flat files that are machine-readable and suitable for input in any software program, such as laboratory information systems. Please find all download links on our website, which is automatically updated with every code change.

Examples

head(WHONET)
#>   Identification number Specimen number Organism         Country
#> 1            fe41d7bafa            1748      SPN         Belgium
#> 2            91f175ec37            1767      eco The Netherlands
#> 3            cc4015056e            1343      eco The Netherlands
#> 4            e864b692f5            1894      MAP         Denmark
#> 5            3d051fe345            1739      PVU         Belgium
#> 6            c80762a08d            1846      103 The Netherlands
#>                               Laboratory  Last name First name Sex Age
#> 1         National Laboratory of Belgium       Abel         B.   F  68
#> 2 National Laboratory of The Netherlands  Delacroix         F.   M  89
#> 3 National Laboratory of The Netherlands   Steensen         F.   M  85
#> 4         National Laboratory of Denmark Beyersdorf         L.   M  62
#> 5         National Laboratory of Belgium     Hummel         W.   M  86
#> 6 National Laboratory of The Netherlands  Eikenboom         J.   F  53
#>   Age category Date of admission Specimen date Specimen type
#> 1        55-74        2005-01-12    2005-01-30         Urine
#> 2          75+        2006-07-30    2006-08-16         Urine
#> 3          75+        2014-03-05    2014-03-14         Urine
#> 4        55-74        2014-10-22    2014-11-01         Urine
#> 5          75+        2016-10-28    2016-11-17         Urine
#> 6        25-54        2017-12-21    2018-01-02         Urine
#>   Specimen type (Numeric)  Reason Isolate number Organism type Serotype
#> 1                       2 Unknown           1748      Bacteria         
#> 2                       2 Unknown           1767      Bacteria         
#> 3                       2 Unknown           1343      Bacteria         
#> 4                       2 Unknown           1894      Bacteria         
#> 5                       2 Unknown           1739      Bacteria         
#> 6                       2 Unknown           1846      Bacteria         
#>   Beta-lactamase  ESBL Carbapenemase MRSA screening test
#> 1          FALSE FALSE         FALSE               FALSE
#> 2          FALSE FALSE         FALSE               FALSE
#> 3          FALSE FALSE         FALSE               FALSE
#> 4          FALSE FALSE         FALSE               FALSE
#> 5          FALSE FALSE         FALSE               FALSE
#> 6          FALSE FALSE         FALSE               FALSE
#>   Inducible clindamycin resistance Comment Date of data entry AMP_ND10 AMC_ED20
#> 1                            FALSE                 2005-01-30        S        S
#> 2                            FALSE                 2006-08-16     <NA>        S
#> 3                            FALSE                 2014-03-14        S        S
#> 4                            FALSE                 2014-11-01        R     <NA>
#> 5                            FALSE                 2016-11-17        R        R
#> 6                            FALSE                 2018-01-02        S        S
#>   TZP_ED30 FEP_ED30 CTX_ED5 FOX_ED30 CAZ_ED10 CRO_ED30 CIP_ED5 AMK_ED30
#> 1        S     <NA>    <NA>     <NA>        R     <NA>    <NA>     <NA>
#> 2     <NA>     <NA>    <NA>     <NA>        R     <NA>    <NA>     <NA>
#> 3        S     <NA>    <NA>     <NA>     <NA>     <NA>    <NA>     <NA>
#> 4     <NA>     <NA>    <NA>     <NA>        R     <NA>    <NA>     <NA>
#> 5        S     <NA>       S        R        S        S       S     <NA>
#> 6        S     <NA>       S        S        S        S       S     <NA>
#>   GEN_ED10 TOB_ED10 SXT_ED1.2 IPM_ND10 PEN_ND1 AMP_ND2 AMC_ND2 CHL_ND30 VAN_ED5
#> 1        R        R         S     <NA>       S       S       S     <NA>       S
#> 2        S     <NA>         R     <NA>       R    <NA>       S     <NA>       S
#> 3     <NA>     <NA>      <NA>     <NA>       S       S       S     <NA>    <NA>
#> 4        R        R      <NA>     <NA>       R       R    <NA>     <NA>    <NA>
#> 5        S        S         R        I       R       R       R     <NA>       R
#> 6        S        S         S        S       R       S       S     <NA>       R
#>   OXA_ED1 ERY_ED15 CLI_ED2 TCY_ED30 RIF_ED5 PEN_EE AMP_EE CRO_EE CIP_EE
#> 1    <NA>        S    <NA>        S    <NA>      S      S   <NA>   <NA>
#> 2    <NA>        S       S        S    <NA>      R   <NA>   <NA>   <NA>
#> 3    <NA>     <NA>       S     <NA>    <NA>      S      S   <NA>   <NA>
#> 4    <NA>     <NA>       S     <NA>    <NA>      R      R   <NA>   <NA>
#> 5    <NA>        R    <NA>        R       R      R      R      S      S
#> 6    <NA>        R    <NA>     <NA>       R      R      S      S      S