Data Set with 500 Isolates - WHONET Example

This example data set has the exact same structure as an export file from WHONET. Such files can be used with this package, as this example data set shows. The antibiotic results are from our example_isolates data set. All patient names are created using online surname generators and are only in place for practice purposes.

Usage

WHONET

Format

A tibble with 500 observations and 53 variables:

Identification number
ID of the sample

Specimen number
ID of the specimen

Organism
Name of the microorganism. Before analysis, you should transform this to a valid microbial class, using as.mo().

Country
Country of origin

Laboratory
Name of laboratory

Last name
Fictitious last name of patient

First name
Fictitious initial of patient

Sex
Fictitious gender of patient

Age
Fictitious age of patient

Age category
Age group, can also be looked up using age_groups()

Date of admission
Date of hospital admission

Specimen date
Date when specimen was received at laboratory

Specimen type
Specimen type or group

Specimen type (Numeric)
Translation of "Specimen type"

Reason
Reason of request with Differential Diagnosis

Isolate number
ID of isolate

Organism type
Type of microorganism, can also be looked up using mo_type()

Serotype
Serotype of microorganism

Beta-lactamase
Microorganism produces beta-lactamase?

ESBL
Microorganism produces extended spectrum beta-lactamase?

Carbapenemase
Microorganism produces carbapenemase?

MRSA screening test
Microorganism is possible MRSA?

Inducible clindamycin resistance
Clindamycin can be induced?

Comment
Other comments

Date of data entry
Date this data was entered in WHONET

AMP_ND10:CIP_EE
0 different antibiotics. You can lookup the abbreviations in the antibiotics data set, or use e.g. ab_name("AMP") to get the official name immediately. Before analysis, you should transform this to a valid antibiotic class, using as.sir().

Details

Like all data sets in this package, this data set is publicly available for download in the following formats: R, MS Excel, Apache Feather, Apache Parquet, SPSS, SAS, and Stata. Please visit our website for the download links. The actual files are of course available on our GitHub repository.

Examples

WHONET
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> Warning: * The 'rsi' class has been replaced with 'sir'. Transform your 'rsi'
#> columns to 'sir' with as.sir(), e.g.:
#>   your_data %>% mutate_if(is.rsi, as.sir)
#> # A tibble: 500 × 53
#>    Identif…¹ Speci…² Organ…³ Country Labor…⁴ Last …⁵ First…⁶ Sex     Age Age c…⁷
#>    <chr>       <int> <chr>   <chr>   <chr>   <chr>   <chr>   <chr> <dbl> <chr>  
#>  1 fe41d7ba…    1748 SPN     Belgium Nation… Abel    B.      F        68 55-74  
#>  2 91f175ec…    1767 eco     The Ne… Nation… Delacr… F.      M        89 75+    
#>  3 cc401505…    1343 eco     The Ne… Nation… Steens… F.      M        85 75+    
#>  4 e864b692…    1894 MAP     Denmark Nation… Beyers… L.      M        62 55-74  
#>  5 3d051fe3…    1739 PVU     Belgium Nation… Hummel  W.      M        86 75+    
#>  6 c80762a0…    1846 103     The Ne… Nation… Eikenb… J.      F        53 25-54  
#>  7 8022d372…    1628 103     Denmark Nation… Leclerc S.      F        77 75+    
#>  8 f3dc5f55…    1493 eco     The Ne… Nation… Delacr… W.      M        53 25-54  
#>  9 15add38f…    1847 eco     France  Nation… Van La… S.      F        63 55-74  
#> 10 fd41248d…    1458 eco     Germany Nation… Moulin  O.      F        75 75+    
#> # … with 490 more rows, 43 more variables: `Date of admission` <date>,
#> #   `Specimen date` <date>, `Specimen type` <chr>,
#> #   `Specimen type (Numeric)` <dbl>, Reason <chr>, `Isolate number` <int>,
#> #   `Organism type` <chr>, Serotype <chr>, `Beta-lactamase` <lgl>, ESBL <lgl>,
#> #   Carbapenemase <lgl>, `MRSA screening test` <lgl>,
#> #   `Inducible clindamycin resistance` <lgl>, Comment <chr>,
#> #   `Date of data entry` <date>, AMP_ND10 <rsi[!]>, AMC_ED20 <rsi[!]>, …