Age in Years of Individuals

Calculates age in years based on a reference date, which is the system date at default.

Usage

age(x, reference = Sys.Date(), exact = FALSE, na.rm = FALSE, ...)

Arguments

x

date(s), character (vectors) will be coerced with as.POSIXlt()

reference

reference date(s) (defaults to today), character (vectors) will be coerced with as.POSIXlt()

exact

a logical to indicate whether age calculation should be exact, i.e. with decimals. It divides the number of days of year-to-date (YTD) of x by the number of days in the year of reference (either 365 or 366).

na.rm

a logical to indicate whether missing values should be removed

...

arguments passed on to as.POSIXlt(), such as origin

Value

An integer (no decimals) if exact = FALSE, a double (with decimals) otherwise

Details

Ages below 0 will be returned as NA with a warning. Ages above 120 will only give a warning.

This function vectorises over both x and reference, meaning that either can have a length of 1 while the other argument has a larger length.

See also

Examples

# 10 random pre-Y2K birth dates
df <- data.frame(birth_date = as.Date("2000-01-01") - runif(10) * 25000)

# add ages
df$age <- age(df$birth_date)

# add exact ages
df$age_exact <- age(df$birth_date, exact = TRUE)

# add age at millenium switch
df$age_at_y2k <- age(df$birth_date, "2000-01-01")

df
#>    birth_date age age_exact age_at_y2k
#> 1  1982-03-20  40  40.44110         17
#> 2  1996-11-03  25  25.81644          3
#> 3  1970-05-07  52  52.30959         29
#> 4  1970-11-29  51  51.74521         29
#> 5  1934-01-13  88  88.62192         65
#> 6  1953-11-17  68  68.77808         46
#> 7  1970-08-02  52  52.07123         29
#> 8  1965-04-20  57  57.35616         34
#> 9  1943-07-28  79  79.08493         56
#> 10 1992-11-12  29  29.79178          7