Age in Years of Individuals

Calculates age in years based on a reference date, which is the system date at default.

Usage

age(x, reference = Sys.Date(), exact = FALSE, na.rm = FALSE, ...)

Arguments

x

date(s), character (vectors) will be coerced with as.POSIXlt()

reference

reference date(s) (default is today), character (vectors) will be coerced with as.POSIXlt()

exact

a logical to indicate whether age calculation should be exact, i.e. with decimals. It divides the number of days of year-to-date (YTD) of x by the number of days in the year of reference (either 365 or 366).

na.rm

a logical to indicate whether missing values should be removed

...

arguments passed on to as.POSIXlt(), such as origin

Value

An integer (no decimals) if exact = FALSE, a double (with decimals) otherwise

Details

Ages below 0 will be returned as NA with a warning. Ages above 120 will only give a warning.

This function vectorises over both x and reference, meaning that either can have a length of 1 while the other argument has a larger length.

See also

Examples

# 10 random pre-Y2K birth dates
df <- data.frame(birth_date = as.Date("2000-01-01") - runif(10) * 25000)

# add ages
df$age <- age(df$birth_date)

# add exact ages
df$age_exact <- age(df$birth_date, exact = TRUE)

# add age at millenium switch
df$age_at_y2k <- age(df$birth_date, "2000-01-01")

df
#>    birth_date age age_exact age_at_y2k
#> 1  1999-06-30  25  25.70411          0
#> 2  1968-01-29  57  57.12055         31
#> 3  1965-12-05  59  59.27123         34
#> 4  1980-03-01  45  45.03562         19
#> 5  1949-11-01  75  75.36438         50
#> 6  1947-02-14  78  78.07671         52
#> 7  1940-02-19  85  85.06301         59
#> 8  1988-01-10  37  37.17260         11
#> 9  1997-08-27  27  27.54521          2
#> 10 1978-01-26  47  47.12877         21