Antibiotic Selectors

These functions allow for filtering rows and selecting columns based on antibiotic test results that are of a specific antibiotic class or group, without the need to define the columns or antibiotic abbreviations. In short, if you have a column name that resembles an antimicrobial agent, it will be picked up by any of these functions that matches its pharmaceutical class: "cefazolin", "CZO" and "J01DB04" will all be picked up by cephalosporins().

Usage

ab_class(ab_class, only_rsi_columns = FALSE, only_treatable = TRUE, ...)

ab_selector(filter, only_rsi_columns = FALSE, only_treatable = TRUE, ...)

aminoglycosides(only_rsi_columns = FALSE, only_treatable = TRUE, ...)

aminopenicillins(only_rsi_columns = FALSE, ...)

antifungals(only_rsi_columns = FALSE, ...)

antimycobacterials(only_rsi_columns = FALSE, ...)

betalactams(only_rsi_columns = FALSE, only_treatable = TRUE, ...)

carbapenems(only_rsi_columns = FALSE, only_treatable = TRUE, ...)

cephalosporins(only_rsi_columns = FALSE, ...)

cephalosporins_1st(only_rsi_columns = FALSE, ...)

cephalosporins_2nd(only_rsi_columns = FALSE, ...)

cephalosporins_3rd(only_rsi_columns = FALSE, ...)

cephalosporins_4th(only_rsi_columns = FALSE, ...)

cephalosporins_5th(only_rsi_columns = FALSE, ...)

fluoroquinolones(only_rsi_columns = FALSE, ...)

glycopeptides(only_rsi_columns = FALSE, ...)

lincosamides(only_rsi_columns = FALSE, ...)

lipoglycopeptides(only_rsi_columns = FALSE, ...)

macrolides(only_rsi_columns = FALSE, ...)

oxazolidinones(only_rsi_columns = FALSE, ...)

penicillins(only_rsi_columns = FALSE, ...)

polymyxins(only_rsi_columns = FALSE, only_treatable = TRUE, ...)

streptogramins(only_rsi_columns = FALSE, ...)

quinolones(only_rsi_columns = FALSE, ...)

tetracyclines(only_rsi_columns = FALSE, ...)

trimethoprims(only_rsi_columns = FALSE, ...)

ureidopenicillins(only_rsi_columns = FALSE, ...)

administrable_per_os(only_rsi_columns = FALSE, ...)

administrable_iv(only_rsi_columns = FALSE, ...)

not_intrinsic_resistant(
  only_rsi_columns = FALSE,
  col_mo = NULL,
  version_expertrules = 3.3,
  ...
)

Arguments

ab_class

an antimicrobial class or a part of it, such as "carba" and "carbapenems". The columns group, atc_group1 and atc_group2 of the antibiotics data set will be searched (case-insensitive) for this value.

only_rsi_columns

a logical to indicate whether only columns of class <rsi> must be selected (defaults to FALSE), see as.rsi()

only_treatable

a logical to indicate whether agents that are only for laboratory tests should be excluded (defaults to TRUE), such as gentamicin-high (GEH) and imipenem/EDTA (IPE)

...

ignored, only in place to allow future extensions

filter

an expression to be evaluated in the antibiotics data set, such as name %like% "trim"

col_mo

column name of the IDs of the microorganisms (see as.mo()), defaults to the first column of class mo. Values will be coerced using as.mo().

version_expertrules

the version number to use for the EUCAST Expert Rules and Intrinsic Resistance guideline. Can be either "3.3", "3.2" or "3.1".

Value

(internally) a character vector of column names, with additional class "ab_selector"

Details

These functions can be used in data set calls for selecting columns and filtering rows. They are heavily inspired by the Tidyverse selection helpers such as everything(), but also work in base R and not only in dplyr verbs. Nonetheless, they are very convenient to use with dplyr functions such as select(), filter() and summarise(), see Examples.

All columns in the data in which these functions are called will be searched for known antibiotic names, abbreviations, brand names, and codes (ATC, EARS-Net, WHO, etc.) according to the antibiotics data set. This means that a selector such as aminoglycosides() will pick up column names like 'gen', 'genta', 'J01GB03', 'tobra', 'Tobracin', etc.

The ab_class() function can be used to filter/select on a manually defined antibiotic class. It searches for results in the antibiotics data set within the columns group, atc_group1 and atc_group2.

The ab_selector() function can be used to internally filter the antibiotics data set on any results, see Examples. It allows for filtering on a (part of) a certain name, and/or a group name or even a minimum of DDDs for oral treatment. This function yields the highest flexibility, but is also the least user-friendly, since it requires a hard-coded filter to set.

The administrable_per_os() and administrable_iv() functions also rely on the antibiotics data set - antibiotic columns will be matched where a DDD (defined daily dose) for resp. oral and IV treatment is available in the antibiotics data set.

The not_intrinsic_resistant() function can be used to only select antibiotic columns that pose no intrinsic resistance for the microorganisms in the data set. For example, if a data set contains only microorganism codes or names of E. coli and K. pneumoniae and contains a column "vancomycin", this column will be removed (or rather, unselected) using this function. It currently applies 'EUCAST Expert Rules' and 'EUCAST Intrinsic Resistance and Unusual Phenotypes' v3.3 (2021) to determine intrinsic resistance, using the eucast_rules() function internally. Because of this determination, this function is quite slow in terms of performance.

Full list of supported (antibiotic) classes

aminoglycosides() can select:
amikacin (AMK), amikacin/fosfomycin (AKF), amphotericin B-high (AMH), apramycin (APR), arbekacin (ARB), astromicin (AST), bekanamycin (BEK), dibekacin (DKB), framycetin (FRM), gentamicin (GEN), gentamicin-high (GEH), habekacin (HAB), hygromycin (HYG), isepamicin (ISE), kanamycin (KAN), kanamycin-high (KAH), kanamycin/cephalexin (KAC), micronomicin (MCR), neomycin (NEO), netilmicin (NET), pentisomicin (PIM), plazomicin (PLZ), propikacin (PKA), ribostamycin (RST), sisomicin (SIS), streptoduocin (STR), streptomycin (STR1), streptomycin-high (STH), tobramycin (TOB) and tobramycin-high (TOH)

aminopenicillins() can select:
amoxicillin (AMX) and ampicillin (AMP)

antifungals() can select:
5-fluorocytosine (FCT), amphotericin B (AMB), anidulafungin (ANI), butoconazole (BUT), caspofungin (CAS), ciclopirox (CIX), clotrimazole (CTR), econazole (ECO), fluconazole (FLU), fosfluconazole (FFL), griseofulvin (GRI), hachimycin (HCH), ibrexafungerp (IBX), isavuconazole (ISV), isoconazole (ISO), itraconazole (ITR), ketoconazole (KET), manogepix (MGX), micafungin (MIF), miconazole (MCZ), nystatin (NYS), pimaricin (PMR), posaconazole (POS), rezafungin (RZF), ribociclib (RBC), sulconazole (SUC), terbinafine (TRB), terconazole (TRC) and voriconazole (VOR)

antimycobacterials() can select:
4-aminosalicylic acid (AMA), calcium aminosalicylate (CLA), capreomycin (CAP), clofazimine (CLF), delamanid (DLM), enviomycin (ENV), ethambutol (ETH), ethambutol/isoniazid (ETI), ethionamide (ETI1), isoniazid (INH), morinamide (MRN), p-aminosalicylic acid (PAS), pretomanid (PMD), prothionamide (PTH), pyrazinamide (PZA), rifabutin (RIB), rifampicin (RIF), rifampicin/isoniazid (RFI), rifampicin/pyrazinamide/ethambutol/isoniazid (RPEI), rifampicin/pyrazinamide/isoniazid (RPI), rifamycin (RFM), rifapentine (RFP), simvastatin/fenofibrate (SMF), sodium aminosalicylate (SDA), streptomycin/isoniazid (STI), terizidone (TRZ), thioacetazone/isoniazid (THI1), tiocarlide (TCR) and viomycin (VIO)

betalactams() can select:
amoxicillin (AMX), amoxicillin/clavulanic acid (AMC), amoxicillin/sulbactam (AXS), ampicillin (AMP), ampicillin/sulbactam (SAM), apalcillin (APL), aspoxicillin (APX), avibactam (AVB), azidocillin (AZD), azlocillin (AZL), aztreonam (ATM), aztreonam/avibactam (AZA), aztreonam/nacubactam (ANC), bacampicillin (BAM), benzathine benzylpenicillin (BNB), benzathine phenoxymethylpenicillin (BNP), benzylpenicillin (PEN), biapenem (BIA), carbenicillin (CRB), carindacillin (CRN), cefacetrile (CAC), cefaclor (CEC), cefadroxil (CFR), cefaloridine (RID), cefamandole (MAN), cefatrizine (CTZ), cefazedone (CZD), cefazolin (CZO), cefcapene (CCP), cefcapene pivoxil (CCX), cefdinir (CDR), cefditoren (DIT), cefditoren pivoxil (DIX), cefepime (FEP), cefepime/clavulanic acid (CPC), cefepime/nacubactam (FNC), cefepime/tazobactam (FPT), cefetamet (CAT), cefetamet pivoxil (CPI), cefetecol (CCL), cefetrizole (CZL), cefixime (CFM), cefmenoxime (CMX), cefmetazole (CMZ), cefodizime (DIZ), cefonicid (CID), cefoperazone (CFP), cefoperazone/sulbactam (CSL), ceforanide (CND), cefoselis (CSE), cefotaxime (CTX), cefotaxime/clavulanic acid (CTC), cefotaxime/sulbactam (CTS), cefotetan (CTT), cefotiam (CTF), cefotiam hexetil (CHE), cefovecin (FOV), cefoxitin (FOX), cefoxitin screening (FOX1), cefozopran (ZOP), cefpimizole (CFZ), cefpiramide (CPM), cefpirome (CPO), cefpodoxime (CPD), cefpodoxime proxetil (CPX), cefpodoxime/clavulanic acid (CDC), cefprozil (CPR), cefquinome (CEQ), cefroxadine (CRD), cefsulodin (CFS), cefsumide (CSU), ceftaroline (CPT), ceftaroline/avibactam (CPA), ceftazidime (CAZ), ceftazidime/avibactam (CZA), ceftazidime/clavulanic acid (CCV), cefteram (CEM), cefteram pivoxil (CPL), ceftezole (CTL), ceftibuten (CTB), ceftiofur (TIO), ceftizoxime (CZX), ceftizoxime alapivoxil (CZP), ceftobiprole (BPR), ceftobiprole medocaril (CFM1), ceftolozane/enzyme inhibitor (CEI), ceftolozane/tazobactam (CZT), ceftriaxone (CRO), cefuroxime (CXM), cefuroxime axetil (CXA), cephalexin (LEX), cephalothin (CEP), cephapirin (HAP), cephradine (CED), ciclacillin (CIC), clometocillin (CLM), cloxacillin (CLO), dicloxacillin (DIC), doripenem (DOR), epicillin (EPC), ertapenem (ETP), flucloxacillin (FLC), hetacillin (HET), imipenem (IPM), imipenem/EDTA (IPE), imipenem/relebactam (IMR), latamoxef (LTM), lenampicillin (LEN), loracarbef (LOR), mecillinam (MEC), meropenem (MEM), meropenem/nacubactam (MNC), meropenem/vaborbactam (MEV), metampicillin (MTM), methicillin (MET), mezlocillin (MEZ), mezlocillin/sulbactam (MSU), nacubactam (NAC), nafcillin (NAF), oxacillin (OXA), panipenem (PAN), penamecillin (PNM), penicillin/novobiocin (PNO), penicillin/sulbactam (PSU), phenethicillin (PHE), phenoxymethylpenicillin (PHN), piperacillin (PIP), piperacillin/sulbactam (PIS), piperacillin/tazobactam (TZP), piridicillin (PRC), pivampicillin (PVM), pivmecillinam (PME), procaine benzylpenicillin (PRB), propicillin (PRP), razupenem (RZM), ritipenem (RIT), ritipenem acoxil (RIA), sarmoxicillin (SRX), sulbactam (SUL), sulbenicillin (SBC), sultamicillin (SLT6), talampicillin (TAL), tazobactam (TAZ), tebipenem (TBP), temocillin (TEM), ticarcillin (TIC) and ticarcillin/clavulanic acid (TCC)

carbapenems() can select:
biapenem (BIA), doripenem (DOR), ertapenem (ETP), imipenem (IPM), imipenem/EDTA (IPE), imipenem/relebactam (IMR), meropenem (MEM), meropenem/nacubactam (MNC), meropenem/vaborbactam (MEV), panipenem (PAN), razupenem (RZM), ritipenem (RIT), ritipenem acoxil (RIA) and tebipenem (TBP)

cephalosporins() can select:
cefacetrile (CAC), cefaclor (CEC), cefadroxil (CFR), cefaloridine (RID), cefamandole (MAN), cefatrizine (CTZ), cefazedone (CZD), cefazolin (CZO), cefcapene (CCP), cefcapene pivoxil (CCX), cefdinir (CDR), cefditoren (DIT), cefditoren pivoxil (DIX), cefepime (FEP), cefepime/clavulanic acid (CPC), cefepime/tazobactam (FPT), cefetamet (CAT), cefetamet pivoxil (CPI), cefetecol (CCL), cefetrizole (CZL), cefixime (CFM), cefmenoxime (CMX), cefmetazole (CMZ), cefodizime (DIZ), cefonicid (CID), cefoperazone (CFP), cefoperazone/sulbactam (CSL), ceforanide (CND), cefoselis (CSE), cefotaxime (CTX), cefotaxime/clavulanic acid (CTC), cefotaxime/sulbactam (CTS), cefotetan (CTT), cefotiam (CTF), cefotiam hexetil (CHE), cefovecin (FOV), cefoxitin (FOX), cefoxitin screening (FOX1), cefozopran (ZOP), cefpimizole (CFZ), cefpiramide (CPM), cefpirome (CPO), cefpodoxime (CPD), cefpodoxime proxetil (CPX), cefpodoxime/clavulanic acid (CDC), cefprozil (CPR), cefquinome (CEQ), cefroxadine (CRD), cefsulodin (CFS), cefsumide (CSU), ceftaroline (CPT), ceftaroline/avibactam (CPA), ceftazidime (CAZ), ceftazidime/avibactam (CZA), ceftazidime/clavulanic acid (CCV), cefteram (CEM), cefteram pivoxil (CPL), ceftezole (CTL), ceftibuten (CTB), ceftiofur (TIO), ceftizoxime (CZX), ceftizoxime alapivoxil (CZP), ceftobiprole (BPR), ceftobiprole medocaril (CFM1), ceftolozane/enzyme inhibitor (CEI), ceftolozane/tazobactam (CZT), ceftriaxone (CRO), cefuroxime (CXM), cefuroxime axetil (CXA), cephalexin (LEX), cephalothin (CEP), cephapirin (HAP), cephradine (CED), latamoxef (LTM) and loracarbef (LOR)

cephalosporins_1st() can select:
cefacetrile (CAC), cefadroxil (CFR), cefaloridine (RID), cefatrizine (CTZ), cefazedone (CZD), cefazolin (CZO), cefroxadine (CRD), ceftezole (CTL), cephalexin (LEX), cephalothin (CEP), cephapirin (HAP) and cephradine (CED)

cephalosporins_2nd() can select:
cefaclor (CEC), cefamandole (MAN), cefmetazole (CMZ), cefonicid (CID), ceforanide (CND), cefotetan (CTT), cefotiam (CTF), cefoxitin (FOX), cefoxitin screening (FOX1), cefprozil (CPR), cefuroxime (CXM), cefuroxime axetil (CXA) and loracarbef (LOR)

cephalosporins_3rd() can select:
cefcapene (CCP), cefcapene pivoxil (CCX), cefdinir (CDR), cefditoren (DIT), cefditoren pivoxil (DIX), cefetamet (CAT), cefetamet pivoxil (CPI), cefixime (CFM), cefmenoxime (CMX), cefodizime (DIZ), cefoperazone (CFP), cefoperazone/sulbactam (CSL), cefotaxime (CTX), cefotaxime/clavulanic acid (CTC), cefotaxime/sulbactam (CTS), cefotiam hexetil (CHE), cefovecin (FOV), cefpimizole (CFZ), cefpiramide (CPM), cefpodoxime (CPD), cefpodoxime proxetil (CPX), cefpodoxime/clavulanic acid (CDC), cefsulodin (CFS), ceftazidime (CAZ), ceftazidime/avibactam (CZA), ceftazidime/clavulanic acid (CCV), cefteram (CEM), cefteram pivoxil (CPL), ceftibuten (CTB), ceftiofur (TIO), ceftizoxime (CZX), ceftizoxime alapivoxil (CZP), ceftriaxone (CRO) and latamoxef (LTM)

cephalosporins_4th() can select:
cefepime (FEP), cefepime/clavulanic acid (CPC), cefepime/tazobactam (FPT), cefetecol (CCL), cefoselis (CSE), cefozopran (ZOP), cefpirome (CPO) and cefquinome (CEQ)

cephalosporins_5th() can select:
ceftaroline (CPT), ceftaroline/avibactam (CPA), ceftobiprole (BPR), ceftobiprole medocaril (CFM1), ceftolozane/enzyme inhibitor (CEI) and ceftolozane/tazobactam (CZT)

fluoroquinolones() can select:
besifloxacin (BES), ciprofloxacin (CIP), clinafloxacin (CLX), danofloxacin (DAN), delafloxacin (DFX), difloxacin (DIF), enoxacin (ENX), enrofloxacin (ENR), finafloxacin (FIN), fleroxacin (FLE), garenoxacin (GRN), gatifloxacin (GAT), gemifloxacin (GEM), grepafloxacin (GRX), levofloxacin (LVX), levonadifloxacin (LND), lomefloxacin (LOM), marbofloxacin (MAR), metioxate (MXT), miloxacin (MIL), moxifloxacin (MFX), nadifloxacin (NAD), nifuroquine (NIF), norfloxacin (NOR), ofloxacin (OFX), orbifloxacin (ORB), pazufloxacin (PAZ), pefloxacin (PEF), pradofloxacin (PRA), premafloxacin (PRX), prulifloxacin (PRU), rufloxacin (RFL), sarafloxacin (SAR), sitafloxacin (SIT), sparfloxacin (SPX), temafloxacin (TMX), tilbroquinol (TBQ), tioxacin (TXC), tosufloxacin (TFX) and trovafloxacin (TVA)

glycopeptides() can select:
avoparcin (AVO), dalbavancin (DAL), norvancomycin (NVA), oritavancin (ORI), ramoplanin (RAM), teicoplanin (TEC), teicoplanin-macromethod (TCM), telavancin (TLV), vancomycin (VAN) and vancomycin-macromethod (VAM)

lincosamides() can select:
acetylmidecamycin (ACM), acetylspiramycin (ASP), clindamycin (CLI), gamithromycin (GAM), kitasamycin (KIT), lincomycin (LIN), meleumycin (MEL), nafithromycin (ZWK), pirlimycin (PRL), primycin (PRM), solithromycin (SOL), tildipirosin (TIP), tilmicosin (TIL), tulathromycin (TUL), tylosin (TYL) and tylvalosin (TYL1)

lipoglycopeptides() can select:
dalbavancin (DAL), oritavancin (ORI) and telavancin (TLV)

macrolides() can select:
acetylmidecamycin (ACM), acetylspiramycin (ASP), azithromycin (AZM), clarithromycin (CLR), dirithromycin (DIR), erythromycin (ERY), flurithromycin (FLR1), gamithromycin (GAM), josamycin (JOS), kitasamycin (KIT), meleumycin (MEL), midecamycin (MID), miocamycin (MCM), nafithromycin (ZWK), oleandomycin (OLE), pirlimycin (PRL), primycin (PRM), rokitamycin (ROK), roxithromycin (RXT), solithromycin (SOL), spiramycin (SPI), telithromycin (TLT), tildipirosin (TIP), tilmicosin (TIL), troleandomycin (TRL), tulathromycin (TUL), tylosin (TYL) and tylvalosin (TYL1)

oxazolidinones() can select:
cadazolid (CDZ), cycloserine (CYC), linezolid (LNZ), tedizolid (TZD) and thiacetazone (THA)

penicillins() can select:
amoxicillin (AMX), amoxicillin/clavulanic acid (AMC), amoxicillin/sulbactam (AXS), ampicillin (AMP), ampicillin/sulbactam (SAM), apalcillin (APL), aspoxicillin (APX), avibactam (AVB), azidocillin (AZD), azlocillin (AZL), aztreonam (ATM), aztreonam/avibactam (AZA), aztreonam/nacubactam (ANC), bacampicillin (BAM), benzathine benzylpenicillin (BNB), benzathine phenoxymethylpenicillin (BNP), benzylpenicillin (PEN), carbenicillin (CRB), carindacillin (CRN), cefepime/nacubactam (FNC), ciclacillin (CIC), clometocillin (CLM), cloxacillin (CLO), dicloxacillin (DIC), epicillin (EPC), flucloxacillin (FLC), hetacillin (HET), lenampicillin (LEN), mecillinam (MEC), metampicillin (MTM), methicillin (MET), mezlocillin (MEZ), mezlocillin/sulbactam (MSU), nacubactam (NAC), nafcillin (NAF), oxacillin (OXA), penamecillin (PNM), penicillin/novobiocin (PNO), penicillin/sulbactam (PSU), phenethicillin (PHE), phenoxymethylpenicillin (PHN), piperacillin (PIP), piperacillin/sulbactam (PIS), piperacillin/tazobactam (TZP), piridicillin (PRC), pivampicillin (PVM), pivmecillinam (PME), procaine benzylpenicillin (PRB), propicillin (PRP), sarmoxicillin (SRX), sulbactam (SUL), sulbenicillin (SBC), sultamicillin (SLT6), talampicillin (TAL), tazobactam (TAZ), temocillin (TEM), ticarcillin (TIC) and ticarcillin/clavulanic acid (TCC)

polymyxins() can select:
colistin (COL), polymyxin B (PLB) and polymyxin B/polysorbate 80 (POP)

quinolones() can select:
besifloxacin (BES), cinoxacin (CIN), ciprofloxacin (CIP), clinafloxacin (CLX), danofloxacin (DAN), delafloxacin (DFX), difloxacin (DIF), enoxacin (ENX), enrofloxacin (ENR), finafloxacin (FIN), fleroxacin (FLE), flumequine (FLM), garenoxacin (GRN), gatifloxacin (GAT), gemifloxacin (GEM), grepafloxacin (GRX), levofloxacin (LVX), levonadifloxacin (LND), lomefloxacin (LOM), marbofloxacin (MAR), metioxate (MXT), miloxacin (MIL), moxifloxacin (MFX), nadifloxacin (NAD), nalidixic acid (NAL), nifuroquine (NIF), nitroxoline (NTR), norfloxacin (NOR), ofloxacin (OFX), orbifloxacin (ORB), oxolinic acid (OXO), pazufloxacin (PAZ), pefloxacin (PEF), pipemidic acid (PPA), piromidic acid (PIR), pradofloxacin (PRA), premafloxacin (PRX), prulifloxacin (PRU), rosoxacin (ROS), rufloxacin (RFL), sarafloxacin (SAR), sitafloxacin (SIT), sparfloxacin (SPX), temafloxacin (TMX), tilbroquinol (TBQ), tioxacin (TXC), tosufloxacin (TFX) and trovafloxacin (TVA)

streptogramins() can select:
pristinamycin (PRI) and quinupristin/dalfopristin (QDA)

tetracyclines() can select:
cetocycline (CTO), chlortetracycline (CTE), clomocycline (CLM1), demeclocycline (DEM), doxycycline (DOX), eravacycline (ERV), lymecycline (LYM), metacycline (MTC), minocycline (MNO), omadacycline (OMC), oxytetracycline (OXY), penimepicycline (PNM1), rolitetracycline (RLT), tetracycline (TCY) and tigecycline (TGC)

trimethoprims() can select:
brodimoprim (BDP), sulfadiazine (SDI), sulfadiazine/tetroxoprim (SLT), sulfadiazine/trimethoprim (SLT1), sulfadimethoxine (SUD), sulfadimidine (SDM), sulfadimidine/trimethoprim (SLT2), sulfafurazole (SLF), sulfaisodimidine (SLF1), sulfalene (SLF2), sulfamazone (SZO), sulfamerazine (SLF3), sulfamerazine/trimethoprim (SLT3), sulfamethizole (SLF4), sulfamethoxazole (SMX), sulfamethoxypyridazine (SLF5), sulfametomidine (SLF6), sulfametoxydiazine (SLF7), sulfametrole/trimethoprim (SLT4), sulfamoxole (SLF8), sulfamoxole/trimethoprim (SLT5), sulfanilamide (SLF9), sulfaperin (SLF10), sulfaphenazole (SLF11), sulfapyridine (SLF12), sulfathiazole (SUT), sulfathiourea (SLF13), trimethoprim (TMP) and trimethoprim/sulfamethoxazole (SXT)

ureidopenicillins() can select:
azlocillin (AZL), mezlocillin (MEZ), piperacillin (PIP) and piperacillin/tazobactam (TZP)

Reference Data Publicly Available

All reference data sets (about microorganisms, antibiotics, R/SI interpretation, EUCAST rules, etc.) in this AMR package are publicly and freely available. We continually export our data sets to formats for use in R, SPSS, SAS, Stata and Excel. We also supply flat files that are machine-readable and suitable for input in any software program, such as laboratory information systems. Please find all download links on our website, which is automatically updated with every code change.

Examples

# `example_isolates` is a data set available in the AMR package.
# See ?example_isolates.
df <- example_isolates[ , c("hospital_id", "mo",
                            "AMP", "AMC", "TZP", "CXM", "CRO", "GEN",
                            "TOB", "COL", "IPM", "MEM", "TEC", "VAN")]

# base R ------------------------------------------------------------------

# select columns 'IPM' (imipenem) and 'MEM' (meropenem)
df[, carbapenems()]
#> ℹ For `carbapenems()` using columns 'IPM' (imipenem) and 'MEM' (meropenem)
#> # A tibble: 2,000 × 2
#>    IPM   MEM  
#>    <rsi> <rsi>
#>  1 NA    NA   
#>  2 NA    NA   
#>  3 NA    NA   
#>  4 NA    NA   
#>  5 NA    NA   
#>  6 NA    NA   
#>  7 NA    NA   
#>  8 NA    NA   
#>  9 NA    NA   
#> 10 NA    NA   
#> # … with 1,990 more rows

# select columns 'mo', 'AMK', 'GEN', 'KAN' and 'TOB'
df[, c("mo", aminoglycosides())]
#> ℹ For `aminoglycosides()` using columns 'GEN' (gentamicin) and 'TOB'
#>   (tobramycin)
#> # A tibble: 2,000 × 3
#>    mo           GEN   TOB  
#>    <mo>         <rsi> <rsi>
#>  1 B_ESCHR_COLI NA    NA   
#>  2 B_ESCHR_COLI NA    NA   
#>  3 B_STPHY_EPDR NA    NA   
#>  4 B_STPHY_EPDR NA    NA   
#>  5 B_STPHY_EPDR NA    NA   
#>  6 B_STPHY_EPDR NA    NA   
#>  7 B_STPHY_AURS NA    S    
#>  8 B_STPHY_AURS NA    S    
#>  9 B_STPHY_EPDR NA    NA   
#> 10 B_STPHY_EPDR NA    NA   
#> # … with 1,990 more rows

# select only antibiotic columns with DDDs for oral treatment
df[, administrable_per_os()]
#> ℹ For `administrable_per_os()` using columns 'AMP' (ampicillin), 'AMC'
#>   (amoxicillin/clavulanic acid), 'CXM' (cefuroxime), 'COL' (colistin) and
#>   'VAN' (vancomycin)
#> # A tibble: 2,000 × 5
#>    AMP   AMC   CXM   COL   VAN  
#>    <rsi> <rsi> <rsi> <rsi> <rsi>
#>  1 NA    I     I     NA    R    
#>  2 NA    I     I     NA    R    
#>  3 NA    NA    R     R     S    
#>  4 NA    NA    R     R     S    
#>  5 NA    NA    R     R     S    
#>  6 NA    NA    R     R     S    
#>  7 R     S     S     R     S    
#>  8 R     S     S     R     S    
#>  9 NA    NA    R     R     S    
#> 10 NA    NA    S     R     S    
#> # … with 1,990 more rows

# filter using any() or all()
df[any(carbapenems() == "R"), ]
#> ℹ For `carbapenems()` using columns 'IPM' (imipenem) and 'MEM' (meropenem)
#> # A tibble: 55 × 14
#>    hospital…¹ mo           AMP   AMC   TZP   CXM   CRO   GEN   TOB   COL   IPM  
#>    <fct>      <mo>         <rsi> <rsi> <rsi> <rsi> <rsi> <rsi> <rsi> <rsi> <rsi>
#>  1 A          B_ENTRC_FACM NA    NA    NA    R     R     R     R     R     R    
#>  2 A          B_ENTRC_FACM NA    NA    NA    R     R     R     R     R     R    
#>  3 B          B_STNTR_MLTP R     R     R     R     R     R     R     R     R    
#>  4 B          B_ENTRC      NA    NA    R     NA    NA    NA    NA    R     R    
#>  5 B          B_ENTRC      NA    NA    R     NA    NA    NA    NA    R     R    
#>  6 B          B_ENTRC_FACM NA    NA    R     R     R     R     R     R     R    
#>  7 D          B_ENTRC_FACM NA    NA    R     R     R     R     R     R     R    
#>  8 A          B_ENTRC_FACM NA    NA    R     R     R     R     R     R     R    
#>  9 A          B_PROTS_MRBL NA    S     S     S     S     S     S     R     R    
#> 10 A          B_PROTS_MRBL NA    S     S     S     S     S     S     R     R    
#> # … with 45 more rows, 3 more variables: MEM <rsi>, TEC <rsi>, VAN <rsi>, and
#> #   abbreviated variable name ¹​hospital_id
subset(df, any(carbapenems() == "R"))
#> ℹ For `carbapenems()` using columns 'IPM' (imipenem) and 'MEM' (meropenem)
#> # A tibble: 55 × 14
#>    hospital…¹ mo           AMP   AMC   TZP   CXM   CRO   GEN   TOB   COL   IPM  
#>    <fct>      <mo>         <rsi> <rsi> <rsi> <rsi> <rsi> <rsi> <rsi> <rsi> <rsi>
#>  1 A          B_ENTRC_FACM NA    NA    NA    R     R     R     R     R     R    
#>  2 A          B_ENTRC_FACM NA    NA    NA    R     R     R     R     R     R    
#>  3 B          B_STNTR_MLTP R     R     R     R     R     R     R     R     R    
#>  4 B          B_ENTRC      NA    NA    R     NA    NA    NA    NA    R     R    
#>  5 B          B_ENTRC      NA    NA    R     NA    NA    NA    NA    R     R    
#>  6 B          B_ENTRC_FACM NA    NA    R     R     R     R     R     R     R    
#>  7 D          B_ENTRC_FACM NA    NA    R     R     R     R     R     R     R    
#>  8 A          B_ENTRC_FACM NA    NA    R     R     R     R     R     R     R    
#>  9 A          B_PROTS_MRBL NA    S     S     S     S     S     S     R     R    
#> 10 A          B_PROTS_MRBL NA    S     S     S     S     S     S     R     R    
#> # … with 45 more rows, 3 more variables: MEM <rsi>, TEC <rsi>, VAN <rsi>, and
#> #   abbreviated variable name ¹​hospital_id

# filter on any or all results in the carbapenem columns (i.e., IPM, MEM):
df[any(carbapenems()), ]
#> ℹ For `carbapenems()` using columns 'IPM' (imipenem) and 'MEM' (meropenem)
#> ℹ Filtering any of columns 'IPM' and 'MEM' to contain value "R", "S" or "I"
#> # A tibble: 962 × 14
#>    hospital…¹ mo           AMP   AMC   TZP   CXM   CRO   GEN   TOB   COL   IPM  
#>    <fct>      <mo>         <rsi> <rsi> <rsi> <rsi> <rsi> <rsi> <rsi> <rsi> <rsi>
#>  1 D          B_ESCHR_COLI NA    I     NA    S     S     NA    S     NA    S    
#>  2 D          B_ESCHR_COLI NA    I     NA    S     S     NA    S     NA    S    
#>  3 B          B_PROTS_MRBL NA    I     NA    S     S     NA    NA    R     S    
#>  4 B          B_PROTS_MRBL NA    I     NA    S     S     NA    NA    R     S    
#>  5 B          B_SERRT_MRCS R     R     NA    R     NA    NA    NA    R     S    
#>  6 B          B_SERRT_MRCS R     R     NA    R     NA    NA    NA    R     S    
#>  7 B          B_SERRT_MRCS R     R     NA    R     NA    NA    NA    R     S    
#>  8 D          B_KLBSL_PNMN R     I     NA    S     S     S     S     NA    S    
#>  9 D          B_KLBSL_PNMN R     I     NA    S     S     S     S     NA    S    
#> 10 C          B_ESCHR_COLI R     R     NA    R     S     S     NA    NA    S    
#> # … with 952 more rows, 3 more variables: MEM <rsi>, TEC <rsi>, VAN <rsi>, and
#> #   abbreviated variable name ¹​hospital_id
df[all(carbapenems()), ]
#> ℹ For `carbapenems()` using columns 'IPM' (imipenem) and 'MEM' (meropenem)
#> ℹ Filtering all of columns 'IPM' and 'MEM' to contain value "R", "S" or "I"
#> # A tibble: 756 × 14
#>    hospita…¹ mo            AMP   AMC   TZP   CXM   CRO   GEN   TOB   COL   IPM  
#>    <fct>     <mo>          <rsi> <rsi> <rsi> <rsi> <rsi> <rsi> <rsi> <rsi> <rsi>
#>  1 C         B_STRPT_GRPB  S     S     S     S     S     R     R     R     S    
#>  2 B         B_STRPT_PYGN  S     S     S     S     S     R     R     R     S    
#>  3 B         B_STRPT_GRPA  S     S     S     S     S     R     R     R     S    
#>  4 B         B_STRPT_GRPA  S     S     S     S     S     R     R     R     S    
#>  5 D         B_STRPT_GRPB  S     S     S     S     S     R     R     R     S    
#>  6 B         B_ESCHR_COLI  R     R     S     S     S     S     S     NA    S    
#>  7 B         B_ENTRBC_CLOC R     R     S     R     NA    S     S     NA    S    
#>  8 B         B_ENTRBC_CLOC R     R     S     R     NA    S     S     NA    S    
#>  9 B         B_ESCHR_COLI  NA    S     S     S     S     S     S     NA    S    
#> 10 B         B_ESCHR_COLI  NA    S     S     S     S     S     S     NA    S    
#> # … with 746 more rows, 3 more variables: MEM <rsi>, TEC <rsi>, VAN <rsi>, and
#> #   abbreviated variable name ¹​hospital_id

# filter with multiple antibiotic selectors using c()
df[all(c(carbapenems(), aminoglycosides()) == "R"), ]
#> ℹ For `carbapenems()` using columns 'IPM' (imipenem) and 'MEM' (meropenem)
#> ℹ For `aminoglycosides()` using columns 'GEN' (gentamicin) and 'TOB'
#>   (tobramycin)
#> # A tibble: 26 × 14
#>    hospital…¹ mo           AMP   AMC   TZP   CXM   CRO   GEN   TOB   COL   IPM  
#>    <fct>      <mo>         <rsi> <rsi> <rsi> <rsi> <rsi> <rsi> <rsi> <rsi> <rsi>
#>  1 B          B_STNTR_MLTP R     R     R     R     R     R     R     R     R    
#>  2 B          B_ENTRC_FACM NA    NA    R     R     R     R     R     R     R    
#>  3 D          B_ENTRC_FACM NA    NA    R     R     R     R     R     R     R    
#>  4 A          B_ENTRC_FACM NA    NA    R     R     R     R     R     R     R    
#>  5 D          B_STNTR_MLTP R     R     R     R     R     R     R     R     R    
#>  6 D          B_ENTRC_FACM NA    NA    R     R     R     R     R     R     R    
#>  7 B          B_STPHY_CONS R     R     R     R     R     R     R     R     R    
#>  8 A          B_STPHY_CONS R     R     R     R     R     R     R     R     R    
#>  9 D          B_STPHY_CONS R     R     R     R     R     R     R     R     R    
#> 10 D          B_STPHY_CONS R     R     R     R     R     R     R     R     R    
#> # … with 16 more rows, 3 more variables: MEM <rsi>, TEC <rsi>, VAN <rsi>, and
#> #   abbreviated variable name ¹​hospital_id

# filter + select in one go: get penicillins in carbapenems-resistant strains
df[any(carbapenems() == "R"), penicillins()]
#> ℹ For `carbapenems()` using columns 'IPM' (imipenem) and 'MEM' (meropenem)
#> ℹ For `penicillins()` using columns 'AMP' (ampicillin), 'AMC'
#>   (amoxicillin/clavulanic acid) and 'TZP' (piperacillin/tazobactam)
#> # A tibble: 55 × 3
#>    AMP   AMC   TZP  
#>    <rsi> <rsi> <rsi>
#>  1 NA    NA    NA   
#>  2 NA    NA    NA   
#>  3 R     R     R    
#>  4 NA    NA    R    
#>  5 NA    NA    R    
#>  6 NA    NA    R    
#>  7 NA    NA    R    
#>  8 NA    NA    R    
#>  9 NA    S     S    
#> 10 NA    S     S    
#> # … with 45 more rows

# You can combine selectors with '&' to be more specific. For example,
# penicillins() would select benzylpenicillin ('peni G') and
# administrable_per_os() would select erythromycin. Yet, when combined these
# drugs are both omitted since benzylpenicillin is not administrable per os
# and erythromycin is not a penicillin:
df[, penicillins() & administrable_per_os()]
#> ℹ For `penicillins()` using columns 'AMP' (ampicillin), 'AMC'
#>   (amoxicillin/clavulanic acid) and 'TZP' (piperacillin/tazobactam)
#> ℹ For `administrable_per_os()` using columns 'AMP' (ampicillin), 'AMC'
#>   (amoxicillin/clavulanic acid), 'CXM' (cefuroxime), 'COL' (colistin) and
#>   'VAN' (vancomycin)
#> # A tibble: 2,000 × 2
#>    AMP   AMC  
#>    <rsi> <rsi>
#>  1 NA    I    
#>  2 NA    I    
#>  3 NA    NA   
#>  4 NA    NA   
#>  5 NA    NA   
#>  6 NA    NA   
#>  7 R     S    
#>  8 R     S    
#>  9 NA    NA   
#> 10 NA    NA   
#> # … with 1,990 more rows

# ab_selector() applies a filter in the `antibiotics` data set and is thus very
# flexible. For instance, to select antibiotic columns with an oral DDD of at
# least 1 gram:
df[, ab_selector(oral_ddd > 1 & oral_units == "g")]
#> ℹ For `ab_selector(oral_ddd > 1 & oral_units == "g")` using columns 'AMP'
#>   (ampicillin), 'AMC' (amoxicillin/clavulanic acid) and 'VAN' (vancomycin)
#> # A tibble: 2,000 × 3
#>    AMP   AMC   VAN  
#>    <rsi> <rsi> <rsi>
#>  1 NA    I     R    
#>  2 NA    I     R    
#>  3 NA    NA    S    
#>  4 NA    NA    S    
#>  5 NA    NA    S    
#>  6 NA    NA    S    
#>  7 R     S     S    
#>  8 R     S     S    
#>  9 NA    NA    S    
#> 10 NA    NA    S    
#> # … with 1,990 more rows

# dplyr -------------------------------------------------------------------
# \donttest{
if (require("dplyr")) {

  # get AMR for all aminoglycosides e.g., per hospital:
  df %>%
    group_by(hospital_id) %>% 
    summarise(across(aminoglycosides(), resistance))
    
  # You can combine selectors with '&' to be more specific:
  df %>%
    select(penicillins() & administrable_per_os())
  
  # get AMR for only drugs that matter - no intrinsic resistance:
  df %>%
    filter(mo_genus() %in% c("Escherichia", "Klebsiella")) %>% 
    group_by(hospital_id) %>% 
    summarise(across(not_intrinsic_resistant(), resistance))
    
  # get susceptibility for antibiotics whose name contains "trim":
  df %>%
    filter(first_isolate()) %>% 
    group_by(hospital_id) %>% 
    summarise(across(ab_selector(name %like% "trim"), susceptibility))

  # this will select columns 'IPM' (imipenem) and 'MEM' (meropenem):
  df %>% 
    select(carbapenems())
    
  # this will select columns 'mo', 'AMK', 'GEN', 'KAN' and 'TOB':
  df %>% 
    select(mo, aminoglycosides())
    
 # any() and all() work in dplyr's filter() too:
 df %>% 
    filter(any(aminoglycosides() == "R"),
           all(cephalosporins_2nd() == "R"))
    
 # also works with c():
 df %>% 
    filter(any(c(carbapenems(), aminoglycosides()) == "R"))
    
 # not setting any/all will automatically apply all():
 df %>% 
    filter(aminoglycosides() == "R")
    
  # this will select columns 'mo' and all antimycobacterial drugs ('RIF'):
  df %>% 
    select(mo, ab_class("mycobact"))
    
  # get bug/drug combinations for only glycopeptides in Gram-positives:
  df %>% 
    filter(mo_is_gram_positive()) %>% 
    select(mo, glycopeptides()) %>% 
    bug_drug_combinations() %>%
    format()
    
  data.frame(some_column = "some_value",
             J01CA01 = "S") %>%   # ATC code of ampicillin
    select(penicillins())         # only the 'J01CA01' column will be selected
    
    
  # with recent versions of dplyr this is all equal:
  x <- df[carbapenems() == "R", ]
  y <- df %>% filter(carbapenems() == "R")
  z <- df %>% filter(if_all(carbapenems(), ~.x == "R"))
  identical(x, y)
  identical(y, z)
}
#> ℹ For `aminoglycosides()` using columns 'GEN' (gentamicin) and 'TOB'
#>   (tobramycin)
#> ℹ For `penicillins()` using columns 'AMP' (ampicillin), 'AMC'
#>   (amoxicillin/clavulanic acid) and 'TZP' (piperacillin/tazobactam)
#> ℹ For `administrable_per_os()` using columns 'AMP' (ampicillin), 'AMC'
#>   (amoxicillin/clavulanic acid), 'CXM' (cefuroxime), 'COL' (colistin) and
#>   'VAN' (vancomycin)
#> ℹ Using column 'mo' as input for `mo_genus()`
#> ℹ For `not_intrinsic_resistant()` removing columns 'TEC' (teicoplanin) and
#>   'VAN' (vancomycin)
#> Warning: Introducing NA: only 23 results available for COL in group: hospital_id =
#> "C" (`minimum` = 30).
#> Error in filter(., first_isolate()): Problem while computing `..1 = first_isolate()`.
#> Caused by error:
#> ! in first_isolate(): `col_date` must be set
# }