Data Sets with 566 Antimicrobial Drugs

Two data sets containing all antibiotics/antimycotics and antivirals. Use as.ab() or one of the ab_* functions to retrieve values from the antibiotics data set. Three identifiers are included in this data set: an antibiotic ID (ab, primarily used in this package) as defined by WHONET/EARS-Net, an ATC code (atc) as defined by the WHO, and a Compound ID (cid) as found in PubChem. Other properties in this data set are derived from one or more of these codes. Note that some drugs have multiple ATC codes.

Usage

antibiotics

antivirals

Format

For the antibiotics data set: a data.frame with 464 observations and 14 variables:

ab
Antibiotic ID as used in this package (such as AMC), using the official EARS-Net (European Antimicrobial Resistance Surveillance Network) codes where available

cid
Compound ID as found in PubChem

name
Official name as used by WHONET/EARS-Net or the WHO

group
A short and concise group name, based on WHONET and WHOCC definitions

atc
ATC codes (Anatomical Therapeutic Chemical) as defined by the WHOCC, like J01CR02

atc_group1
Official pharmacological subgroup (3rd level ATC code) as defined by the WHOCC, like "Macrolides, lincosamides and streptogramins"

atc_group2
Official chemical subgroup (4th level ATC code) as defined by the WHOCC, like "Macrolides"

abbr
List of abbreviations as used in many countries, also for antibiotic susceptibility testing (AST)

synonyms
Synonyms (often trade names) of a drug, as found in PubChem based on their compound ID

oral_ddd
Defined Daily Dose (DDD), oral treatment, currently available for 170 drugs

oral_units
Units of oral_ddd

iv_ddd
Defined Daily Dose (DDD), parenteral (intravenous) treatment, currently available for 145 drugs

iv_units
Units of iv_ddd

loinc
All LOINC codes (Logical Observation Identifiers Names and Codes) associated with the name of the antimicrobial agent. Use ab_loinc() to retrieve them quickly, see ab_property().

For the antivirals data set: a data.frame with 102 observations and 9 variables:

atc
ATC codes (Anatomical Therapeutic Chemical) as defined by the WHOCC

cid
Compound ID as found in PubChem

name
Official name as used by WHONET/EARS-Net or the WHO

atc_group
Official pharmacological subgroup (3rd level ATC code) as defined by the WHOCC

synonyms
Synonyms (often trade names) of a drug, as found in PubChem based on their compound ID

oral_ddd
Defined Daily Dose (DDD), oral treatment

oral_units
Units of oral_ddd

iv_ddd
Defined Daily Dose (DDD), parenteral treatment

iv_units
Units of iv_ddd

An object of class data.frame with 102 rows and 9 columns.

Source

World Health Organization (WHO) Collaborating Centre for Drug Statistics Methodology (WHOCC): https://www.whocc.no/atc_ddd_index/

European Commission Public Health PHARMACEUTICALS - COMMUNITY REGISTER: https://ec.europa.eu/health/documents/community-register/html/reg_hum_atc.htm

Details

Properties that are based on an ATC code are only available when an ATC is available. These properties are: atc_group1, atc_group2, oral_ddd, oral_units, iv_ddd and iv_units.

Synonyms (i.e. trade names) were derived from the Compound ID (cid) and consequently only available where a CID is available.

Direct download

These data sets are available as 'flat files' for use even without R - you can find the files here:

https://github.com/msberends/AMR/raw/main/data-raw/antibiotics.txt

https://github.com/msberends/AMR/raw/main/data-raw/antivirals.txt

Files in R format (with preserved data structure) can be found here:

https://github.com/msberends/AMR/raw/main/data/antibiotics.rda

https://github.com/msberends/AMR/raw/main/data/antivirals.rda

Reference Data Publicly Available

All reference data sets (about microorganisms, antibiotics, R/SI interpretation, EUCAST rules, etc.) in this AMR package are publicly and freely available. We continually export our data sets to formats for use in R, SPSS, SAS, Stata and Excel. We also supply flat files that are machine-readable and suitable for input in any software program, such as laboratory information systems. Please find all download links on our website, which is automatically updated with every code change.

WHOCC

This package contains all ~550 antibiotic, antimycotic and antiviral drugs and their Anatomical Therapeutic Chemical (ATC) codes, ATC groups and Defined Daily Dose (DDD) from the World Health Organization Collaborating Centre for Drug Statistics Methodology (WHOCC, https://www.whocc.no) and the Pharmaceuticals Community Register of the European Commission (https://ec.europa.eu/health/documents/community-register/html/reg_hum_atc.htm).

These have become the gold standard for international drug utilisation monitoring and research.

The WHOCC is located in Oslo at the Norwegian Institute of Public Health and funded by the Norwegian government. The European Commission is the executive of the European Union and promotes its general interest.

NOTE: The WHOCC copyright does not allow use for commercial purposes, unlike any other info from this package. See https://www.whocc.no/copyright_disclaimer/.

See also

Examples

head(antibiotics)
#>    ab      cid                  name                    group
#> 1 AMA     4649 4-aminosalicylic acid       Antimycobacterials
#> 2 FCT     3366      5-fluorocytosine Antifungals/antimycotics
#> 3 ACM  6450012     Acetylmidecamycin  Macrolides/lincosamides
#> 4 ASP 49787020      Acetylspiramycin  Macrolides/lincosamides
#> 5 ALS     8954    Aldesulfone sodium     Other antibacterials
#> 6 AMK    37768              Amikacin          Aminoglycosides
#>                         atc                          atc_group1
#> 1                   J04AA01 Drugs for treatment of tuberculosis
#> 2                   D01AE21         Antifungals for topical use
#> 3                        NA                                <NA>
#> 4                        NA                                <NA>
#> 5                   J04BA03        Drugs for treatment of lepra
#> 6 D06AX12, J01GB06, S01AA21       Aminoglycoside antibacterials
#>                            atc_group2          abbreviations
#> 1 Aminosalicylic acid and derivatives                       
#> 2   Other antifungals for topical use  5flc, fcu, fluo, fluy
#> 3                                <NA>                       
#> 4                                <NA>                       
#> 5        Drugs for treatment of lepra                       
#> 6               Other aminoglycosides ak, ami, amik, amk, an
#>                                                                                                                                                                                                                                                                                                                                                                synonyms
#> 1 aminopar, aminosalicylic, aminosalicylic acid, aminosalyl, aminox, apacil, deapasil, entepas, ferrosan, gabbropas, helipidyl, hellipidyl, neopasalate, osacyl, pamacyl, pamisyl, paramycin, parasal, parasalicil, parasalindon, pasalon, pasara, pascorbic, pasdium, paser granules, paskalium, pasmed, pasnodia, pasolac, propasa, rezipas, teebacin, wln: zr cq dvq
#> 2                                                                                                                                                                                                                          alcobon, ancobon, ancotil, ancotyl, flucitosina, flucystine, flucytosin, flucytosine, flucytosinum, flucytosone, fluocytosine, fluorcytosine
#> 3                                                                                                                                                                                                                                                                                                                                                                      
#> 4                                                                                                                                                                                                                                                                                                                        acetylspiramycin, foromacidin b, spiramycin ii
#> 5                                                                                                                                                 adesulfone sodium, aldapsone, aldesulfona sodica, aldesulfone, aldesulfone sodique, aldesulfone sodium, diamidin, diasone, diasone sodium, diazon, novotrone, sodium aldesulphone, sodium sulfoxone, sulfoxone sodium
#> 6                                                                                                                                            amicacin, amikacillin, amikacin, amikacin base, amikacin dihydrate, amikacin sulfate, amikacina, amikacine, amikacinum, amikavet, amikin, amiklin, amikozit, amukin, arikace, briclin, lukadin, mikavir, pierami, potentox
#>   oral_ddd oral_units iv_ddd iv_units
#> 1    12.00          g     NA     <NA>
#> 2       NA       <NA>     NA     <NA>
#> 3       NA       <NA>     NA     <NA>
#> 4       NA       <NA>     NA     <NA>
#> 5     0.33          g     NA     <NA>
#> 6       NA       <NA>      1        g
#>                                                                                                                                loinc
#> 1                                                                                                                                   
#> 2                                                                                10974-4, 23805-5, 25142-1, 25143-9, 3639-2, 46218-4
#> 3                                                                                                                                   
#> 4                                                                                                                                   
#> 5                                                                                                                                   
#> 6 13546-7, 15098-7, 17798-0, 31097-9, 31098-7, 31099-5, 3319-1, 3320-9, 3321-7, 35669-1, 50802-8, 50803-6, 56628-1, 59378-0, 80972-3
head(antivirals)
#>       atc       cid               name
#> 1 J05AF06    441300           Abacavir
#> 2 J05AB01 135398513          Aciclovir
#> 3 J05AF08     60871 Adefovir dipivoxil
#> 4 J05AE05     65016         Amprenavir
#> 5 J05AP06  16076883        Asunaprevir
#> 6 J05AE08    148192         Atazanavir
#>                                                            atc_group
#> 1         Nucleoside and nucleotide reverse transcriptase inhibitors
#> 2 Nucleosides and nucleotides excl. reverse transcriptase inhibitors
#> 3         Nucleoside and nucleotide reverse transcriptase inhibitors
#> 4                                                Protease inhibitors
#> 5                         Antivirals for treatment of HCV infections
#> 6                                                Protease inhibitors
#>                                                                                                                                                                                                                                                                                                          synonyms
#> 1                                                                                                                                                                                                                                                                              Abacavir, Abacavir sulfate, Ziagen
#> 2 Acicloftal, Aciclovier, Aciclovir, Aciclovirum, Activir, AcycloFoam, Acycloguanosine, Acyclovir, Acyclovir Lauriad, ACYCLOVIR SODIUM, Avirax, Cargosil, Cyclovir, Genvir, Gerpevir, Hascovir, Herpevir, Maynar, Poviral, Sitavig, Sitavir, Vipral, Virolex, Viropump, Virorax, Zovirax, Zovirax topical, Zyclir
#> 3                                                                                                                                 Adefovir di ester, Adefovir dipivoxil, Adefovir Dipivoxil, Adefovir dipivoxyl, Adefovir pivoxil, Adefovirdipivoxl, Bisadenine, BISADENINE, BisPMEA, Hepsera, Preveon, YouHeDing
#> 4                                                                                                                                                                                                                                                             Agenerase, Amprenavir, Amprenavirum, Prozei, Vertex
#> 5                                                                                                                                                                                                                                                                                           Asunaprevir, Sunvepra
#> 6                                                                                                                                                                                                                                                      Atazanavir, Atazanavir Base, Latazanavir, Reyataz, Zrivada
#>   oral_ddd oral_units iv_ddd iv_units
#> 1      0.6          g     NA     <NA>
#> 2      4.0          g      4        g
#> 3     10.0         mg     NA     <NA>
#> 4      1.2          g     NA     <NA>
#> 5       NA       <NA>     NA     <NA>
#> 6      0.3          g     NA     <NA>