Data sets with 558 antimicrobials

Two data sets containing all antibiotics/antimycotics and antivirals. Use as.ab() or one of the ab_property() functions to retrieve values from the antibiotics data set. Three identifiers are included in this data set: an antibiotic ID (ab, primarily used in this package) as defined by WHONET/EARS-Net, an ATC code (atc) as defined by the WHO, and a Compound ID (cid) as found in PubChem. Other properties in this data set are derived from one or more of these codes.

antibiotics

antivirals

Format

For the antibiotics data set: a data.frame with 456 observations and 14 variables:

ab
Antibiotic ID as used in this package (like AMC), using the official EARS-Net (European Antimicrobial Resistance Surveillance Network) codes where available

atc
ATC code (Anatomical Therapeutic Chemical) as defined by the WHOCC, like J01CR02

cid
Compound ID as found in PubChem

name
Official name as used by WHONET/EARS-Net or the WHO

group
A short and concise group name, based on WHONET and WHOCC definitions

atc_group1
Official pharmacological subgroup (3rd level ATC code) as defined by the WHOCC, like "Macrolides, lincosamides and streptogramins"

atc_group2
Official chemical subgroup (4th level ATC code) as defined by the WHOCC, like "Macrolides"

abbr
List of abbreviations as used in many countries, also for antibiotic susceptibility testing (AST)

synonyms
Synonyms (often trade names) of a drug, as found in PubChem based on their compound ID

oral_ddd
Defined Daily Dose (DDD), oral treatment

oral_units
Units of oral_ddd

iv_ddd
Defined Daily Dose (DDD), parenteral treatment

iv_units
Units of iv_ddd

loinc
All LOINC codes (Logical Observation Identifiers Names and Codes) associated with the name of the antimicrobial agent. Use ab_loinc() to retrieve them quickly, see ab_property().

For the antivirals data set: a data.frame with 102 observations and 9 variables:

atc
ATC code (Anatomical Therapeutic Chemical) as defined by the WHOCC

cid
Compound ID as found in PubChem

name
Official name as used by WHONET/EARS-Net or the WHO

atc_group
Official pharmacological subgroup (3rd level ATC code) as defined by the WHOCC

synonyms
Synonyms (often trade names) of a drug, as found in PubChem based on their compound ID

oral_ddd
Defined Daily Dose (DDD), oral treatment

oral_units
Units of oral_ddd

iv_ddd
Defined Daily Dose (DDD), parenteral treatment

iv_units
Units of iv_ddd

An object of class data.frame with 102 rows and 9 columns.

Source

World Health Organization (WHO) Collaborating Centre for Drug Statistics Methodology (WHOCC): https://www.whocc.no/atc_ddd_index/

WHONET 2019 software: http://www.whonet.org/software.html

European Commission Public Health PHARMACEUTICALS - COMMUNITY REGISTER: http://ec.europa.eu/health/documents/community-register/html/atc.htm

Details

Properties that are based on an ATC code are only available when an ATC is available. These properties are: atc_group1, atc_group2, oral_ddd, oral_units, iv_ddd and iv_units.

Synonyms (i.e. trade names) are derived from the Compound ID (cid) and consequently only available where a CID is available.

Direct download

These data sets are available as 'flat files' for use even without R - you can find the files here:

https://gitlab.com/msberends/AMR/raw/master/data-raw/antibiotics.txt

https://gitlab.com/msberends/AMR/raw/master/data-raw/antivirals.txt

Files in R format (with preserved data structure) can be found here:

https://gitlab.com/msberends/AMR/raw/master/data/antibiotics.rda

https://gitlab.com/msberends/AMR/raw/master/data/antivirals.rda

WHOCC

This package contains all ~550 antibiotic, antimycotic and antiviral drugs and their Anatomical Therapeutic Chemical (ATC) codes, ATC groups and Defined Daily Dose (DDD) from the World Health Organization Collaborating Centre for Drug Statistics Methodology (WHOCC, https://www.whocc.no) and the Pharmaceuticals Community Register of the European Commission (http://ec.europa.eu/health/documents/community-register/html/atc.htm).

These have become the gold standard for international drug utilisation monitoring and research.

The WHOCC is located in Oslo at the Norwegian Institute of Public Health and funded by the Norwegian government. The European Commission is the executive of the European Union and promotes its general interest.

NOTE: The WHOCC copyright does not allow use for commercial purposes, unlike any other info from this package. See https://www.whocc.no/copyright_disclaimer/.

Read more on our website!

On our website https://msberends.gitlab.io/AMR you can find a comprehensive tutorial about how to conduct AMR analysis, the complete documentation of all functions (which reads a lot easier than here in R) and an example analysis using WHONET data.

See also