Data Sets with 603 Antimicrobial Drugs

Two data sets containing all antibiotics/antimycotics and antivirals. Use as.ab() or one of the ab_* functions to retrieve values from the antibiotics data set. Three identifiers are included in this data set: an antibiotic ID (ab, primarily used in this package) as defined by WHONET/EARS-Net, an ATC code (atc) as defined by the WHO, and a Compound ID (cid) as found in PubChem. Other properties in this data set are derived from one or more of these codes. Note that some drugs have multiple ATC codes.

Usage

antibiotics

antivirals

Format

For the antibiotics data set: a tibble with 483 observations and 14 variables:

ab
Antibiotic ID as used in this package (such as AMC), using the official EARS-Net (European Antimicrobial Resistance Surveillance Network) codes where available

cid
Compound ID as found in PubChem

name
Official name as used by WHONET/EARS-Net or the WHO

group
A short and concise group name, based on WHONET and WHOCC definitions

atc
ATC codes (Anatomical Therapeutic Chemical) as defined by the WHOCC, like J01CR02

atc_group1
Official pharmacological subgroup (3rd level ATC code) as defined by the WHOCC, like "Macrolides, lincosamides and streptogramins"

atc_group2
Official chemical subgroup (4th level ATC code) as defined by the WHOCC, like "Macrolides"

abbr
List of abbreviations as used in many countries, also for antibiotic susceptibility testing (AST)

synonyms
Synonyms (often trade names) of a drug, as found in PubChem based on their compound ID

oral_ddd
Defined Daily Dose (DDD), oral treatment, currently available for 174 drugs

oral_units
Units of oral_ddd

iv_ddd
Defined Daily Dose (DDD), parenteral (intravenous) treatment, currently available for 146 drugs

iv_units
Units of iv_ddd

loinc
All LOINC codes (Logical Observation Identifiers Names and Codes) associated with the name of the antimicrobial drug. Use ab_loinc() to retrieve them quickly, see ab_property().

For the antivirals data set: a tibble with 120 observations and 11 variables:

av
Antibiotic ID as used in this package (such as AMC), using the official EARS-Net (European Antimicrobial Resistance Surveillance Network) codes where available

name
Official name as used by WHONET/EARS-Net or the WHO

atc
ATC codes (Anatomical Therapeutic Chemical) as defined by the WHOCC

cid
Compound ID as found in PubChem

atc_group
Official pharmacological subgroup (3rd level ATC code) as defined by the WHOCC

synonyms
Synonyms (often trade names) of a drug, as found in PubChem based on their compound ID

oral_ddd
Defined Daily Dose (DDD), oral treatment

oral_units
Units of oral_ddd

iv_ddd
Defined Daily Dose (DDD), parenteral treatment

iv_units
Units of iv_ddd

loinc
All LOINC codes (Logical Observation Identifiers Names and Codes) associated with the name of the antimicrobial drug.

An object of class tbl_df (inherits from tbl, data.frame) with 120 rows and 11 columns.

Source

World Health Organization (WHO) Collaborating Centre for Drug Statistics Methodology (WHOCC): https://www.whocc.no/atc_ddd_index/

Logical Observation Identifiers Names and Codes (LOINC), Version 2.73 (8 August, 2022). Accessed from https://loinc.org on 30 October, 2022.

European Commission Public Health PHARMACEUTICALS - COMMUNITY REGISTER: https://ec.europa.eu/health/documents/community-register/html/reg_hum_atc.htm

Details

Properties that are based on an ATC code are only available when an ATC is available. These properties are: atc_group1, atc_group2, oral_ddd, oral_units, iv_ddd and iv_units.

Synonyms (i.e. trade names) were derived from the PubChem Compound ID (column cid) and consequently only available where a CID is available.

Direct download

Like all data sets in this package, these data sets are publicly available for download in the following formats: R, MS Excel, Apache Feather, Apache Parquet, SPSS, SAS, and Stata. Please visit our website for the download links. The actual files are of course available on our GitHub repository.

WHOCC

This package contains all ~550 antibiotic, antimycotic and antiviral drugs and their Anatomical Therapeutic Chemical (ATC) codes, ATC groups and Defined Daily Dose (DDD) from the World Health Organization Collaborating Centre for Drug Statistics Methodology (WHOCC, https://www.whocc.no) and the Pharmaceuticals Community Register of the European Commission (https://ec.europa.eu/health/documents/community-register/html/reg_hum_atc.htm).

These have become the gold standard for international drug utilisation monitoring and research.

The WHOCC is located in Oslo at the Norwegian Institute of Public Health and funded by the Norwegian government. The European Commission is the executive of the European Union and promotes its general interest.

NOTE: The WHOCC copyright does not allow use for commercial purposes, unlike any other info from this package. See https://www.whocc.no/copyright_disclaimer/.

See also

Examples

antibiotics
#> # A tibble: 483 × 14
#>    ab        cid name        group atc   atc_g…¹ atc_g…² abbre…³ synon…⁴ oral_…⁵
#>    <ab>    <dbl> <chr>       <chr> <lis> <chr>   <chr>   <list>  <named>   <dbl>
#>  1 AMA      4649 4-aminosal… Anti… <chr> Drugs … Aminos… <chr>   <chr>     12   
#>  2 ACM   6450012 Acetylmide… Macr… <chr> NA      NA      <chr>   <chr>     NA   
#>  3 ASP  49787020 Acetylspir… Macr… <chr> NA      NA      <chr>   <chr>     NA   
#>  4 ALS      8954 Aldesulfon… Othe… <chr> Drugs … Drugs … <chr>   <chr>      0.33
#>  5 AMK     37768 Amikacin    Amin… <chr> Aminog… Other … <chr>   <chr>     NA   
#>  6 AKF        NA Amikacin/f… Amin… <chr> NA      NA      <chr>   <chr>     NA   
#>  7 AMX     33613 Amoxicillin Beta… <chr> Beta-l… Penici… <chr>   <chr>      1.5 
#>  8 AMC  23665637 Amoxicilli… Beta… <chr> Beta-l… Combin… <chr>   <chr>      1.5 
#>  9 AXS    465441 Amoxicilli… Beta… <chr> NA      NA      <chr>   <chr>     NA   
#> 10 AMB   5280965 Amphoteric… Anti… <chr> Antimy… Antibi… <chr>   <chr>     40   
#> # … with 473 more rows, 4 more variables: oral_units <chr>, iv_ddd <dbl>,
#> #   iv_units <chr>, loinc <list>, and abbreviated variable names ¹​atc_group1,
#> #   ²​atc_group2, ³​abbreviations, ⁴​synonyms, ⁵​oral_ddd
antivirals
#> # A tibble: 120 × 11
#>    av    name  atc      cid atc_g…¹ synon…² oral_…³ oral_…⁴ iv_ddd iv_un…⁵ loinc
#>    <av>  <chr> <chr>  <dbl> <chr>   <list>    <dbl> <chr>    <dbl> <chr>   <lis>
#>  1 ABA   Abac… J05A… 4.41e5 Nucleo… <chr>       0.6 g           NA NA      <chr>
#>  2 ACI   Acic… J05A… 1.35e8 Nucleo… <chr>       4   g            4 g       <chr>
#>  3 ADD   Adef… J05A… 6.09e4 Nucleo… <chr>      10   mg          NA NA      <chr>
#>  4 AME   Amen… J05A… 1.14e7 Other … <chr>       0.4 g           NA NA      <chr>
#>  5 AMP   Ampr… J05A… 6.50e4 Protea… <chr>       1.2 g           NA NA      <chr>
#>  6 ASU   Asun… J05A… 1.61e7 Antivi… <chr>       0.2 g           NA NA      <chr>
#>  7 ATA   Ataz… J05A… 1.48e5 Protea… <chr>       0.3 g           NA NA      <chr>
#>  8 ATA+… Ataz… J05A… 8.66e7 Antivi… <chr>      NA   NA          NA NA      <chr>
#>  9 ATA+… Ataz… J05A… 2.51e7 Antivi… <chr>       0.3 g           NA NA      <chr>
#> 10 BAM   Balo… J05A… 1.24e8 Other … <chr>      40   mg          NA NA      <chr>
#> # … with 110 more rows, and abbreviated variable names ¹​atc_group, ²​synonyms,
#> #   ³​oral_ddd, ⁴​oral_units, ⁵​iv_units