Class 'disk'

This transforms a vector to a new class disk, which is a growth zone size (around an antibiotic disk) in millimeters between 6 and 99.

as.disk(x, na.rm = FALSE)

is.disk(x)

Arguments

x
vector

na.rm
a logical indicating whether missing values should be removed

Value

Ordered integer factor with new class disk

Details

Interpret disk values as RSI values with as.rsi. It supports guidelines from EUCAST and CLSI.

Read more on our website!

On our website https://msberends.gitlab.io/AMR you can find a comprehensive tutorial about how to conduct AMR analysis, the complete documentation of all functions (which reads a lot easier than here in R) and an example analysis using WHONET data.

See also

Examples

# NOT RUN {
# interpret disk values
as.rsi(x = 12,
       mo = as.mo("S. pneumoniae"),
       ab = "AMX",
       guideline = "EUCAST")
as.rsi(x = 12,
       mo = as.mo("S. pneumoniae"),
       ab = "AMX",
       guideline = "CLSI")
# }