Get ATC properties from WHOCC website

Gets data from the WHO to determine properties of an ATC (e.g. an antibiotic) like name, defined daily dose (DDD) or standard unit.

atc_online_property(
  atc_code,
  property,
  administration = "O",
  url = "https://www.whocc.no/atc_ddd_index/?code=%s&showdescription=no"
)

atc_online_groups(atc_code, ...)

atc_online_ddd(atc_code, ...)

Arguments

atc_code
a character or character vector with ATC code(s) of antibiotic(s)

property
property of an ATC code. Valid values are "ATC", "Name", "DDD", "U" ("unit"), "Adm.R", "Note" and groups. For this last option, all hierarchical groups of an ATC code will be returned, see Examples.

administration
type of administration when using property = "Adm.R", see Details

url
url of website of the WHO. The sign %s can be used as a placeholder for ATC codes.

...
parameters to pass on to atc_property

Source

https://www.whocc.no/atc_ddd_alterations__cumulative/ddd_alterations/abbrevations/

Details

Options for parameter administration:

"Implant" = Implant

"Inhal" = Inhalation

"Instill" = Instillation

"N" = nasal

"O" = oral

"P" = parenteral

"R" = rectal

"SL" = sublingual/buccal

"TD" = transdermal

"V" = vaginal

Abbreviations of return values when using property = "U" (unit):

"g" = gram

"mg" = milligram

`"mcg"`` = microgram

"U" = unit

"TU" = thousand units

"MU" = million units

"mmol" = millimole

"ml" = milliliter (e.g. eyedrops)

N.B. This function requires an internet connection and only works if the following packages are installed: curl, rvest, xml2.

Stable lifecycle

The lifecycle of this function is stable. In a stable function, major changes are unlikely. This means that the unlying code will generally evolve by adding new arguments; removing arguments or changing the meaning of existing arguments will be avoided.

If the unlying code needs breaking changes, they will occur gradually. For example, a parameter will be deprecated and first continue to work, but will emit an message informing you of the change. Next, typically after at least one newly released version on CRAN, the message will be transformed to an error.

Read more on our website!

On our website https://msberends.github.io/AMR you can find a comprehensive tutorial about how to conduct AMR analysis, the complete documentation of all functions (which reads a lot easier than here in R) and an example analysis using WHONET data. As we would like to better understand the backgrounds and needs of our users, please participate in our survey!

Examples

if (FALSE) {
# oral DDD (Defined Daily Dose) of amoxicillin
atc_online_property("J01CA04", "DDD", "O")
# parenteral DDD (Defined Daily Dose) of amoxicillin
atc_online_property("J01CA04", "DDD", "P")

atc_online_property("J01CA04", property = "groups") # search hierarchical groups of amoxicillin
# [1] "ANTIINFECTIVES FOR SYSTEMIC USE"
# [2] "ANTIBACTERIALS FOR SYSTEMIC USE"
# [3] "BETA-LACTAM ANTIBACTERIALS, PENICILLINS"
# [4] "Penicillins with extended spectrum"
}