Version info of included Catalogue of Life

This function returns information about the included data from the Catalogue of Life.

catalogue_of_life_version()

Value

a list, which prints in pretty format

Details

For DSMZ, see microorganisms.

Catalogue of Life

This package contains the complete taxonomic tree of almost all microorganisms (~70,000 species) from the authoritative and comprehensive Catalogue of Life (http://www.catalogueoflife.org). The Catalogue of Life is the most comprehensive and authoritative global index of species currently available.

Click here for more information about the included taxa. Check which version of the Catalogue of Life was included in this package with catalogue_of_life_version().

Read more on our website!

On our website https://msberends.gitlab.io/AMR you can find a comprehensive tutorial about how to conduct AMR analysis, the complete documentation of all functions (which reads a lot easier than here in R) and an example analysis using WHONET data.

See also

Examples

library(dplyr)
microorganisms %>% freq(kingdom)
microorganisms %>% group_by(kingdom) %>% freq(phylum, nmax = NULL)