Data Set with Treatment Dosages as Defined by EUCAST

EUCAST breakpoints used in this package are based on the dosages in this data set. They can be retrieved with eucast_dosage().

dosage

Format

A data.frame with 169 observations and 9 variables:

ab
Antibiotic ID as used in this package (such as AMC), using the official EARS-Net (European Antimicrobial Resistance Surveillance Network) codes where available

name
Official name of the antimicrobial agent as used by WHONET/EARS-Net or the WHO

type
Type of the dosage, either "high_dosage", "standard_dosage" or "uncomplicated_uti"

dose
Dose, such as "2 g" or "25 mg/kg"

dose_times
Number of times a dose must be administered

administration
Route of administration, either "im", "iv" or "oral"

notes
Additional dosage notes

original_txt
Original text in the PDF file of EUCAST

eucast_version
Version number of the EUCAST Clinical Breakpoints guideline to which these dosages apply

Details

'EUCAST Clinical Breakpoint Tables' v11.0 (2021) are based on the dosages in this data set.

Reference Data Publicly Available

All reference data sets (about microorganisms, antibiotics, R/SI interpretation, EUCAST rules, etc.) in this AMR package are publicly and freely available. We continually export our data sets to formats for use in R, SPSS, SAS, Stata and Excel. We also supply flat files that are machine-readable and suitable for input in any software program, such as laboratory information systems. Please find all download links on our website, which is automatically updated with every code change.

Read more on Our Website!

On our website https://msberends.github.io/AMR/ you can find a comprehensive tutorial about how to conduct AMR data analysis, the complete documentation of all functions and an example analysis using WHONET data.