Data Set with Treatment Dosages as Defined by EUCAST

EUCAST breakpoints used in this package are based on the dosages in this data set. They can be retrieved with eucast_dosage().

Usage

dosage

Format

A tibble with 336 observations and 9 variables:

ab
Antibiotic ID as used in this package (such as AMC), using the official EARS-Net (European Antimicrobial Resistance Surveillance Network) codes where available

name
Official name of the antimicrobial drug as used by WHONET/EARS-Net or the WHO

type
Type of the dosage, either "high_dosage", "standard_dosage", or "uncomplicated_uti"

dose
Dose, such as "2 g" or "25 mg/kg"

dose_times
Number of times a dose must be administered

administration
Route of administration, either "im", "iv", or "oral"

notes
Additional dosage notes

original_txt
Original text in the PDF file of EUCAST

eucast_version
Version number of the EUCAST Clinical Breakpoints guideline to which these dosages apply

Details

This data set is based on 'EUCAST Clinical Breakpoint Tables' v12.0 (2022) and 'EUCAST Clinical Breakpoint Tables' v11.0 (2021).

Direct download

Like all data sets in this package, this data set is publicly available for download in the following formats: R, MS Excel, Apache Feather, Apache Parquet, SPSS, SAS, and Stata. Please visit our website for the download links. The actual files are of course available on our GitHub repository.

Examples

dosage
#> # A tibble: 336 × 9
#>    ab   name                   type  dose  dose_…¹ admin…² notes origi…³ eucas…⁴
#>    <ab> <chr>                  <chr> <chr>   <int> <chr>   <chr> <chr>     <dbl>
#>  1 AMK  Amikacin               stan… 25-3…       1 iv      ""    25-30 …      12
#>  2 AMX  Amoxicillin            high… 2 g         6 iv      ""    2 g x …      12
#>  3 AMX  Amoxicillin            stan… 1 g         3 iv      ""    1 g x …      12
#>  4 AMX  Amoxicillin            high… 0.75…       3 oral    ""    0.75-1…      12
#>  5 AMX  Amoxicillin            stan… 0.5 g       3 oral    ""    0.5 g …      12
#>  6 AMX  Amoxicillin            unco… 0.5 g       3 oral    ""    0.5 g …      12
#>  7 AMC  Amoxicillin/clavulani… high… 2 g …       3 iv      ""    (2 g a…      12
#>  8 AMC  Amoxicillin/clavulani… stan… 1 g …       3 iv      ""    (1 g a…      12
#>  9 AMC  Amoxicillin/clavulani… high… 0.87…       3 oral    ""    (0.875…      12
#> 10 AMC  Amoxicillin/clavulani… stan… 0.5 …       3 oral    ""    (0.5 g…      12
#> # … with 326 more rows, and abbreviated variable names ¹​dose_times,
#> #   ²​administration, ³​original_txt, ⁴​eucast_version