Data Set with Unclean Data

A data set containing 3 000 microbial isolates that are not cleaned up and consequently not ready for AMR data analysis. This data set can be used for practice.

Usage

example_isolates_unclean

Format

A tibble with 3 000 observations and 8 variables:

patient_id
ID of the patient

date
date of receipt at the laboratory

hospital
ID of the hospital, from A to C

bacteria
info about microorganism that can be transformed with as.mo(), see also microorganisms

AMX:GEN
4 different antimicrobials that have to be transformed with as.sir()

Details

Like all data sets in this package, this data set is publicly available for download in the following formats: R, MS Excel, Apache Feather, Apache Parquet, SPSS, and Stata. Please visit our website for the download links. The actual files are of course available on our GitHub repository.

Examples

example_isolates_unclean
#> # A tibble: 3,000 × 8
#>    patient_id hospital date       bacteria      AMX   AMC   CIP   GEN  
#>    <chr>      <chr>    <date>     <chr>         <chr> <chr> <chr> <chr>
#>  1 J3         A        2012-11-21 E. coli       R     I     S     S    
#>  2 R7         A        2018-04-03 K. pneumoniae R     I     S     S    
#>  3 P3         A        2014-09-19 E. coli       R     S     S     S    
#>  4 P10        A        2015-12-10 E. coli       S     I     S     S    
#>  5 B7         A        2015-03-02 E. coli       S     S     S     S    
#>  6 W3         A        2018-03-31 S. aureus     R     S     R     S    
#>  7 J8         A        2016-06-14 E. coli       R     S     S     S    
#>  8 M3         A        2015-10-25 E. coli       R     S     S     S    
#>  9 J3         A        2019-06-19 E. coli       S     S     S     S    
#> 10 G6         A        2015-04-27 S. aureus     S     S     S     S    
#> # ℹ 2,990 more rows