Determine (New) Episodes for Patients

These functions determine which items in a vector can be considered (the start of) a new episode, based on the argument episode_days. This can be used to determine clinical episodes for any epidemiological analysis. The get_episode() function returns the index number of the episode per group, while the is_new_episode() function returns values TRUE/FALSE to indicate whether an item in a vector is the start of a new episode.

Usage

get_episode(x, episode_days, ...)

is_new_episode(x, episode_days, ...)

Arguments

x

vector of dates (class Date or POSIXt), will be sorted internally to determine episodes

episode_days

required episode length in days, can also be less than a day or Inf, see Details

...

ignored, only in place to allow future extensions

Value

get_episode(): a double vector

is_new_episode(): a logical vector

Details

Dates are first sorted from old to new. The oldest date will mark the start of the first episode. After this date, the next date will be marked that is at least episode_days days later than the start of the first episode. From that second marked date on, the next date will be marked that is at least episode_days days later than the start of the second episode which will be the start of the third episode, and so on. Before the vector is being returned, the original order will be restored.

The first_isolate() function is a wrapper around the is_new_episode() function, but is more efficient for data sets containing microorganism codes or names and allows for different isolate selection methods.

The dplyr package is not required for these functions to work, but these functions do support variable grouping and work conveniently inside dplyr verbs such as filter(), mutate() and summarise().

See also

Examples

# `example_isolates` is a data set available in the AMR package.
# See ?example_isolates
df <- example_isolates[sample(seq_len(2000), size = 200), ]

get_episode(df$date, episode_days = 60) # indices
#>   [1] 60 59 40 62 19 23 22  5 49 47  2 53 35 32 20 62 40 56 17 42 59 57 38 15 61
#>  [26] 14 26 13  6  9 27 28 59 34 36 44  9  7 47 17  5 16  5 14 13 24 48  9 62  5
#>  [51] 56  1 57  6 26 52 22 13 16 51 42 19 30 13 15  2 38 63  4  9 13 34 13  1 24
#>  [76] 29 44 19 11 58  1 61 21 37 61 64 41 26 39 40 19 31 21  1 43  1 23 52 27  5
#> [101] 31 31 23 43 57  3  6 59 55 22 58 29 33 12 32 63 25 11 61 29 37 62 26 21 57
#> [126] 61 25 26 34 54  7 10 13 11 18 56 33 23  2 34 45 24 20 64 24 30 57 50 39 16
#> [151] 60  8 56 62 45 43 61 53  8  4  9 51 56 15 49 11 10 10 41 26 49 36 63 13 26
#> [176] 46 13 13 38 54 14 27 10 58 55 29 25 39 27 22 60 32  8 59 25  8 23 56 57 24
is_new_episode(df$date, episode_days = 60) # TRUE/FALSE
#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE
#>  [13]  TRUE FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE  TRUE FALSE FALSE  TRUE
#>  [25] FALSE  TRUE FALSE  TRUE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE FALSE
#>  [49] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE
#>  [61]  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE
#>  [73] FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE
#>  [85]  TRUE  TRUE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE  TRUE
#>  [97] FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE  TRUE FALSE FALSE
#> [109]  TRUE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE FALSE  TRUE FALSE FALSE FALSE
#> [121] FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE  TRUE
#> [133] FALSE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE  TRUE FALSE  TRUE FALSE
#> [145] FALSE  TRUE FALSE  TRUE  TRUE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE
#> [157] FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#> [169]  TRUE FALSE FALSE  TRUE  TRUE FALSE FALSE  TRUE FALSE FALSE  TRUE  TRUE
#> [181] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [193] FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE

# filter on results from the third 60-day episode only, using base R
df[which(get_episode(df$date, 60) == 3), ]
#> # A tibble: 1 × 46
#>   date       patient   age gender ward     mo           PEN   OXA   FLC   AMX  
#>   <date>     <chr>   <dbl> <chr>  <chr>    <mo>         <rsi> <rsi> <rsi> <rsi>
#> 1 2002-07-21 955940     82 F      Clinical B_PSDMN_AERG R     NA    NA    R    
#> # … with 36 more variables: AMC <rsi>, AMP <rsi>, TZP <rsi>, CZO <rsi>,
#> #   FEP <rsi>, CXM <rsi>, FOX <rsi>, CTX <rsi>, CAZ <rsi>, CRO <rsi>,
#> #   GEN <rsi>, TOB <rsi>, AMK <rsi>, KAN <rsi>, TMP <rsi>, SXT <rsi>,
#> #   NIT <rsi>, FOS <rsi>, LNZ <rsi>, CIP <rsi>, MFX <rsi>, VAN <rsi>,
#> #   TEC <rsi>, TCY <rsi>, TGC <rsi>, DOX <rsi>, ERY <rsi>, CLI <rsi>,
#> #   AZM <rsi>, IPM <rsi>, MEM <rsi>, MTR <rsi>, CHL <rsi>, COL <rsi>,
#> #   MUP <rsi>, RIF <rsi>

# the functions also work for less than a day, e.g. to include one per hour:
get_episode(c(
  Sys.time(),
  Sys.time() + 60 * 60
),
episode_days = 1 / 24
)
#> [1] 1 2

# \donttest{
if (require("dplyr")) {
  # is_new_episode() can also be used in dplyr verbs to determine patient
  # episodes based on any (combination of) grouping variables:
  df %>%
    mutate(condition = sample(
      x = c("A", "B", "C"),
      size = 200,
      replace = TRUE
    )) %>%
    group_by(condition) %>%
    mutate(new_episode = is_new_episode(date, 365)) %>%
    select(patient, date, condition, new_episode)
}
#> # A tibble: 200 × 4
#> # Groups:   condition [3]
#>    patient date       condition new_episode
#>    <chr>   <date>     <chr>     <lgl>      
#>  1 671620  2016-12-04 B         FALSE      
#>  2 BF4515  2016-09-25 C         FALSE      
#>  3 116866  2011-11-07 B         FALSE      
#>  4 422833  2017-05-26 B         FALSE      
#>  5 344184  2006-08-01 A         FALSE      
#>  6 92F410  2007-09-03 C         TRUE       
#>  7 534816  2007-04-16 B         TRUE       
#>  8 088256  2003-01-25 B         TRUE       
#>  9 959835  2013-11-16 A         TRUE       
#> 10 255339  2013-07-22 B         FALSE      
#> # … with 190 more rows
if (require("dplyr")) {
  df %>%
    group_by(ward, patient) %>%
    transmute(date,
      patient,
      new_index = get_episode(date, 60),
      new_logical = is_new_episode(date, 60)
    )
}
#> # A tibble: 200 × 5
#> # Groups:   ward, patient [182]
#>    ward     date       patient new_index new_logical
#>    <chr>    <date>     <chr>       <dbl> <lgl>      
#>  1 Clinical 2016-12-04 671620          1 TRUE       
#>  2 Clinical 2016-09-25 BF4515          1 TRUE       
#>  3 Clinical 2011-11-07 116866          1 TRUE       
#>  4 Clinical 2017-05-26 422833          1 TRUE       
#>  5 Clinical 2006-08-01 344184          1 TRUE       
#>  6 Clinical 2007-09-03 92F410          1 TRUE       
#>  7 Clinical 2007-04-16 534816          1 TRUE       
#>  8 ICU      2003-01-25 088256          1 TRUE       
#>  9 Clinical 2013-11-16 959835          1 TRUE       
#> 10 Clinical 2013-07-22 255339          1 TRUE       
#> # … with 190 more rows
if (require("dplyr")) {
  df %>%
    group_by(ward) %>%
    summarise(
      n_patients = n_distinct(patient),
      n_episodes_365 = sum(is_new_episode(date, episode_days = 365)),
      n_episodes_60 = sum(is_new_episode(date, episode_days = 60)),
      n_episodes_30 = sum(is_new_episode(date, episode_days = 30))
    )
}
#> # A tibble: 3 × 5
#>   ward       n_patients n_episodes_365 n_episodes_60 n_episodes_30
#>   <chr>           <int>          <int>         <int>         <int>
#> 1 Clinical          127             15            55            79
#> 2 ICU                44             11            31            34
#> 3 Outpatient         11              6             9            10
if (require("dplyr")) {

  # grouping on patients and microorganisms leads to the same
  # results as first_isolate() when using 'episode-based':
  x <- df %>%
    filter_first_isolate(
      include_unknown = TRUE,
      method = "episode-based"
    )

  y <- df %>%
    group_by(patient, mo) %>%
    filter(is_new_episode(date, 365)) %>%
    ungroup()

  identical(x, y)
}
#> Including isolates from ICU.
#> [1] FALSE
if (require("dplyr")) {

  # but is_new_episode() has a lot more flexibility than first_isolate(),
  # since you can now group on anything that seems relevant:
  df %>%
    group_by(patient, mo, ward) %>%
    mutate(flag_episode = is_new_episode(date, 365)) %>%
    select(group_vars(.), flag_episode)
}
#> # A tibble: 200 × 4
#> # Groups:   patient, mo, ward [194]
#>    patient mo            ward     flag_episode
#>    <chr>   <mo>          <chr>    <lgl>       
#>  1 671620  B_ESCHR_COLI  Clinical TRUE        
#>  2 BF4515  B_STPHY_EPDR  Clinical TRUE        
#>  3 116866  B_STPHY_CONS  Clinical TRUE        
#>  4 422833  B_ENTRC_FCLS  Clinical TRUE        
#>  5 344184  B_STPHY_CONS  Clinical TRUE        
#>  6 92F410  B_ENTRBC_CLOC Clinical TRUE        
#>  7 534816  B_ENTRC_FCLS  Clinical TRUE        
#>  8 088256  B_STPHY_HMNS  ICU      TRUE        
#>  9 959835  B_STPHY_CONS  Clinical TRUE        
#> 10 255339  B_ESCHR_COLI  Clinical TRUE        
#> # … with 190 more rows
# }