Determine (New) Episodes for Patients

These functions determine which items in a vector can be considered (the start of) a new episode, based on the argument episode_days. This can be used to determine clinical episodes for any epidemiological analysis. The get_episode() function returns the index number of the episode per group, while the is_new_episode() function returns values TRUE/FALSE to indicate whether an item in a vector is the start of a new episode.

Usage

get_episode(x, episode_days, ...)

is_new_episode(x, episode_days, ...)

Arguments

x

vector of dates (class Date or POSIXt), will be sorted internally to determine episodes

episode_days

required episode length in days, can also be less than a day or Inf, see Details

...

ignored, only in place to allow future extensions

Value

get_episode(): a double vector

is_new_episode(): a logical vector

Details

Dates are first sorted from old to new. The oldest date will mark the start of the first episode. After this date, the next date will be marked that is at least episode_days days later than the start of the first episode. From that second marked date on, the next date will be marked that is at least episode_days days later than the start of the second episode which will be the start of the third episode, and so on. Before the vector is being returned, the original order will be restored.

The first_isolate() function is a wrapper around the is_new_episode() function, but is more efficient for data sets containing microorganism codes or names and allows for different isolate selection methods.

The dplyr package is not required for these functions to work, but these functions do support variable grouping and work conveniently inside dplyr verbs such as filter(), mutate() and summarise().

See also

Examples

# `example_isolates` is a data set available in the AMR package.
# See ?example_isolates
df <- example_isolates[sample(seq_len(2000), size = 200), ]

get_episode(df$date, episode_days = 60) # indices
#>   [1] 30 12  7 57  9 44 25 31  3 56  6 43 38 46 39 12 53 24 46 60 48 28 31  8 29
#>  [26] 41 38 35 37 17 38 58 35 56 18 65 31 32 28 56 37 30 51 17 62 34 30  2 21 42
#>  [51] 12 62  1 48 63 20 27 22 37 10 47  5 13 23 38 65 15 21 12 59 32 57 20 41 64
#>  [76] 26 12 38 44 42 55 20 57 45  8 29 19  5  8 48 19 61 55  7 49  2 57 20 18 63
#> [101]  3 26 55 22 37 62 38 54 58 60 18 15 18 15  4 40 20 47 42 60 18  7 38  3 51
#> [126]  8 24 33 11 46 16 63 58  7 21 48 50 22 29  4 40 25 13 43 48  9 36 14 10  5
#> [151] 44  8  3  1 21 15 22 46  4 10 61 15 39 22 28 58  1 16  8 60 36 19 49 53 61
#> [176]  6 55 29 17 63 26 14 23 52 23 18  6 63 64 18  9 41  8 52 39  3 23 54 10 23
is_new_episode(df$date, episode_days = 60) # TRUE/FALSE
#>   [1] FALSE FALSE  TRUE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE  TRUE FALSE  TRUE FALSE  TRUE
#>  [25] FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE
#>  [37] FALSE FALSE FALSE  TRUE FALSE  TRUE  TRUE FALSE  TRUE  TRUE FALSE  TRUE
#>  [49]  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE
#>  [61] FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE FALSE
#>  [73] FALSE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE
#>  [85] FALSE  TRUE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97]  TRUE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE
#> [109] FALSE FALSE  TRUE FALSE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [121] FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE
#> [133] FALSE FALSE FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE
#> [145] FALSE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#> [157] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE  TRUE
#> [169] FALSE FALSE FALSE  TRUE  TRUE FALSE  TRUE  TRUE FALSE FALSE FALSE  TRUE
#> [181] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
#> [193] FALSE  TRUE  TRUE FALSE FALSE  TRUE FALSE FALSE

# filter on results from the third 60-day episode only, using base R
df[which(get_episode(df$date, 60) == 3), ]
#> # A tibble: 5 × 46
#>   date       patient   age gender ward     mo           PEN   OXA   FLC   AMX  
#>   <date>     <chr>   <dbl> <chr>  <chr>    <mo>         <rsi> <rsi> <rsi> <rsi>
#> 1 2002-11-14 058917     76 F      ICU      B_STPHY_HMNS R     NA    S     NA   
#> 2 2002-11-12 CF9318     29 M      ICU      B_STPHY_CONS R     NA    R     R    
#> 3 2002-11-16 762305     87 F      Clinical B_PROTS_MRBL R     NA    NA    NA   
#> 4 2002-11-18 956065     89 F      Clinical B_ESCHR_COLI R     NA    NA    NA   
#> 5 2002-11-18 956065     89 F      Clinical B_ESCHR_COLI R     NA    NA    NA   
#> # … with 36 more variables: AMC <rsi>, AMP <rsi>, TZP <rsi>, CZO <rsi>,
#> #   FEP <rsi>, CXM <rsi>, FOX <rsi>, CTX <rsi>, CAZ <rsi>, CRO <rsi>,
#> #   GEN <rsi>, TOB <rsi>, AMK <rsi>, KAN <rsi>, TMP <rsi>, SXT <rsi>,
#> #   NIT <rsi>, FOS <rsi>, LNZ <rsi>, CIP <rsi>, MFX <rsi>, VAN <rsi>,
#> #   TEC <rsi>, TCY <rsi>, TGC <rsi>, DOX <rsi>, ERY <rsi>, CLI <rsi>,
#> #   AZM <rsi>, IPM <rsi>, MEM <rsi>, MTR <rsi>, CHL <rsi>, COL <rsi>,
#> #   MUP <rsi>, RIF <rsi>

# the functions also work for less than a day, e.g. to include one per hour:
get_episode(c(
  Sys.time(),
  Sys.time() + 60 * 60
),
episode_days = 1 / 24
)
#> [1] 1 2

# \donttest{
if (require("dplyr")) {
  # is_new_episode() can also be used in dplyr verbs to determine patient
  # episodes based on any (combination of) grouping variables:
  df %>%
    mutate(condition = sample(
      x = c("A", "B", "C"),
      size = 200,
      replace = TRUE
    )) %>%
    group_by(condition) %>%
    mutate(new_episode = is_new_episode(date, 365)) %>%
    select(patient, date, condition, new_episode)
}
#> # A tibble: 200 × 4
#> # Groups:   condition [3]
#>    patient date       condition new_episode
#>    <chr>   <date>     <chr>     <lgl>      
#>  1 F68330  2009-04-09 A         TRUE       
#>  2 F35553  2004-12-29 C         FALSE      
#>  3 218456  2003-09-26 A         FALSE      
#>  4 438196  2016-02-16 B         FALSE      
#>  5 386739  2004-04-15 B         FALSE      
#>  6 C541BC  2012-09-03 C         FALSE      
#>  7 192790  2008-02-15 C         FALSE      
#>  8 008218  2009-04-27 C         FALSE      
#>  9 058917  2002-11-14 A         FALSE      
#> 10 375975  2015-11-19 A         FALSE      
#> # … with 190 more rows
if (require("dplyr")) {
  df %>%
    group_by(ward, patient) %>%
    transmute(date,
      patient,
      new_index = get_episode(date, 60),
      new_logical = is_new_episode(date, 60)
    )
}
#> # A tibble: 200 × 5
#> # Groups:   ward, patient [184]
#>    ward       date       patient new_index new_logical
#>    <chr>      <date>     <chr>       <dbl> <lgl>      
#>  1 Clinical   2009-04-09 F68330          1 TRUE       
#>  2 ICU        2004-12-29 F35553          4 TRUE       
#>  3 Clinical   2003-09-26 218456          1 TRUE       
#>  4 ICU        2016-02-16 438196          1 TRUE       
#>  5 ICU        2004-04-15 386739          1 TRUE       
#>  6 Clinical   2012-09-03 C541BC          1 TRUE       
#>  7 Outpatient 2008-02-15 192790          1 TRUE       
#>  8 Clinical   2009-04-27 008218          1 TRUE       
#>  9 ICU        2002-11-14 058917          1 TRUE       
#> 10 Clinical   2015-11-19 375975          1 TRUE       
#> # … with 190 more rows
if (require("dplyr")) {
  df %>%
    group_by(ward) %>%
    summarise(
      n_patients = n_distinct(patient),
      n_episodes_365 = sum(is_new_episode(date, episode_days = 365)),
      n_episodes_60 = sum(is_new_episode(date, episode_days = 60)),
      n_episodes_30 = sum(is_new_episode(date, episode_days = 30))
    )
}
#> # A tibble: 3 × 5
#>   ward       n_patients n_episodes_365 n_episodes_60 n_episodes_30
#>   <chr>           <int>          <int>         <int>         <int>
#> 1 Clinical          118             15            55            72
#> 2 ICU                54             12            35            40
#> 3 Outpatient         12              7            10            11
if (require("dplyr")) {

  # grouping on patients and microorganisms leads to the same
  # results as first_isolate() when using 'episode-based':
  x <- df %>%
    filter_first_isolate(
      include_unknown = TRUE,
      method = "episode-based"
    )

  y <- df %>%
    group_by(patient, mo) %>%
    filter(is_new_episode(date, 365)) %>%
    ungroup()

  identical(x, y)
}
#> Including isolates from ICU.
#> [1] FALSE
if (require("dplyr")) {

  # but is_new_episode() has a lot more flexibility than first_isolate(),
  # since you can now group on anything that seems relevant:
  df %>%
    group_by(patient, mo, ward) %>%
    mutate(flag_episode = is_new_episode(date, 365)) %>%
    select(group_vars(.), flag_episode)
}
#> # A tibble: 200 × 4
#> # Groups:   patient, mo, ward [192]
#>    patient mo           ward       flag_episode
#>    <chr>   <mo>         <chr>      <lgl>       
#>  1 F68330  B_STRPT_PNMN Clinical   TRUE        
#>  2 F35553  B_SERRT_MRCS ICU        TRUE        
#>  3 218456  B_PROTS_MRBL Clinical   TRUE        
#>  4 438196  B_STPHY_AURS ICU        TRUE        
#>  5 386739  B_ESCHR_COLI ICU        TRUE        
#>  6 C541BC  B_ESCHR_COLI Clinical   TRUE        
#>  7 192790  B_STPHY_CONS Outpatient TRUE        
#>  8 008218  B_STRPT_GRPA Clinical   TRUE        
#>  9 058917  B_STPHY_HMNS ICU        TRUE        
#> 10 375975  B_STPHY_AURS Clinical   TRUE        
#> # … with 190 more rows
# }