Determine (New) Episodes for Patients

These functions determine which items in a vector can be considered (the start of) a new episode, based on the argument episode_days. This can be used to determine clinical episodes for any epidemiological analysis. The get_episode() function returns the index number of the episode per group, while the is_new_episode() function returns values TRUE/FALSE to indicate whether an item in a vector is the start of a new episode.

Usage

get_episode(x, episode_days, ...)

is_new_episode(x, episode_days, ...)

Arguments

x

vector of dates (class Date or POSIXt), will be sorted internally to determine episodes

episode_days

required episode length in days, can also be less than a day or Inf, see Details

...

ignored, only in place to allow future extensions

Value

get_episode(): a double vector

is_new_episode(): a logical vector

Details

Dates are first sorted from old to new. The oldest date will mark the start of the first episode. After this date, the next date will be marked that is at least episode_days days later than the start of the first episode. From that second marked date on, the next date will be marked that is at least episode_days days later than the start of the second episode which will be the start of the third episode, and so on. Before the vector is being returned, the original order will be restored.

The first_isolate() function is a wrapper around the is_new_episode() function, but is more efficient for data sets containing microorganism codes or names and allows for different isolate selection methods.

The dplyr package is not required for these functions to work, but these functions do support variable grouping and work conveniently inside dplyr verbs such as filter(), mutate() and summarise().

See also

Examples

# `example_isolates` is a data set available in the AMR package.
# See ?example_isolates
df <- example_isolates[sample(seq_len(2000), size = 200), ]

get_episode(df$date, episode_days = 60) # indices
#>   [1] 10 21 49  7 48 16 36 11 37 21 63 39 16 44 22 61 14 54 44 23 12 40  8  1 49
#>  [26] 42  8 52 64 26 13 52  2 14  1 22 58 63 33 53 14 31 31  3  4 25 22 29 31 64
#>  [51] 66 17 17 29 46 41 14  6  1 10 60 18 38 20  6 33 37  9 64 14 47 59 64 39 43
#>  [76] 29 50 50  6 59 21 40  4 15 62 46 59 54 12  8 55 16 15 31 59 58 33 66 36 30
#> [101] 27 14 42 21 49  8  3 49 42 25 19 12 63 45 51 13 45 32 56 52 24  5 35 20 43
#> [126] 65 34 14 11 66 47 36 33 62 38 57  1 47 12 66 10 25 11 33 57 66 54 17 23 12
#> [151] 14  4 43 48 55 44 58 65 63  4 19 11 36 38 13 66 36 18 19  8 24 39 64 14 49
#> [176] 62 19 41  8 51 48 19 35  4 28  3 61 62 44 11 52  5  5 34 40 35 29 35 10 20
is_new_episode(df$date, episode_days = 60) # TRUE/FALSE
#>   [1] FALSE FALSE FALSE  TRUE FALSE  TRUE  TRUE FALSE FALSE  TRUE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE  TRUE FALSE  TRUE  TRUE  TRUE FALSE  TRUE
#>  [37]  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49]  TRUE FALSE FALSE  TRUE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE  TRUE
#>  [61]  TRUE FALSE FALSE  TRUE FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE
#>  [73]  TRUE  TRUE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE  TRUE
#>  [85]  TRUE  TRUE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#>  [97]  TRUE FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE
#> [109]  TRUE  TRUE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE  TRUE  TRUE FALSE
#> [121]  TRUE FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE
#> [133] FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE
#> [145]  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE  TRUE FALSE  TRUE
#> [157] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE
#> [169]  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
#> [181] FALSE FALSE FALSE  TRUE  TRUE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE
#> [193]  TRUE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE

# filter on results from the third 60-day episode only, using base R
df[which(get_episode(df$date, 60) == 3), ]
#> # A tibble: 3 × 46
#>   date       patient   age gender ward     mo           PEN   OXA   FLC   AMX  
#>   <date>     <chr>   <dbl> <chr>  <chr>    <mo>         <rsi> <rsi> <rsi> <rsi>
#> 1 2002-08-19 A49852     70 M      Clinical B_ESCHR_COLI R     NA    NA    NA   
#> 2 2002-07-30 218912     76 F      ICU      B_ESCHR_COLI R     NA    NA    NA   
#> 3 2002-08-31 149442     80 F      ICU      B_STPHY_AURS R     NA    S     R    
#> # … with 36 more variables: AMC <rsi>, AMP <rsi>, TZP <rsi>, CZO <rsi>,
#> #   FEP <rsi>, CXM <rsi>, FOX <rsi>, CTX <rsi>, CAZ <rsi>, CRO <rsi>,
#> #   GEN <rsi>, TOB <rsi>, AMK <rsi>, KAN <rsi>, TMP <rsi>, SXT <rsi>,
#> #   NIT <rsi>, FOS <rsi>, LNZ <rsi>, CIP <rsi>, MFX <rsi>, VAN <rsi>,
#> #   TEC <rsi>, TCY <rsi>, TGC <rsi>, DOX <rsi>, ERY <rsi>, CLI <rsi>,
#> #   AZM <rsi>, IPM <rsi>, MEM <rsi>, MTR <rsi>, CHL <rsi>, COL <rsi>,
#> #   MUP <rsi>, RIF <rsi>

# the functions also work for less than a day, e.g. to include one per hour:
get_episode(c(
  Sys.time(),
  Sys.time() + 60 * 60
),
episode_days = 1 / 24
)
#> [1] 1 2

# \donttest{
if (require("dplyr")) {
  # is_new_episode() can also be used in dplyr verbs to determine patient
  # episodes based on any (combination of) grouping variables:
  df %>%
    mutate(condition = sample(
      x = c("A", "B", "C"),
      size = 200,
      replace = TRUE
    )) %>%
    group_by(condition) %>%
    mutate(new_episode = is_new_episode(date, 365)) %>%
    select(patient, date, condition, new_episode)
}
#> # A tibble: 200 × 4
#> # Groups:   condition [3]
#>    patient date       condition new_episode
#>    <chr>   <date>     <chr>     <lgl>      
#>  1 419655  2004-02-21 C         FALSE      
#>  2 F3F09F  2006-06-03 A         FALSE      
#>  3 545388  2013-12-05 B         FALSE      
#>  4 E45CF0  2003-07-06 B         FALSE      
#>  5 992282  2013-10-01 A         TRUE       
#>  6 A92398  2005-05-14 B         FALSE      
#>  7 E58549  2010-03-06 C         TRUE       
#>  8 529296  2004-06-09 A         FALSE      
#>  9 F54287  2010-07-05 C         FALSE      
#> 10 578848  2006-05-08 C         FALSE      
#> # … with 190 more rows
if (require("dplyr")) {
  df %>%
    group_by(ward, patient) %>%
    transmute(date,
      patient,
      new_index = get_episode(date, 60),
      new_logical = is_new_episode(date, 60)
    )
}
#> # A tibble: 200 × 5
#> # Groups:   ward, patient [177]
#>    ward     date       patient new_index new_logical
#>    <chr>    <date>     <chr>       <dbl> <lgl>      
#>  1 ICU      2004-02-21 419655          1 TRUE       
#>  2 Clinical 2006-06-03 F3F09F          1 TRUE       
#>  3 Clinical 2013-12-05 545388          1 TRUE       
#>  4 Clinical 2003-07-06 E45CF0          1 TRUE       
#>  5 Clinical 2013-10-01 992282          1 TRUE       
#>  6 Clinical 2005-05-14 A92398          1 TRUE       
#>  7 Clinical 2010-03-06 E58549          1 TRUE       
#>  8 ICU      2004-06-09 529296          1 TRUE       
#>  9 Clinical 2010-07-05 F54287          1 TRUE       
#> 10 Clinical 2006-05-08 578848          1 TRUE       
#> # … with 190 more rows
if (require("dplyr")) {
  df %>%
    group_by(ward) %>%
    summarise(
      n_patients = n_distinct(patient),
      n_episodes_365 = sum(is_new_episode(date, episode_days = 365)),
      n_episodes_60 = sum(is_new_episode(date, episode_days = 60)),
      n_episodes_30 = sum(is_new_episode(date, episode_days = 30))
    )
}
#> # A tibble: 3 × 5
#>   ward       n_patients n_episodes_365 n_episodes_60 n_episodes_30
#>   <chr>           <int>          <int>         <int>         <int>
#> 1 Clinical          113             15            54            71
#> 2 ICU                48             13            30            38
#> 3 Outpatient         16              7            13            15
if (require("dplyr")) {

  # grouping on patients and microorganisms leads to the same
  # results as first_isolate() when using 'episode-based':
  x <- df %>%
    filter_first_isolate(
      include_unknown = TRUE,
      method = "episode-based"
    )

  y <- df %>%
    group_by(patient, mo) %>%
    filter(is_new_episode(date, 365)) %>%
    ungroup()

  identical(x, y)
}
#> Including isolates from ICU.
#> [1] FALSE
if (require("dplyr")) {

  # but is_new_episode() has a lot more flexibility than first_isolate(),
  # since you can now group on anything that seems relevant:
  df %>%
    group_by(patient, mo, ward) %>%
    mutate(flag_episode = is_new_episode(date, 365)) %>%
    select(group_vars(.), flag_episode)
}
#> # A tibble: 200 × 4
#> # Groups:   patient, mo, ward [188]
#>    patient mo           ward     flag_episode
#>    <chr>   <mo>         <chr>    <lgl>       
#>  1 419655  B_STRPT_PNMN ICU      TRUE        
#>  2 F3F09F  B_STRPT_ORLS Clinical TRUE        
#>  3 545388  B_STPHY_HMNS Clinical TRUE        
#>  4 E45CF0  B_KLBSL_PNMN Clinical TRUE        
#>  5 992282  B_STRPT_GRPA Clinical TRUE        
#>  6 A92398  B_ESCHR_COLI Clinical TRUE        
#>  7 E58549  B_ESCHR_COLI Clinical TRUE        
#>  8 529296  B_ENTRC_FACM ICU      TRUE        
#>  9 F54287  B_STRPT_ANGN Clinical TRUE        
#> 10 578848  B_ESCHR_COLI Clinical TRUE        
#> # … with 190 more rows
# }