Determine (Clinical) Episodes

These functions determine which items in a vector can be considered (the start of) a new episode, based on the argument episode_days. This can be used to determine clinical episodes for any epidemiological analysis. The get_episode() function returns the index number of the episode per group, while the is_new_episode() function returns TRUE for every new get_episode() index, and is thus equal to !duplicated(get_episode(...)).

Usage

get_episode(x, episode_days, ...)

is_new_episode(x, episode_days, ...)

Arguments

x

vector of dates (class Date or POSIXt), will be sorted internally to determine episodes

episode_days

required episode length in days, can also be less than a day or Inf, see Details

...

ignored, only in place to allow future extensions

Value

get_episode(): an integer vector

is_new_episode(): a logical vector

Details

The functions get_episode() and is_new_episode() differ in this way when setting episode_days to 365:

person_id	date	get_episode()	is_new_episode()

A	2019-01-01	1	TRUE

A	2019-03-01	1	FALSE

A	2021-01-01	2	TRUE

B	2008-01-01	1	TRUE

B	2008-01-01	1	FALSE

C	2020-01-01	1	TRUE

Dates are first sorted from old to new. The oldest date will mark the start of the first episode. After this date, the next date will be marked that is at least episode_days days later than the start of the first episode. From that second marked date on, the next date will be marked that is at least episode_days days later than the start of the second episode which will be the start of the third episode, and so on. Before the vector is being returned, the original order will be restored.

The first_isolate() function is a wrapper around the is_new_episode() function, but is more efficient for data sets containing microorganism codes or names and allows for different isolate selection methods.

The dplyr package is not required for these functions to work, but these episode functions do support variable grouping and work conveniently inside dplyr verbs such as filter(), mutate() and summarise().

See also

Examples

# `example_isolates` is a data set available in the AMR package.
# See ?example_isolates
df <- example_isolates[sample(seq_len(2000), size = 100), ]

get_episode(df$date, episode_days = 60) # indices
#>   [1] 35 16 14 38  5  7 10 13  8 26 33  1 21 23 50 43 19 21 21 18 10 31 42 49  7
#>  [26] 18 14  9 45 24 21 47 40 44 48 35 27 31 12 21 15 28 37 10 25 35 44 18 29 46
#>  [51] 18 42 18 30 17 31 45  3 20 43 25 40 38 47 10 10 22 34 19 36 13 11 41 49 12
#>  [76] 12 39 32 15 35 22  7 44 26  4 50 51 16 42 26  6 35 10 14  2 38 14 27 41 38
is_new_episode(df$date, episode_days = 60) # TRUE/FALSE
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE  TRUE FALSE  TRUE  TRUE  TRUE
#>  [25] FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE FALSE
#>  [37]  TRUE FALSE  TRUE FALSE  TRUE  TRUE  TRUE FALSE  TRUE FALSE FALSE FALSE
#>  [49]  TRUE  TRUE FALSE FALSE FALSE  TRUE  TRUE FALSE FALSE  TRUE  TRUE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE FALSE  TRUE FALSE  TRUE
#>  [73]  TRUE FALSE FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85]  TRUE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE
#>  [97] FALSE FALSE FALSE FALSE

# filter on results from the third 60-day episode only, using base R
df[which(get_episode(df$date, 60) == 3), ]
#> # A tibble: 1 × 46
#>   date       patient   age gender ward  mo           PEN   OXA   FLC   AMX  
#>   <date>     <chr>   <dbl> <chr>  <chr> <mo>         <sir> <sir> <sir> <sir>
#> 1 2002-07-23 F35553     51 M      ICU   B_STPHY_AURS R     NA    S     R    
#> # … with 36 more variables: AMC <sir>, AMP <sir>, TZP <sir>, CZO <sir>,
#> #   FEP <sir>, CXM <sir>, FOX <sir>, CTX <sir>, CAZ <sir>, CRO <sir>,
#> #   GEN <sir>, TOB <sir>, AMK <sir>, KAN <sir>, TMP <sir>, SXT <sir>,
#> #   NIT <sir>, FOS <sir>, LNZ <sir>, CIP <sir>, MFX <sir>, VAN <sir>,
#> #   TEC <sir>, TCY <sir>, TGC <sir>, DOX <sir>, ERY <sir>, CLI <sir>,
#> #   AZM <sir>, IPM <sir>, MEM <sir>, MTR <sir>, CHL <sir>, COL <sir>,
#> #   MUP <sir>, RIF <sir>

# the functions also work for less than a day, e.g. to include one per hour:
get_episode(
  c(
    Sys.time(),
    Sys.time() + 60 * 60
  ),
  episode_days = 1 / 24
)
#> [1] 1 2

# \donttest{
if (require("dplyr")) {
  # is_new_episode() can also be used in dplyr verbs to determine patient
  # episodes based on any (combination of) grouping variables:
  df %>%
    mutate(condition = sample(
      x = c("A", "B", "C"),
      size = 100,
      replace = TRUE
    )) %>%
    group_by(patient, condition) %>%
    mutate(new_episode = is_new_episode(date, 365)) %>%
    select(patient, date, condition, new_episode) %>%
    arrange(patient, condition, date)
}
#> # A tibble: 100 × 4
#> # Groups:   patient, condition [96]
#>    patient date       condition new_episode
#>    <chr>   <date>     <chr>     <lgl>      
#>  1 008268  2007-02-20 B         TRUE       
#>  2 067927  2002-01-07 C         TRUE       
#>  3 0E2483  2007-04-06 A         TRUE       
#>  4 0F9638  2014-09-22 A         TRUE       
#>  5 124128  2006-10-02 C         TRUE       
#>  6 1435C8  2004-03-03 C         TRUE       
#>  7 151041  2006-02-10 C         TRUE       
#>  8 16DC39  2015-11-19 A         TRUE       
#>  9 180928  2008-06-14 A         TRUE       
#> 10 22B987  2009-10-19 C         TRUE       
#> # … with 90 more rows

if (require("dplyr")) {
  df %>%
    group_by(ward, patient) %>%
    transmute(date,
      patient,
      new_index = get_episode(date, 60),
      new_logical = is_new_episode(date, 60)
    ) %>%
    arrange(patient, ward, date)
}
#> # A tibble: 100 × 5
#> # Groups:   ward, patient [91]
#>    ward     date       patient new_index new_logical
#>    <chr>    <date>     <chr>       <int> <lgl>      
#>  1 ICU      2007-02-20 008268          1 TRUE       
#>  2 ICU      2002-01-07 067927          1 TRUE       
#>  3 Clinical 2007-04-06 0E2483          1 TRUE       
#>  4 Clinical 2014-09-22 0F9638          1 TRUE       
#>  5 Clinical 2006-10-02 124128          1 TRUE       
#>  6 Clinical 2004-03-03 1435C8          1 TRUE       
#>  7 Clinical 2006-02-10 151041          1 TRUE       
#>  8 ICU      2015-11-19 16DC39          1 TRUE       
#>  9 Clinical 2008-06-14 180928          1 TRUE       
#> 10 Clinical 2009-10-19 22B987          1 TRUE       
#> # … with 90 more rows

if (require("dplyr")) {
  df %>%
    group_by(ward) %>%
    summarise(
      n_patients = n_distinct(patient),
      n_episodes_365 = sum(is_new_episode(date, episode_days = 365)),
      n_episodes_60 = sum(is_new_episode(date, episode_days = 60)),
      n_episodes_30 = sum(is_new_episode(date, episode_days = 30))
    )
}
#> # A tibble: 3 × 5
#>   ward       n_patients n_episodes_365 n_episodes_60 n_episodes_30
#>   <chr>           <int>          <int>         <int>         <int>
#> 1 Clinical           67             13            40            48
#> 2 ICU                20             12            20            21
#> 3 Outpatient          4              4             4             4

# grouping on patients and microorganisms leads to the same
# results as first_isolate() when using 'episode-based':
if (require("dplyr")) {
  x <- df %>%
    filter_first_isolate(
      include_unknown = TRUE,
      method = "episode-based"
    )

  y <- df %>%
    group_by(patient, mo) %>%
    filter(is_new_episode(date, 365)) %>%
    ungroup()

  identical(x, y)
}
#> [1] TRUE

# but is_new_episode() has a lot more flexibility than first_isolate(),
# since you can now group on anything that seems relevant:
if (require("dplyr")) {
  df %>%
    group_by(patient, mo, ward) %>%
    mutate(flag_episode = is_new_episode(date, 365)) %>%
    select(group_vars(.), flag_episode)
}
#> # A tibble: 100 × 4
#> # Groups:   patient, mo, ward [94]
#>    patient mo           ward     flag_episode
#>    <chr>   <mo>         <chr>    <lgl>       
#>  1 8697C7  B_STPHY_CONS Clinical TRUE        
#>  2 F3F09F  B_STRPT_ORLS Clinical TRUE        
#>  3 A90606  B_STRPT_PNMN Clinical TRUE        
#>  4 435623  B_CTRBC_FRND ICU      TRUE        
#>  5 6BC362  B_CRYNB      ICU      TRUE        
#>  6 739C43  B_ESCHR_COLI Clinical TRUE        
#>  7 F58364  B_STRPT_MITS Clinical TRUE        
#>  8 969581  B_STPHY_CONS Clinical TRUE        
#>  9 8DB5B8  B_STPHY_CONS Clinical TRUE        
#> 10 882122  B_STRPT_GRPC Clinical TRUE        
#> # … with 90 more rows
# }