Determine (New) Episodes for Patients

These functions determine which items in a vector can be considered (the start of) a new episode, based on the argument episode_days. This can be used to determine clinical episodes for any epidemiological analysis. The get_episode() function returns the index number of the episode per group, while the is_new_episode() function returns values TRUE/FALSE to indicate whether an item in a vector is the start of a new episode.

Usage

get_episode(x, episode_days, ...)

is_new_episode(x, episode_days, ...)

Arguments

x

vector of dates (class Date or POSIXt), will be sorted internally to determine episodes

episode_days

required episode length in days, can also be less than a day or Inf, see Details

...

ignored, only in place to allow future extensions

Value

get_episode(): a double vector

is_new_episode(): a logical vector

Details

Dates are first sorted from old to new. The oldest date will mark the start of the first episode. After this date, the next date will be marked that is at least episode_days days later than the start of the first episode. From that second marked date on, the next date will be marked that is at least episode_days days later than the start of the second episode which will be the start of the third episode, and so on. Before the vector is being returned, the original order will be restored.

The first_isolate() function is a wrapper around the is_new_episode() function, but is more efficient for data sets containing microorganism codes or names and allows for different isolate selection methods.

The dplyr package is not required for these functions to work, but these functions do support variable grouping and work conveniently inside dplyr verbs such as filter(), mutate() and summarise().

See also

Examples

# `example_isolates` is a data set available in the AMR package.
# See ?example_isolates
df <- example_isolates[sample(seq_len(2000), size = 200), ]

get_episode(df$date, episode_days = 60) # indices
#>   [1] 16 17 44  6  5 17 43 61 27  6 39 12  2 55 15 47 40 62 45 26 62 62 33 17  1
#>  [26] 23 10 37 30 25 35 43 37  8  5 29  7 39 57 50 62  8 43  5 51 29 64 62 18 22
#>  [51] 49 56 37 32  9 31 29 33  6 24 40 32 12 62 62 17 40 23 57 64  1 54 15 59 38
#>  [76] 62 60 21 48  1 58 54 57  8  8 29 19 49 64 62 63 39 11  7 27 34 63 22 18 54
#> [101]  3 41 15 23  4  5  9 32  3 24 12 56 40 21 30 19 31 18 62 44  7 51 56  3 12
#> [126] 34 38 46  4 17 42 24 28 52 61 22 59 23 53  1  1 36 57 29  6  9 54 22 36 55
#> [151] 47 33 42  9 36 18 61 56  8 18 54 42 29 59 13  5 52  7 15  2 26  9 39 53 54
#> [176] 42 20 46 14 13 45 55  2 41 20 58  7 42 10 43  5 53  6  6 60 63  3 14  5 60
is_new_episode(df$date, episode_days = 60) # TRUE/FALSE
#>   [1]  TRUE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE  TRUE  TRUE
#>  [13] FALSE  TRUE FALSE  TRUE FALSE  TRUE  TRUE FALSE FALSE FALSE  TRUE FALSE
#>  [25] FALSE FALSE  TRUE FALSE FALSE  TRUE  TRUE FALSE  TRUE FALSE FALSE FALSE
#>  [37]  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE
#>  [49] FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE
#>  [73]  TRUE  TRUE  TRUE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE
#>  [85]  TRUE  TRUE FALSE FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE  TRUE FALSE
#>  [97] FALSE  TRUE  TRUE  TRUE FALSE  TRUE FALSE  TRUE  TRUE FALSE  TRUE  TRUE
#> [109]  TRUE  TRUE FALSE FALSE  TRUE FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE
#> [121] FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE  TRUE  TRUE FALSE
#> [133]  TRUE  TRUE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE
#> [145] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [157]  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE
#> [169] FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE
#> [181] FALSE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE
#> [193] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE

# filter on results from the third 60-day episode only, using base R
df[which(get_episode(df$date, 60) == 3), ]
#> # A tibble: 4 × 46
#>   date       patient   age gender ward     mo           PEN   OXA   FLC   AMX  
#>   <date>     <chr>   <dbl> <chr>  <chr>    <mo>         <rsi> <rsi> <rsi> <rsi>
#> 1 2002-06-23 798871     82 M      Clinical B_ENTRC_FCLS NA    NA    NA    NA   
#> 2 2002-06-05 24D393     20 F      Clinical B_ESCHR_COLI R     NA    NA    NA   
#> 3 2002-07-21 955940     82 F      Clinical B_PSDMN_AERG R     NA    NA    R    
#> 4 2002-07-21 955940     82 F      Clinical B_PSDMN_AERG R     NA    NA    R    
#> # … with 36 more variables: AMC <rsi>, AMP <rsi>, TZP <rsi>, CZO <rsi>,
#> #   FEP <rsi>, CXM <rsi>, FOX <rsi>, CTX <rsi>, CAZ <rsi>, CRO <rsi>,
#> #   GEN <rsi>, TOB <rsi>, AMK <rsi>, KAN <rsi>, TMP <rsi>, SXT <rsi>,
#> #   NIT <rsi>, FOS <rsi>, LNZ <rsi>, CIP <rsi>, MFX <rsi>, VAN <rsi>,
#> #   TEC <rsi>, TCY <rsi>, TGC <rsi>, DOX <rsi>, ERY <rsi>, CLI <rsi>,
#> #   AZM <rsi>, IPM <rsi>, MEM <rsi>, MTR <rsi>, CHL <rsi>, COL <rsi>,
#> #   MUP <rsi>, RIF <rsi>

# the functions also work for less than a day, e.g. to include one per hour:
get_episode(c(
  Sys.time(),
  Sys.time() + 60 * 60
),
episode_days = 1 / 24
)
#> [1] 1 2

# \donttest{
if (require("dplyr")) {
  # is_new_episode() can also be used in dplyr verbs to determine patient
  # episodes based on any (combination of) grouping variables:
  df %>%
    mutate(condition = sample(
      x = c("A", "B", "C"),
      size = 200,
      replace = TRUE
    )) %>%
    group_by(condition) %>%
    mutate(new_episode = is_new_episode(date, 365)) %>%
    select(patient, date, condition, new_episode)
}
#> # A tibble: 200 × 4
#> # Groups:   condition [3]
#>    patient date       condition new_episode
#>    <chr>   <date>     <chr>     <lgl>      
#>  1 2B1C03  2005-10-23 C         FALSE      
#>  2 5DF436  2006-01-04 A         FALSE      
#>  3 299616  2012-11-13 C         FALSE      
#>  4 328792  2003-05-28 B         FALSE      
#>  5 374815  2002-12-02 A         FALSE      
#>  6 968584  2005-12-29 A         TRUE       
#>  7 824233  2012-07-10 C         FALSE      
#>  8 E68281  2017-03-02 A         FALSE      
#>  9 195736  2008-08-28 C         TRUE       
#> 10 F35553  2003-04-07 B         TRUE       
#> # … with 190 more rows
if (require("dplyr")) {
  df %>%
    group_by(ward, patient) %>%
    transmute(date,
      patient,
      new_index = get_episode(date, 60),
      new_logical = is_new_episode(date, 60)
    )
}
#> # A tibble: 200 × 5
#> # Groups:   ward, patient [180]
#>    ward       date       patient new_index new_logical
#>    <chr>      <date>     <chr>       <dbl> <lgl>      
#>  1 Clinical   2005-10-23 2B1C03          1 TRUE       
#>  2 ICU        2006-01-04 5DF436          1 TRUE       
#>  3 Clinical   2012-11-13 299616          1 TRUE       
#>  4 Clinical   2003-05-28 328792          1 TRUE       
#>  5 Clinical   2002-12-02 374815          1 TRUE       
#>  6 Clinical   2005-12-29 968584          1 TRUE       
#>  7 Clinical   2012-07-10 824233          1 TRUE       
#>  8 Clinical   2017-03-02 E68281          1 TRUE       
#>  9 Outpatient 2008-08-28 195736          1 TRUE       
#> 10 ICU        2003-04-07 F35553          2 TRUE       
#> # … with 190 more rows
if (require("dplyr")) {
  df %>%
    group_by(ward) %>%
    summarise(
      n_patients = n_distinct(patient),
      n_episodes_365 = sum(is_new_episode(date, episode_days = 365)),
      n_episodes_60 = sum(is_new_episode(date, episode_days = 60)),
      n_episodes_30 = sum(is_new_episode(date, episode_days = 30))
    )
}
#> # A tibble: 3 × 5
#>   ward       n_patients n_episodes_365 n_episodes_60 n_episodes_30
#>   <chr>           <int>          <int>         <int>         <int>
#> 1 Clinical          123             14            55            73
#> 2 ICU                42             10            27            34
#> 3 Outpatient         15              7            12            13
if (require("dplyr")) {

  # grouping on patients and microorganisms leads to the same
  # results as first_isolate() when using 'episode-based':
  x <- df %>%
    filter_first_isolate(
      include_unknown = TRUE,
      method = "episode-based"
    )

  y <- df %>%
    group_by(patient, mo) %>%
    filter(is_new_episode(date, 365)) %>%
    ungroup()

  identical(x, y)
}
#> Including isolates from ICU.
#> [1] FALSE
if (require("dplyr")) {

  # but is_new_episode() has a lot more flexibility than first_isolate(),
  # since you can now group on anything that seems relevant:
  df %>%
    group_by(patient, mo, ward) %>%
    mutate(flag_episode = is_new_episode(date, 365)) %>%
    select(group_vars(.), flag_episode)
}
#> # A tibble: 200 × 4
#> # Groups:   patient, mo, ward [188]
#>    patient mo           ward       flag_episode
#>    <chr>   <mo>         <chr>      <lgl>       
#>  1 2B1C03  B_STPHY_CONS Clinical   TRUE        
#>  2 5DF436  B_ENTRC      ICU        TRUE        
#>  3 299616  B_STPHY_AURS Clinical   TRUE        
#>  4 328792  B_STRPT_MITS Clinical   TRUE        
#>  5 374815  B_STPHY_CONS Clinical   TRUE        
#>  6 968584  B_PSDMN_AERG Clinical   TRUE        
#>  7 824233  B_STPHY_AURS Clinical   TRUE        
#>  8 E68281  B_STPHY_AURS Clinical   TRUE        
#>  9 195736  B_STPHY_AURS Outpatient TRUE        
#> 10 F35553  B_STPHY_AURS ICU        TRUE        
#> # … with 190 more rows
# }