Determine (New) Episodes for Patients

These functions determine which items in a vector can be considered (the start of) a new episode, based on the argument episode_days. This can be used to determine clinical episodes for any epidemiological analysis. The get_episode() function returns the index number of the episode per group, while the is_new_episode() function returns values TRUE/FALSE to indicate whether an item in a vector is the start of a new episode.

Usage

get_episode(x, episode_days, ...)

is_new_episode(x, episode_days, ...)

Arguments

x

vector of dates (class Date or POSIXt), will be sorted internally to determine episodes

episode_days

required episode length in days, can also be less than a day or Inf, see Details

...

ignored, only in place to allow future extensions

Value

get_episode(): a double vector

is_new_episode(): a logical vector

Details

Dates are first sorted from old to new. The oldest date will mark the start of the first episode. After this date, the next date will be marked that is at least episode_days days later than the start of the first episode. From that second marked date on, the next date will be marked that is at least episode_days days later than the start of the second episode which will be the start of the third episode, and so on. Before the vector is being returned, the original order will be restored.

The first_isolate() function is a wrapper around the is_new_episode() function, but is more efficient for data sets containing microorganism codes or names and allows for different isolate selection methods.

The dplyr package is not required for these functions to work, but these functions do support variable grouping and work conveniently inside dplyr verbs such as filter(), mutate() and summarise().

See also

Examples

# `example_isolates` is a data set available in the AMR package.
# See ?example_isolates
df <- example_isolates[sample(seq_len(2000), size = 200), ]

get_episode(df$date, episode_days = 60) # indices
#>   [1]  9 29 29 17 30 21  4 56 31 35 18 50 62  7 28 28 57 37 53 46 15 23 14 39 42
#>  [26] 46 20 62 29 54 18 54  1 61  3 12  6  8  9 43 24  1 61 37 11 42 28 36 12  8
#>  [51] 26 61 55 60  4 19 30  7 20  9 13 57  9 59 47  8 17  1 64  8  2  4 60  7 34
#>  [76] 61 44 20 61 55  7  2 32 33 18  6 18 11  7 10 18  5  3 44 23 54 31 18  4 19
#> [101] 25 17  7 41 43 17 55 21 30 61 11 16 14 57 52 41 40 23 43 60 37 26 12 12  4
#> [126] 10  6 40 11  2 45 30 51 47 29 25 58 10 61 19  8 39  1 24 25 58 10 18 34 20
#> [151] 63 43 57 51 60 41 36 64  4 61 55 22 19 27 53 64 51 15 55 23 22 36 45 57 48
#> [176] 40 44 49 57 38 40 29 32  2 11 50 38 53 10 33  8  9 54  4  6 58 57  3 14 60
is_new_episode(df$date, episode_days = 60) # TRUE/FALSE
#>   [1] FALSE  TRUE FALSE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#>  [13] FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE  TRUE
#>  [25] FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE  TRUE FALSE  TRUE FALSE FALSE
#>  [49] FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#>  [61]  TRUE FALSE  TRUE  TRUE  TRUE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE
#>  [73] FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
#>  [85] FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE
#> [109] FALSE  TRUE FALSE  TRUE FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE  TRUE
#> [121] FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#> [133] FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE
#> [145]  TRUE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#> [157] FALSE FALSE FALSE FALSE FALSE  TRUE FALSE  TRUE  TRUE FALSE  TRUE  TRUE
#> [169]  TRUE FALSE FALSE  TRUE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE FALSE FALSE
#> [181]  TRUE FALSE  TRUE  TRUE FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE
#> [193]  TRUE FALSE FALSE  TRUE FALSE  TRUE  TRUE FALSE

# filter on results from the third 60-day episode only, using base R
df[which(get_episode(df$date, 60) == 3), ]
#> # A tibble: 3 × 46
#>   date       patient   age gender ward     mo           PEN   OXA   FLC   AMX  
#>   <date>     <chr>   <dbl> <chr>  <chr>    <mo>         <sir> <sir> <sir> <sir>
#> 1 2002-08-28 390178     57 M      Clinical B_STRPT_SLVR S     NA    NA    S    
#> 2 2002-08-14 785317     51 F      ICU      B_ESCHR_COLI R     NA    NA    NA   
#> 3 2002-07-24 F35553     51 M      ICU      B_STPHY_AURS R     NA    S     R    
#> # … with 36 more variables: AMC <sir>, AMP <sir>, TZP <sir>, CZO <sir>,
#> #   FEP <sir>, CXM <sir>, FOX <sir>, CTX <sir>, CAZ <sir>, CRO <sir>,
#> #   GEN <sir>, TOB <sir>, AMK <sir>, KAN <sir>, TMP <sir>, SXT <sir>,
#> #   NIT <sir>, FOS <sir>, LNZ <sir>, CIP <sir>, MFX <sir>, VAN <sir>,
#> #   TEC <sir>, TCY <sir>, TGC <sir>, DOX <sir>, ERY <sir>, CLI <sir>,
#> #   AZM <sir>, IPM <sir>, MEM <sir>, MTR <sir>, CHL <sir>, COL <sir>,
#> #   MUP <sir>, RIF <sir>

# the functions also work for less than a day, e.g. to include one per hour:
get_episode(
  c(
    Sys.time(),
    Sys.time() + 60 * 60
  ),
  episode_days = 1 / 24
)
#> [1] 1 2

# \donttest{
if (require("dplyr")) {
  # is_new_episode() can also be used in dplyr verbs to determine patient
  # episodes based on any (combination of) grouping variables:
  df %>%
    mutate(condition = sample(
      x = c("A", "B", "C"),
      size = 200,
      replace = TRUE
    )) %>%
    group_by(condition) %>%
    mutate(new_episode = is_new_episode(date, 365)) %>%
    select(patient, date, condition, new_episode)
}
#> # A tibble: 200 × 4
#> # Groups:   condition [3]
#>    patient date       condition new_episode
#>    <chr>   <date>     <chr>     <lgl>      
#>  1 822116  2004-02-24 C         FALSE      
#>  2 144549  2009-09-01 C         FALSE      
#>  3 982666  2009-09-06 A         FALSE      
#>  4 5DF436  2006-01-04 B         TRUE       
#>  5 023456  2009-11-02 A         FALSE      
#>  6 725063  2007-06-07 A         FALSE      
#>  7 FCC668  2002-10-14 B         FALSE      
#>  8 D28985  2015-12-21 C         FALSE      
#>  9 B26404  2010-01-02 A         FALSE      
#> 10 F83217  2010-12-21 B         TRUE       
#> # … with 190 more rows
if (require("dplyr")) {
  df %>%
    group_by(ward, patient) %>%
    transmute(date,
      patient,
      new_index = get_episode(date, 60),
      new_logical = is_new_episode(date, 60)
    )
}
#> # A tibble: 200 × 5
#> # Groups:   ward, patient [183]
#>    ward       date       patient new_index new_logical
#>    <chr>      <date>     <chr>       <dbl> <lgl>      
#>  1 ICU        2004-02-24 822116          1 TRUE       
#>  2 Clinical   2009-09-01 144549          1 TRUE       
#>  3 Clinical   2009-09-06 982666          1 TRUE       
#>  4 ICU        2006-01-04 5DF436          1 TRUE       
#>  5 Clinical   2009-11-02 023456          1 TRUE       
#>  6 ICU        2007-06-07 725063          1 TRUE       
#>  7 ICU        2002-10-14 FCC668          1 TRUE       
#>  8 Outpatient 2015-12-21 D28985          1 TRUE       
#>  9 Clinical   2010-01-02 B26404          1 TRUE       
#> 10 ICU        2010-12-21 F83217          1 TRUE       
#> # … with 190 more rows
if (require("dplyr")) {
  df %>%
    group_by(ward) %>%
    summarise(
      n_patients = n_distinct(patient),
      n_episodes_365 = sum(is_new_episode(date, episode_days = 365)),
      n_episodes_60 = sum(is_new_episode(date, episode_days = 60)),
      n_episodes_30 = sum(is_new_episode(date, episode_days = 30))
    )
}
#> # A tibble: 3 × 5
#>   ward       n_patients n_episodes_365 n_episodes_60 n_episodes_30
#>   <chr>           <int>          <int>         <int>         <int>
#> 1 Clinical          118             14            51            70
#> 2 ICU                59             13            37            48
#> 3 Outpatient          6              5             6             6
if (require("dplyr")) {
  # grouping on patients and microorganisms leads to the same
  # results as first_isolate() when using 'episode-based':
  x <- df %>%
    filter_first_isolate(
      include_unknown = TRUE,
      method = "episode-based"
    )

  y <- df %>%
    group_by(patient, mo) %>%
    filter(is_new_episode(date, 365)) %>%
    ungroup()

  identical(x, y)
}
#> Including isolates from ICU.
#> [1] FALSE
if (require("dplyr")) {
  # but is_new_episode() has a lot more flexibility than first_isolate(),
  # since you can now group on anything that seems relevant:
  df %>%
    group_by(patient, mo, ward) %>%
    mutate(flag_episode = is_new_episode(date, 365)) %>%
    select(group_vars(.), flag_episode)
}
#> # A tibble: 200 × 4
#> # Groups:   patient, mo, ward [192]
#>    patient mo            ward       flag_episode
#>    <chr>   <mo>          <chr>      <lgl>       
#>  1 822116  B_STPHY_AURS  ICU        TRUE        
#>  2 144549  B_ENTRBC_CLOC Clinical   TRUE        
#>  3 982666  B_STRPT_DYSG  Clinical   TRUE        
#>  4 5DF436  B_STPHY_AURS  ICU        TRUE        
#>  5 023456  B_KLBSL_PNMN  Clinical   TRUE        
#>  6 725063  B_STPHY_CONS  ICU        TRUE        
#>  7 FCC668  B_ACNTB       ICU        TRUE        
#>  8 D28985  B_ESCHR_COLI  Outpatient TRUE        
#>  9 B26404  B_STPHY_CONS  Clinical   TRUE        
#> 10 F83217  B_STPHY_CONS  ICU        TRUE        
#> # … with 190 more rows
# }