Determine (New) Episodes for Patients

These functions determine which items in a vector can be considered (the start of) a new episode, based on the argument episode_days. This can be used to determine clinical episodes for any epidemiological analysis. The get_episode() function returns the index number of the episode per group, while the is_new_episode() function returns values TRUE/FALSE to indicate whether an item in a vector is the start of a new episode.

Usage

get_episode(x, episode_days, ...)

is_new_episode(x, episode_days, ...)

Arguments

x

vector of dates (class Date or POSIXt), will be sorted internally to determine episodes

episode_days

required episode length in days, can also be less than a day or Inf, see Details

...

ignored, only in place to allow future extensions

Value

get_episode(): a double vector

is_new_episode(): a logical vector

Details

Dates are first sorted from old to new. The oldest date will mark the start of the first episode. After this date, the next date will be marked that is at least episode_days days later than the start of the first episode. From that second marked date on, the next date will be marked that is at least episode_days days later than the start of the second episode which will be the start of the third episode, and so on. Before the vector is being returned, the original order will be restored.

The first_isolate() function is a wrapper around the is_new_episode() function, but is more efficient for data sets containing microorganism codes or names and allows for different isolate selection methods.

The dplyr package is not required for these functions to work, but these functions do support variable grouping and work conveniently inside dplyr verbs such as filter(), mutate() and summarise().

See also

Examples

# `example_isolates` is a data set available in the AMR package.
# See ?example_isolates
df <- example_isolates[sample(seq_len(2000), size = 200), ]

get_episode(df$date, episode_days = 60) # indices
#>   [1] 51  1 30 33 57 38 63 41 55 28 20 27 45 11 12 55 39  3 21 59 25 21  9  2 42
#>  [26]  5 49 24 11 32 51 28 21 38 57 36 11 54 24 36 60  6 50 31 60 47 43 20 54 10
#>  [51] 40 35 12 62 31 42 41 37  9 20  1  2 34 48 46 45 51 36 59 45 62 28 54 26  2
#>  [76] 44 49 28 19 17 10 62 40  6  8 38 51 32  5 37 46  7 61 51 38 27 52 33 58 39
#> [101] 38  1 41  7 48 43  9 17 14 31 16 60 19 54 54 28 24 31 54 35 54  3  6 15 18
#> [126] 48 34 34 60 29 41 12 19 57 54 50 45 17 56 62 10 28 13 28  4 33 53 54 47 22
#> [151] 56 31  9 49 40 61 33 11 50 27 23 24 44 15  4  8 30 48 28 61 62  3 40 60 26
#> [176] 21 26 25  4 13 10  3  8 54 63 48 64 64 44 42 47 58 53 14  8 52 22 13 17 23
is_new_episode(df$date, episode_days = 60) # TRUE/FALSE
#>   [1] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE
#>  [13]  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE  TRUE FALSE FALSE
#>  [37]  TRUE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE  TRUE  TRUE  TRUE
#>  [49] FALSE  TRUE FALSE  TRUE  TRUE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE
#>  [73] FALSE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
#>  [97]  TRUE  TRUE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE
#> [109]  TRUE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE
#> [121] FALSE  TRUE FALSE  TRUE  TRUE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE
#> [133] FALSE  TRUE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE
#> [145] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE
#> [157] FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE
#> [169] FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE
#> [181] FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE
#> [193]  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE

# filter on results from the third 60-day episode only, using base R
df[which(get_episode(df$date, 60) == 3), ]
#> # A tibble: 4 × 46
#>   date       patient   age gender ward      mo           PEN   OXA   FLC   AMX  
#>   <date>     <chr>   <dbl> <chr>  <chr>     <mo>         <rsi> <rsi> <rsi> <rsi>
#> 1 2002-10-11 974319     78 M      Outpatie… B_STPHY_EPDR S     NA    S     NA   
#> 2 2002-08-31 149442     80 F      ICU       B_STPHY_AURS R     NA    S     R    
#> 3 2002-10-18 E55128     57 F      ICU       B_STPHY_AURS R     NA    S     R    
#> 4 2002-09-23 F35553     51 M      ICU       B_STPHY_AURS S     NA    S     NA   
#> # … with 36 more variables: AMC <rsi>, AMP <rsi>, TZP <rsi>, CZO <rsi>,
#> #   FEP <rsi>, CXM <rsi>, FOX <rsi>, CTX <rsi>, CAZ <rsi>, CRO <rsi>,
#> #   GEN <rsi>, TOB <rsi>, AMK <rsi>, KAN <rsi>, TMP <rsi>, SXT <rsi>,
#> #   NIT <rsi>, FOS <rsi>, LNZ <rsi>, CIP <rsi>, MFX <rsi>, VAN <rsi>,
#> #   TEC <rsi>, TCY <rsi>, TGC <rsi>, DOX <rsi>, ERY <rsi>, CLI <rsi>,
#> #   AZM <rsi>, IPM <rsi>, MEM <rsi>, MTR <rsi>, CHL <rsi>, COL <rsi>,
#> #   MUP <rsi>, RIF <rsi>

# the functions also work for less than a day, e.g. to include one per hour:
get_episode(c(
  Sys.time(),
  Sys.time() + 60 * 60
),
episode_days = 1 / 24
)
#> [1] 1 2

# \donttest{
if (require("dplyr")) {
  # is_new_episode() can also be used in dplyr verbs to determine patient
  # episodes based on any (combination of) grouping variables:
  df %>%
    mutate(condition = sample(
      x = c("A", "B", "C"),
      size = 200,
      replace = TRUE
    )) %>%
    group_by(condition) %>%
    mutate(new_episode = is_new_episode(date, 365)) %>%
    select(patient, date, condition, new_episode)
}
#> # A tibble: 200 × 4
#> # Groups:   condition [3]
#>    patient date       condition new_episode
#>    <chr>   <date>     <chr>     <lgl>      
#>  1 D53062  2014-07-28 B         FALSE      
#>  2 495616  2002-01-17 C         TRUE       
#>  3 534816  2008-10-28 A         TRUE       
#>  4 D30712  2009-04-10 B         FALSE      
#>  5 092034  2016-08-07 C         FALSE      
#>  6 080086  2010-08-08 A         FALSE      
#>  7 470001  2017-10-04 C         FALSE      
#>  8 F15984  2011-04-01 B         FALSE      
#>  9 D41749  2015-12-15 A         FALSE      
#> 10 945BD5  2008-05-06 A         FALSE      
#> # … with 190 more rows
if (require("dplyr")) {
  df %>%
    group_by(ward, patient) %>%
    transmute(date,
      patient,
      new_index = get_episode(date, 60),
      new_logical = is_new_episode(date, 60)
    )
}
#> # A tibble: 200 × 5
#> # Groups:   ward, patient [185]
#>    ward       date       patient new_index new_logical
#>    <chr>      <date>     <chr>       <dbl> <lgl>      
#>  1 Clinical   2014-07-28 D53062          1 TRUE       
#>  2 Clinical   2002-01-17 495616          1 TRUE       
#>  3 Clinical   2008-10-28 534816          1 TRUE       
#>  4 Clinical   2009-04-10 D30712          1 TRUE       
#>  5 Clinical   2016-08-07 092034          1 TRUE       
#>  6 Clinical   2010-08-08 080086          1 TRUE       
#>  7 Clinical   2017-10-04 470001          1 TRUE       
#>  8 Outpatient 2011-04-01 F15984          1 TRUE       
#>  9 Clinical   2015-12-15 D41749          1 TRUE       
#> 10 Clinical   2008-05-06 945BD5          1 TRUE       
#> # … with 190 more rows
if (require("dplyr")) {
  df %>%
    group_by(ward) %>%
    summarise(
      n_patients = n_distinct(patient),
      n_episodes_365 = sum(is_new_episode(date, episode_days = 365)),
      n_episodes_60 = sum(is_new_episode(date, episode_days = 60)),
      n_episodes_30 = sum(is_new_episode(date, episode_days = 30))
    )
}
#> # A tibble: 3 × 5
#>   ward       n_patients n_episodes_365 n_episodes_60 n_episodes_30
#>   <chr>           <int>          <int>         <int>         <int>
#> 1 Clinical          116             14            53            71
#> 2 ICU                52             11            32            40
#> 3 Outpatient         17             10            16            17
if (require("dplyr")) {

  # grouping on patients and microorganisms leads to the same
  # results as first_isolate() when using 'episode-based':
  x <- df %>%
    filter_first_isolate(
      include_unknown = TRUE,
      method = "episode-based"
    )

  y <- df %>%
    group_by(patient, mo) %>%
    filter(is_new_episode(date, 365)) %>%
    ungroup()

  identical(x, y)
}
#> Including isolates from ICU.
#> [1] FALSE
if (require("dplyr")) {

  # but is_new_episode() has a lot more flexibility than first_isolate(),
  # since you can now group on anything that seems relevant:
  df %>%
    group_by(patient, mo, ward) %>%
    mutate(flag_episode = is_new_episode(date, 365)) %>%
    select(group_vars(.), flag_episode)
}
#> # A tibble: 200 × 4
#> # Groups:   patient, mo, ward [190]
#>    patient mo            ward       flag_episode
#>    <chr>   <mo>          <chr>      <lgl>       
#>  1 D53062  B_HMPHL_INFL  Clinical   TRUE        
#>  2 495616  B_STPHY_EPDR  Clinical   TRUE        
#>  3 534816  F_CANDD_ALBC  Clinical   TRUE        
#>  4 D30712  B_STRPT_PNMN  Clinical   TRUE        
#>  5 092034  B_ESCHR_COLI  Clinical   TRUE        
#>  6 080086  B_STPHY_CONS  Clinical   TRUE        
#>  7 470001  B_KLBSL_PNMN  Clinical   TRUE        
#>  8 F15984  B_STPHY_CONS  Outpatient TRUE        
#>  9 D41749  B_ESCHR_COLI  Clinical   TRUE        
#> 10 945BD5  B_ENTRBC_CLOC Clinical   TRUE        
#> # … with 190 more rows
# }