AMR Plots with ggplot2

Use these functions to create bar plots for AMR data analysis. All functions rely on ggplot2 functions.

ggplot_rsi(
  data,
  position = NULL,
  x = "antibiotic",
  fill = "interpretation",
  facet = NULL,
  breaks = seq(0, 1, 0.1),
  limits = NULL,
  translate_ab = "name",
  combine_SI = TRUE,
  combine_IR = FALSE,
  minimum = 30,
  language = get_locale(),
  nrow = NULL,
  colours = c(S = "#3CAEA3", SI = "#3CAEA3", I = "#F6D55C", IR = "#ED553B", R =
    "#ED553B"),
  datalabels = TRUE,
  datalabels.size = 2.5,
  datalabels.colour = "grey15",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  x.title = "Antimicrobial",
  y.title = "Proportion",
  ...
)

geom_rsi(
  position = NULL,
  x = c("antibiotic", "interpretation"),
  fill = "interpretation",
  translate_ab = "name",
  minimum = 30,
  language = get_locale(),
  combine_SI = TRUE,
  combine_IR = FALSE,
  ...
)

facet_rsi(facet = c("interpretation", "antibiotic"), nrow = NULL)

scale_y_percent(breaks = seq(0, 1, 0.1), limits = NULL)

scale_rsi_colours(..., aesthetics = "fill")

theme_rsi()

labels_rsi_count(
  position = NULL,
  x = "antibiotic",
  translate_ab = "name",
  minimum = 30,
  language = get_locale(),
  combine_SI = TRUE,
  combine_IR = FALSE,
  datalabels.size = 3,
  datalabels.colour = "grey15"
)

Arguments

data
a data.frame with column(s) of class rsi (see as.rsi())

position
position adjustment of bars, either "fill", "stack" or "dodge"

x
variable to show on x axis, either "antibiotic" (default) or "interpretation" or a grouping variable

fill
variable to categorise using the plots legend, either "antibiotic" (default) or "interpretation" or a grouping variable

facet
variable to split plots by, either "interpretation" (default) or "antibiotic" or a grouping variable

breaks
a numeric vector of positions

limits
a numeric vector of length two providing limits of the scale, use NA to refer to the existing minimum or maximum

translate_ab
a column name of the antibiotics data set to translate the antibiotic abbreviations to, using ab_property()

combine_SI
a logical to indicate whether all values of S and I must be merged into one, so the output only consists of S+I vs. R (susceptible vs. resistant). This used to be the argument combine_IR, but this now follows the redefinition by EUCAST about the interpretation of I (increased exposure) in 2019, see section 'Interpretation of S, I and R' below. Default is TRUE.

combine_IR
a logical to indicate whether all values of I and R must be merged into one, so the output only consists of S vs. I+R (susceptible vs. non-susceptible). This is outdated, see argument combine_SI.

minimum
the minimum allowed number of available (tested) isolates. Any isolate count lower than minimum will return NA with a warning. The default number of 30 isolates is advised by the Clinical and Laboratory Standards Institute (CLSI) as best practice, see Source.

language
language of the returned text, defaults to system language (see get_locale()) and can also be set with getOption("AMR_locale"). Use language = NULL or language = "" to prevent translation.

nrow
(when using facet) number of rows

colours
a named vactor with colour to be used for filling. The default colours are colour-blind friendly.

datalabels
show datalabels using labels_rsi_count()

datalabels.size
size of the datalabels

datalabels.colour
colour of the datalabels

title
text to show as title of the plot

subtitle
text to show as subtitle of the plot

caption
text to show as caption of the plot

x.title
text to show as x axis description

y.title
text to show as y axis description

...
other arguments passed on to geom_rsi() or, in case of scale_rsi_colours(), named values to set colours. The default colours are colour-blind friendly, while maintaining the convention that e.g. 'susceptible' should be green and 'resistant' should be red. See Examples.

aesthetics
aesthetics to apply the colours to, defaults to "fill" but can also be (a combination of) "alpha", "colour", "fill", "linetype", "shape" or "size"

Details

At default, the names of antibiotics will be shown on the plots using ab_name(). This can be set with the translate_ab argument. See count_df().

The Functions

geom_rsi() will take any variable from the data that has an rsi class (created with as.rsi()) using rsi_df() and will plot bars with the percentage R, I and S. The default behaviour is to have the bars stacked and to have the different antibiotics on the x axis.

facet_rsi() creates 2d plots (at default based on S/I/R) using ggplot2::facet_wrap().

scale_y_percent() transforms the y axis to a 0 to 100% range using ggplot2::scale_y_continuous().

scale_rsi_colours() sets colours to the bars (green for S, yellow for I, and red for R). with multilingual support. The default colours are colour-blind friendly, while maintaining the convention that e.g. 'susceptible' should be green and 'resistant' should be red.

theme_rsi() is a [ggplot2 theme][ggplot2::theme() with minimal distraction.

labels_rsi_count() print datalabels on the bars with percentage and amount of isolates using ggplot2::geom_text().

ggplot_rsi() is a wrapper around all above functions that uses data as first input. This makes it possible to use this function after a pipe (%>%). See Examples.

Stable Lifecycle

The lifecycle of this function is stable. In a stable function, major changes are unlikely. This means that the unlying code will generally evolve by adding new arguments; removing arguments or changing the meaning of existing arguments will be avoided.

If the unlying code needs breaking changes, they will occur gradually. For example, a argument will be deprecated and first continue to work, but will emit an message informing you of the change. Next, typically after at least one newly released version on CRAN, the message will be transformed to an error.

Read more on Our Website!

On our website https://msberends.github.io/AMR/ you can find a comprehensive tutorial about how to conduct AMR data analysis, the complete documentation of all functions and an example analysis using WHONET data.

Examples

# \donttest{
if (require("ggplot2") & require("dplyr")) {
 
  # get antimicrobial results for drugs against a UTI:
  ggplot(example_isolates %>% select(AMX, NIT, FOS, TMP, CIP)) +
    geom_rsi()
 
  # prettify the plot using some additional functions:
  df <- example_isolates %>% select(AMX, NIT, FOS, TMP, CIP)
  ggplot(df) +
    geom_rsi() +
    scale_y_percent() +
    scale_rsi_colours() +
    labels_rsi_count() +
    theme_rsi()
 
  # or better yet, simplify this using the wrapper function - a single command:
  example_isolates %>%
    select(AMX, NIT, FOS, TMP, CIP) %>%
    ggplot_rsi()
 
  # get only proportions and no counts:
  example_isolates %>%
    select(AMX, NIT, FOS, TMP, CIP) %>%
    ggplot_rsi(datalabels = FALSE)
 
  # add other ggplot2 arguments as you like:
  example_isolates %>%
    select(AMX, NIT, FOS, TMP, CIP) %>%
    ggplot_rsi(width = 0.5,
               colour = "black",
               size = 1,
               linetype = 2,
               alpha = 0.25)

  # you can alter the colours with colour names:
  example_isolates %>%
    select(AMX) %>%
    ggplot_rsi(colours = c(SI = "yellow"))

  # but you can also use the built-in colour-blind friendly colours for
  # your plots, where "S" is green, "I" is yellow and "R" is red:
  data.frame(x = c("Value1", "Value2", "Value3"),
             y = c(1, 2, 3),
             z = c("Value4", "Value5", "Value6")) %>%
    ggplot() +
    geom_col(aes(x = x, y = y, fill = z)) +
    scale_rsi_colours(Value4 = "S", Value5 = "I", Value6 = "R")
  
  # resistance of ciprofloxacine per age group
  example_isolates %>%
    mutate(first_isolate = first_isolate()) %>%
    filter(first_isolate == TRUE,
           mo == as.mo("E. coli")) %>%
    # age_groups() is also a function in this AMR package:
    group_by(age_group = age_groups(age)) %>%
    select(age_group,
           CIP) %>%
    ggplot_rsi(x = "age_group")
  
  # a shorter version which also adjusts data label colours:
  example_isolates %>%
    select(AMX, NIT, FOS, TMP, CIP) %>%
    ggplot_rsi(colours = FALSE)

  # it also supports groups (don't forget to use the group var on `x` or `facet`):
  example_isolates %>%
    select(hospital_id, AMX, NIT, FOS, TMP, CIP) %>%
    group_by(hospital_id) %>%
    ggplot_rsi(x = "hospital_id",
               facet = "antibiotic",
               nrow = 1,
               title = "AMR of Anti-UTI Drugs Per Hospital",
               x.title = "Hospital",
               datalabels = FALSE)
}
# }