Data Set with Bacterial Intrinsic Resistance

Data set containing defined intrinsic resistance by EUCAST of all bug-drug combinations.

Usage,

intrinsic_resistant

Format

A data.frame with 134,956 observations and 2 variables:

microorganism
Official taxonomic name of the microorganism, according to the LPSN

antibiotic
Official name of the antibiotic drug, according to the WHOCC

Details

The repository of this AMR package contains a file comprising this exact data set: https://github.com/msberends/AMR/blob/main/data-raw/intrinsic_resistant.txt. This file allows for machine reading EUCAST guidelines about intrinsic resistance, which is almost impossible with the Excel and PDF files distributed by EUCAST. The file is updated automatically.

This data set is based on 'EUCAST Expert Rules' and 'EUCAST Intrinsic Resistance and Unusual Phenotypes' v3.3 (2021).

Reference Data Publicly Available

All reference data sets (about microorganisms, antibiotics, R/SI interpretation, EUCAST rules, etc.) in this AMR package are publicly and freely available. We continually export our data sets to formats for use in R, SPSS, SAS, Stata and Excel. We also supply flat files that are machine-readable and suitable for input in any software program, such as laboratory information systems. Please find all download links on our website, which is automatically updated with every code change.

Read more on Our Website!

On our website https://msberends.github.io/AMR/ you can find a comprehensive tutorial about how to conduct AMR data analysis, the complete documentation of all functions and an example analysis using WHONET data.

Examples

subset(intrinsic_resistant,
       antibiotic == "Vancomycin" & microorganism %like% "Enterococcus")$microorganism
#> [1] "Enterococcus casseliflavus" "Enterococcus gallinarum"

# \donttest{
if (require("dplyr")) {
  intrinsic_resistant %>%
    filter(antibiotic == "Vancomycin" & microorganism %like% "Enterococcus") %>% 
    pull(microorganism)
  #> [1] "Enterococcus casseliflavus" "Enterococcus gallinarum"
}
# }