Pattern Matching

Convenient wrapper around base::grep() to match a pattern: a %like% b. It always returns a logical vector and is always case-insensitive (use a %like_case% b for case-sensitive matching). Also, pattern (b) can be as long as x (a) to compare items of each index in both vectors, or they both can have the same length to iterate over all cases.

like(x, pattern, ignore.case = TRUE)

x %like% pattern

x %like_case% pattern

Arguments

x
a character vector where matches are sought, or an object which can be coerced by as.character() to a character vector.

pattern
a character string containing a regular expression (or character string for fixed = TRUE) to be matched in the given character vector. Coerced by as.character() to a character string if possible. If a character vector of length 2 or more is supplied, the first element is used with a warning.

ignore.case
if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.

Source

Idea from the like function from the data.table package, but made it case insensitive at default and let it support multiple patterns. Also, if the regex fails the first time, it tries again with perl = TRUE.

Value

A logical vector

Details

Using RStudio? This function can also be inserted from the Addins menu and can have its own Keyboard Shortcut like Ctrl+Shift+L or Cmd+Shift+L (see Tools > Modify Keyboard Shortcuts...).

Stable lifecycle

The lifecycle of this function is stable. In a stable function, we are largely happy with the unlying code, and major changes are unlikely. This means that the unlying code will generally evolve by adding new arguments; we will avoid removing arguments or changing the meaning of existing arguments.

If the unlying code needs breaking changes, they will occur gradually. To begin with, the function or argument will be deprecated; it will continue to work but will emit an message informing you of the change. Next, typically after at least one newly released version on CRAN, the message will be transformed to an error.

Read more on our website!

On our website https://msberends.gitlab.io/AMR you can find a comprehensive tutorial about how to conduct AMR analysis, the complete documentation of all functions (which reads a lot easier than here in R) and an example analysis using WHONET data.

See also

Examples

# simple test
a <- "This is a test"
b <- "TEST"
a %like% b
#> TRUE
b %like% a
#> FALSE

# also supports multiple patterns, length must be equal to x
a <- c("Test case", "Something different", "Yet another thing")
b <- c(     "case",           "diff",      "yet")
a %like% b
#> TRUE TRUE TRUE

# get frequencies of bacteria whose name start with 'Ent' or 'ent'
library(dplyr)
example_isolates %>%
  filter(mo_name(mo) %like% '^ent') %>%
  freq(mo_genus(mo))