Pattern Matching

Convenient wrapper around grep() to match a pattern: x %like% pattern. It always returns a logical vector and is always case-insensitive (use x %like_case% pattern for case-sensitive matching). Also, pattern can be as long as x to compare items of each index in both vectors, or they both can have the same length to iterate over all cases.

like(x, pattern, ignore.case = TRUE)

x %like% pattern

x %like_case% pattern

Arguments

x
a character vector where matches are sought, or an object which can be coerced by as.character() to a character vector.

pattern
a character string containing a regular expression (or character string for fixed = TRUE) to be matched in the given character vector. Coerced by as.character() to a character string if possible. If a character vector of length 2 or more is supplied, the first element is used with a warning.

ignore.case
if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.

Source

Idea from the like function from the data.table package

Value

A logical vector

Details

The %like% function:

Is case-insensitive (use %like_case% for case-sensitive matching)

Supports multiple patterns

Checks if pattern is a regular expression and sets fixed = TRUE if not, to greatly improve speed

Tries again with perl = TRUE if regex fails

Using RStudio? This function can also be inserted from the Addins menu and can have its own Keyboard Shortcut like Ctrl+Shift+L or Cmd+Shift+L (see Tools > Modify Keyboard Shortcuts...).

Stable lifecycle

The lifecycle of this function is stable. In a stable function, major changes are unlikely. This means that the unlying code will generally evolve by adding new arguments; removing arguments or changing the meaning of existing arguments will be avoided.

If the unlying code needs breaking changes, they will occur gradually. For example, a parameter will be deprecated and first continue to work, but will emit an message informing you of the change. Next, typically after at least one newly released version on CRAN, the message will be transformed to an error.

Read more on our website!

On our website https://msberends.github.io/AMR/ you can find a comprehensive tutorial about how to conduct AMR analysis, the complete documentation of all functions and an example analysis using WHONET data. As we would like to better understand the backgrounds and needs of our users, please participate in our survey!

See also

Examples

# simple test
a <- "This is a test"
b <- "TEST"
a %like% b
#> TRUE
b %like% a
#> FALSE

# also supports multiple patterns, length must be equal to x
a <- c("Test case", "Something different", "Yet another thing")
b <- c(     "case",           "diff",      "yet")
a %like% b
#> TRUE TRUE TRUE

# get isolates whose name start with 'Ent' or 'ent'
# \donttest{
if (require("dplyr")) {
  example_isolates %>%
    filter(mo_name(mo) %like% "^ent")
}
# }