Determine multidrug-resistant organisms (MDRO)

Determine which isolates are multidrug-resistant organisms (MDRO) according to country-specific guidelines.

mdro(tbl, country = NULL, col_mo = NULL, info = TRUE,
  amcl = "amcl", amik = "amik", amox = "amox", ampi = "ampi",
  azit = "azit", aztr = "aztr", cefa = "cefa", cfra = "cfra",
  cfep = "cfep", cfot = "cfot", cfox = "cfox", cfta = "cfta",
  cftr = "cftr", cfur = "cfur", chlo = "chlo", cipr = "cipr",
  clar = "clar", clin = "clin", clox = "clox", coli = "coli",
  czol = "czol", dapt = "dapt", doxy = "doxy", erta = "erta",
  eryt = "eryt", fosf = "fosf", fusi = "fusi", gent = "gent",
  imip = "imip", kana = "kana", levo = "levo", linc = "linc",
  line = "line", mero = "mero", metr = "metr", mino = "mino",
  moxi = "moxi", nali = "nali", neom = "neom", neti = "neti",
  nitr = "nitr", novo = "novo", norf = "norf", oflo = "oflo",
  peni = "peni", pipe = "pipe", pita = "pita", poly = "poly",
  qida = "qida", rifa = "rifa", roxi = "roxi", siso = "siso",
  teic = "teic", tetr = "tetr", tica = "tica", tige = "tige",
  tobr = "tobr", trim = "trim", trsu = "trsu", vanc = "vanc")

brmo(..., country = "nl")

mrgn(tbl, country = "de", ...)

eucast_exceptional_phenotypes(tbl, country = "EUCAST", ...)

Arguments

tbl
table with antibiotic columns, like e.g. amox and amcl

country
country code to determine guidelines. EUCAST rules will be used when left empty, see Details. Should be or a code from the list of ISO 3166-1 alpha-2 country codes. Case-insensitive. Currently supported are de (Germany) and nl (the Netherlands).

col_mo
column name of the unique IDs of the microorganisms (see mo), defaults to the first column of class mo. Values will be coerced using as.mo.

info
print progress

amcl
column name of an antibiotic, see Antibiotics

amik
column name of an antibiotic, see Antibiotics

amox
column name of an antibiotic, see Antibiotics

ampi
column name of an antibiotic, see Antibiotics

azit
column name of an antibiotic, see Antibiotics

aztr
column name of an antibiotic, see Antibiotics

cefa
column name of an antibiotic, see Antibiotics

cfra
column name of an antibiotic, see Antibiotics

cfep
column name of an antibiotic, see Antibiotics

cfot
column name of an antibiotic, see Antibiotics

cfox
column name of an antibiotic, see Antibiotics

cfta
column name of an antibiotic, see Antibiotics

cftr
column name of an antibiotic, see Antibiotics

cfur
column name of an antibiotic, see Antibiotics

chlo
column name of an antibiotic, see Antibiotics

cipr
column name of an antibiotic, see Antibiotics

clar
column name of an antibiotic, see Antibiotics

clin
column name of an antibiotic, see Antibiotics

clox
column name of an antibiotic, see Antibiotics

coli
column name of an antibiotic, see Antibiotics

czol
column name of an antibiotic, see Antibiotics

dapt
column name of an antibiotic, see Antibiotics

doxy
column name of an antibiotic, see Antibiotics

erta
column name of an antibiotic, see Antibiotics

eryt
column name of an antibiotic, see Antibiotics

fosf
column name of an antibiotic, see Antibiotics

fusi
column name of an antibiotic, see Antibiotics

gent
column name of an antibiotic, see Antibiotics

imip
column name of an antibiotic, see Antibiotics

kana
column name of an antibiotic, see Antibiotics

levo
column name of an antibiotic, see Antibiotics

linc
column name of an antibiotic, see Antibiotics

line
column name of an antibiotic, see Antibiotics

mero
column name of an antibiotic, see Antibiotics

metr
column name of an antibiotic, see Antibiotics

mino
column name of an antibiotic, see Antibiotics

moxi
column name of an antibiotic, see Antibiotics

nali
column name of an antibiotic, see Antibiotics

neom
column name of an antibiotic, see Antibiotics

neti
column name of an antibiotic, see Antibiotics

nitr
column name of an antibiotic, see Antibiotics

novo
column name of an antibiotic, see Antibiotics

norf
column name of an antibiotic, see Antibiotics

oflo
column name of an antibiotic, see Antibiotics

peni
column name of an antibiotic, see Antibiotics

pipe
column name of an antibiotic, see Antibiotics

pita
column name of an antibiotic, see Antibiotics

poly
column name of an antibiotic, see Antibiotics

qida
column name of an antibiotic, see Antibiotics

rifa
column name of an antibiotic, see Antibiotics

roxi
column name of an antibiotic, see Antibiotics

siso
column name of an antibiotic, see Antibiotics

teic
column name of an antibiotic, see Antibiotics

tetr
column name of an antibiotic, see Antibiotics

tica
column name of an antibiotic, see Antibiotics

tige
column name of an antibiotic, see Antibiotics

tobr
column name of an antibiotic, see Antibiotics

trim
column name of an antibiotic, see Antibiotics

trsu
column name of an antibiotic, see Antibiotics

vanc
column name of an antibiotic, see Antibiotics

...
parameters that are passed on to methods

Value

Ordered factor with levels Negative < Positive, unconfirmed < Positive.

Details

When country will be left blank, guidelines will be taken from EUCAST Expert Rules Version 3.1 "Intrinsic Resistance and Exceptional Phenotypes Tables" (http://www.eucast.org/fileadmin/src/media/PDFs/EUCAST_files/Expert_Rules/Expert_rules_intrinsic_exceptional_V3.1.pdf).

Antibiotics

To define antibiotics column names, input a text (case-insensitive) or use NULL to skip a column (e.g. tica = NULL). Non-existing columns will anyway be skipped with a warning.

Abbrevations of the column containing antibiotics in the form: abbreviation: generic name (ATC code)

amcl: amoxicillin+clavulanic acid (J01CR02), amik: amikacin (J01GB06), amox: amoxicillin (J01CA04), ampi: ampicillin (J01CA01), azit: azithromycin (J01FA10), azlo: azlocillin (J01CA09), aztr: aztreonam (J01DF01), cefa: cefaloridine (J01DB02), cfep: cefepime (J01DE01), cfot: cefotaxime (J01DD01), cfox: cefoxitin (J01DC01), cfra: cefradine (J01DB09), cfta: ceftazidime (J01DD02), cftr: ceftriaxone (J01DD04), cfur: cefuroxime (J01DC02), chlo: chloramphenicol (J01BA01), cipr: ciprofloxacin (J01MA02), clar: clarithromycin (J01FA09), clin: clindamycin (J01FF01), clox: flucloxacillin (J01CF05), coli: colistin (J01XB01), czol: cefazolin (J01DB04), dapt: daptomycin (J01XX09), doxy: doxycycline (J01AA02), erta: ertapenem (J01DH03), eryt: erythromycin (J01FA01), fosf: fosfomycin (J01XX01), fusi: fusidic acid (J01XC01), gent: gentamicin (J01GB03), imip: imipenem (J01DH51), kana: kanamycin (J01GB04), levo: levofloxacin (J01MA12), linc: lincomycin (J01FF02), line: linezolid (J01XX08), mero: meropenem (J01DH02), mezl: mezlocillin (J01CA10), mino: minocycline (J01AA08), moxi: moxifloxacin (J01MA14), nali: nalidixic acid (J01MB02), neom: neomycin (J01GB05), neti: netilmicin (J01GB07), nitr: nitrofurantoin (J01XE01), norf: norfloxacin (J01MA06), novo: novobiocin (an ATCvet code: QJ01XX95), oflo: ofloxacin (J01MA01), peni: penicillin (J01RA01), pipe: piperacillin (J01CA12), pita: piperacillin+tazobactam (J01CR05), poly: polymyxin B (J01XB02), pris: pristinamycin (J01FG01), qida: quinupristin/dalfopristin (J01FG02), rifa: rifampicin (J04AB02), roxi: roxithromycin (J01FA06), siso: sisomicin (J01GB08), teic: teicoplanin (J01XA02), tetr: tetracycline (J01AA07), tica: ticarcillin (J01CA13), tige: tigecycline (J01AA12), tobr: tobramycin (J01GB01), trim: trimethoprim (J01EA01), trsu: sulfamethoxazole and trimethoprim (J01EE01), vanc: vancomycin (J01XA01).

Examples

# NOT RUN {
library(dplyr)

septic_patients %>%
  mutate(EUCAST = mdro(.),
         BRMO = brmo(.))
# }