Determine Multidrug-Resistant Organisms (MDRO)

Determine which isolates are multidrug-resistant organisms (MDRO) according to international, national and custom guidelines.

Usage

mdro(
  x = NULL,
  guideline = "CMI2012",
  col_mo = NULL,
  info = interactive(),
  pct_required_classes = 0.5,
  combine_SI = TRUE,
  verbose = FALSE,
  only_rsi_columns = FALSE,
  ...
)

custom_mdro_guideline(..., as_factor = TRUE)

brmo(x = NULL, only_rsi_columns = FALSE, ...)

mrgn(x = NULL, only_rsi_columns = FALSE, ...)

mdr_tb(x = NULL, only_rsi_columns = FALSE, ...)

mdr_cmi2012(x = NULL, only_rsi_columns = FALSE, ...)

eucast_exceptional_phenotypes(x = NULL, only_rsi_columns = FALSE, ...)

Source

See the supported guidelines above for the list of publications used for this function.

Arguments

x

a data.frame with antibiotics columns, like AMX or amox. Can be left blank for automatic determination.

guideline

a specific guideline to follow, see sections Supported international / national guidelines and Using Custom Guidelines below. When left empty, the publication by Magiorakos et al. (see below) will be followed.

col_mo

column name of the IDs of the microorganisms (see as.mo()), defaults to the first column of class mo. Values will be coerced using as.mo().

info

a logical to indicate whether progress should be printed to the console, defaults to only print while in interactive sessions

pct_required_classes

minimal required percentage of antimicrobial classes that must be available per isolate, rounded down. For example, with the default guideline, 17 antimicrobial classes must be available for S. aureus. Setting this pct_required_classes argument to 0.5 (default) means that for every S. aureus isolate at least 8 different classes must be available. Any lower number of available classes will return NA for that isolate.

combine_SI

a logical to indicate whether all values of S and I must be merged into one, so resistance is only considered when isolates are R, not I. As this is the default behaviour of the mdro() function, it follows the redefinition by EUCAST about the interpretation of I (increased exposure) in 2019, see section 'Interpretation of S, I and R' below. When using combine_SI = FALSE, resistance is considered when isolates are R or I.

verbose

a logical to turn Verbose mode on and off (default is off). In Verbose mode, the function does not return the MDRO results, but instead returns a data set in logbook form with extensive info about which isolates would be MDRO-positive, or why they are not.

only_rsi_columns

a logical to indicate whether only antibiotic columns must be detected that were transformed to class rsi (see as.rsi()) on beforehand (defaults to FALSE)

...

in case of custom_mdro_guideline(): a set of rules, see section Using Custom Guidelines below. Otherwise: column name of an antibiotic, see section Antibiotics below.

as_factor

a logical to indicate whether the returned value should be an ordered factor (TRUE, default), or otherwise a character vector

Value

CMI 2012 paper - function mdr_cmi2012() or mdro():
Ordered factor with levels Negative < Multi-drug-resistant (MDR) < Extensively drug-resistant (XDR) < Pandrug-resistant (PDR)

TB guideline - function mdr_tb() or mdro(..., guideline = "TB"):
Ordered factor with levels Negative < Mono-resistant < Poly-resistant < Multi-drug-resistant < Extensively drug-resistant

German guideline - function mrgn() or mdro(..., guideline = "MRGN"):
Ordered factor with levels Negative < 3MRGN < 4MRGN

Everything else, except for custom guidelines:
Ordered factor with levels Negative < Positive, unconfirmed < Positive. The value "Positive, unconfirmed" means that, according to the guideline, it is not entirely sure if the isolate is multi-drug resistant and this should be confirmed with additional (e.g. molecular) tests

Details

These functions are context-aware. This means that the x argument can be left blank if used inside a data.frame call, see Examples.

For the pct_required_classes argument, values above 1 will be divided by 100. This is to support both fractions (0.75 or 3/4) and percentages (75).

Note: Every test that involves the Enterobacteriaceae family, will internally be performed using its newly named order Enterobacterales, since the Enterobacteriaceae family has been taxonomically reclassified by Adeolu et al. in 2016. Before that, Enterobacteriaceae was the only family under the Enterobacteriales (with an i) order. All species under the old Enterobacteriaceae family are still under the new Enterobacterales (without an i) order, but divided into multiple families. The way tests are performed now by this mdro() function makes sure that results from before 2016 and after 2016 are identical.

Supported International / National Guidelines

Currently supported guidelines are (case-insensitive):

guideline = "CMI2012" (default)

Magiorakos AP, Srinivasan A et al. "Multidrug-resistant, extensively drug-resistant and pandrug-resistant bacteria: an international expert proposal for interim standard definitions for acquired resistance." Clinical Microbiology and Infection (2012) (link)

guideline = "EUCAST3.3" (or simply guideline = "EUCAST")

The European international guideline - EUCAST Expert Rules Version 3.3 "Intrinsic Resistance and Unusual Phenotypes" (link)

guideline = "EUCAST3.2"

The European international guideline - EUCAST Expert Rules Version 3.2 "Intrinsic Resistance and Unusual Phenotypes" (link)

guideline = "EUCAST3.1"

The European international guideline - EUCAST Expert Rules Version 3.1 "Intrinsic Resistance and Exceptional Phenotypes Tables" (link)

guideline = "TB"

The international guideline for multi-drug resistant tuberculosis - World Health Organization "Companion handbook to the WHO guidelines for the programmatic management of drug-resistant tuberculosis" (link)

guideline = "MRGN"

The German national guideline - Mueller et al. (2015) Antimicrobial Resistance and Infection Control 4:7; doi:10.1186/s13756-015-0047-6

guideline = "BRMO"

The Dutch national guideline - Rijksinstituut voor Volksgezondheid en Milieu "WIP-richtlijn BRMO (Bijzonder Resistente Micro-Organismen) (ZKH)" (link)

Please suggest your own (country-specific) guidelines by letting us know: https://github.com/msberends/AMR/issues/new.

Using Custom Guidelines

Custom guidelines can be set with the custom_mdro_guideline() function. This is of great importance if you have custom rules to determine MDROs in your hospital, e.g., rules that are dependent on ward, state of contact isolation or other variables in your data.

If you are familiar with the case_when() function of the dplyr package, you will recognise the input method to set your own rules. Rules must be set using what R considers to be the 'formula notation'. The rule is written before the tilde (~) and the consequence of the rule is written after the tilde:

custom <- custom_mdro_guideline(CIP == "R" & age > 60 ~ "Elderly Type A",
                                ERY == "R" & age > 60 ~ "Elderly Type B")

If a row/an isolate matches the first rule, the value after the first ~ (in this case 'Elderly Type A') will be set as MDRO value. Otherwise, the second rule will be tried and so on. The number of rules is unlimited.

You can print the rules set in the console for an overview. Colours will help reading it if your console supports colours.

custom
#> A set of custom MDRO rules:
#>   1. CIP is "R" and age is higher than 60 -> Elderly Type A
#>   2. ERY is "R" and age is higher than 60 -> Elderly Type B
#>   3. Otherwise -> Negative
#>
#> Unmatched rows will return NA.

The outcome of the function can be used for the guideline argument in the mdro() function:

x <- mdro(example_isolates,
          guideline = custom)
table(x)
#>       Negative Elderly Type A Elderly Type B
#>           1070            198            732

Rules can also be combined with other custom rules by using c():

x <- mdro(example_isolates,
          guideline = c(custom,
                        custom_mdro_guideline(ERY == "R" & age > 50 ~ "Elderly Type C")))
table(x)
#>       Negative Elderly Type A Elderly Type B Elderly Type C
#>            961            198            732            109

The rules set (the custom object in this case) could be exported to a shared file location using saveRDS() if you collaborate with multiple users. The custom rules set could then be imported using readRDS().

Antibiotics

To define antibiotics column names, leave as it is to determine it automatically with guess_ab_col() or input a text (case-insensitive), or use NULL to skip a column (e.g. TIC = NULL to skip ticarcillin). Manually defined but non-existing columns will be skipped with a warning.

The following antibiotics are eligible for the functions eucast_rules() and mdro(). These are shown below in the format 'name (antimicrobial ID, ATC code)', sorted alphabetically:

Amikacin (AMK, S01AE08), amoxicillin (AMX, J01MA02), amoxicillin/clavulanic acid (AMC, J01MA23), ampicillin (AMP, J01MA04), ampicillin/sulbactam (SAM, J01MA08), arbekacin (ARB, J01MA19), aspoxicillin (APX, J01MA16), azidocillin (AZD, J01MA15), azithromycin (AZM, J01MA11), azlocillin (AZL, J01MA25), aztreonam (ATM, J01MA12), bacampicillin (BAM, J01MA24), bekanamycin (BEK, J01MA07), benzathine benzylpenicillin (BNB, J01MA14), benzathine phenoxymethylpenicillin (BNP, D10AF05), benzylpenicillin (PEN, J01MA06), besifloxacin (BES, J01MA01), biapenem (BIA, J01MA18), carbenicillin (CRB, J01MA03), carindacillin (CRN, J01MA17), cefacetrile (CAC, J01MA10), cefaclor (CEC, J01MA21), cefadroxil (CFR, J01MA09), cefalexin (LEX, J01MA05), cefaloridine (RID, P01AA05), cefalotin (CEP, J01MA22), cefamandole (MAN, J01MA13), cefapirin (HAP, J01CA01), cefatrizine (CTZ, J01CA04), cefazedone (CZD, J01CA12), cefazolin (CZO, J01CR05), cefcapene (CCP, J01CA13), cefdinir (CDR, J01AA02), cefditoren (DIT, J01FA10), cefepime (FEP, J01FA09), cefetamet (CAT, J01CR02), cefixime (CFM, J01AA08), cefmenoxime (CMX, J01FA06), cefmetazole (CMZ, J01CF04), cefodizime (DIZ, J01CF05), cefonicid (CID, J01CR01), cefoperazone (CFP, J01CA19), cefoperazone/sulbactam (CSL, J01CE04), ceforanide (CND, J01CA09), cefotaxime (CTX, J01DF01), cefotaxime/clavulanic acid (CTC, J01CA06), cefotetan (CTT, J01CE08), cefotiam (CTF, J01CE10), cefoxitin (FOX, J01CE01), cefozopran (ZOP, J01CA03), cefpiramide (CPM, J01CA05), cefpirome (CPO, J01CE07), cefpodoxime (CPD, J01CF02), cefprozil (CPR, J01CF01), cefroxadine (CRD, J01CA07), cefsulodin (CFS, J01CA18), ceftaroline (CPT, J01CA11), ceftazidime (CAZ, J01CA14), ceftazidime/clavulanic acid (CCV, J01CF03), cefteram (CEM, J01CA10), ceftezole (CTL, J01CF06), ceftibuten (CTB, J01CE06), ceftizoxime (CZX, J01CE05), ceftobiprole medocaril (CFM1, J01CE02), ceftolozane/tazobactam (CZT, J01CA02), ceftriaxone (CRO, J01CA08), ceftriaxone/beta-lactamase inhibitor (CEB, J01CE09), cefuroxime (CXM, J01CE03), cephradine (CED, J01CG01), chloramphenicol (CHL, J01CA16), ciprofloxacin (CIP, J01CR04), clarithromycin (CLR, J01CA15), clindamycin (CLI, J01CG02), clometocillin (CLM, J01CA17), cloxacillin (CLO, J01CR03), colistin (COL, J01DB10), cycloserine (CYC, J01DC04), dalbavancin (DAL, J01DB05), daptomycin (DAP, J01DB01), delafloxacin (DFX, J01DB02), dibekacin (DKB, J01DB03), dicloxacillin (DIC, J01DC03), dirithromycin (DIR, J01DB08), doripenem (DOR, J01DB07), doxycycline (DOX, J01DB06), enoxacin (ENX, J01DB04), epicillin (EPC, J01DD17), ertapenem (ETP, J01DD15), erythromycin (ERY, J01DD16), fleroxacin (FLE, J01DE01), flucloxacillin (FLC, J01DD10), flurithromycin (FLR1, J01DD08), fosfomycin (FOS, J01DD05), framycetin (FRM, J01DC09), fusidic acid (FUS, J01DD09), garenoxacin (GRN, J01DC06), gatifloxacin (GAT, J01DD12), gemifloxacin (GEM, J01DD62), gentamicin (GEN, J01DC11), grepafloxacin (GRX, J01DD01), hetacillin (HET, J01DD51), imipenem (IPM, J01DC05), imipenem/relebactam (IMR, J01DC07), isepamicin (ISE, J01DC01), josamycin (JOS, J01DE03), kanamycin (KAN, J01DD11), lascufloxacin (LSC, J01DE02), latamoxef (LTM, J01DD13), levofloxacin (LVX, J01DC10), levonadifloxacin (LND, J01DB11), lincomycin (LIN, J01DD03), linezolid (LNZ, J01DI02), lomefloxacin (LOM, J01DD02), loracarbef (LOR, J01DD52), mecillinam (MEC, J01DD18), meropenem (MEM, J01DB12), meropenem/vaborbactam (MEV, J01DD14), metampicillin (MTM, J01DD07), meticillin (MET, J01DI01), mezlocillin (MEZ, J01DI54), micronomicin (MCR, J01DD04), midecamycin (MID, J01DD63), minocycline (MNO, J01DC02), miocamycin (MCM, J01DB09), moxifloxacin (MFX, J01DD06), nadifloxacin (NAD, J01DC08), nafcillin (NAF, J01DH05), nalidixic acid (NAL, J01DH04), neomycin (NEO, J01DH03), netilmicin (NET, J01DH51), nitrofurantoin (NIT, J01DH56), norfloxacin (NOR, J01DH02), ofloxacin (OFX, J01DH52), oleandomycin (OLE, J01DH55), oritavancin (ORI, J01DH06), oxacillin (OXA, J01XA02), panipenem (PAN, J01XA01), pazufloxacin (PAZ, J01XC01), pefloxacin (PEF, J01FA13), penamecillin (PNM, J01FA01), pheneticillin (PHE, J01FA14), phenoxymethylpenicillin (PHN, J01FA07), piperacillin (PIP, J01FA03), piperacillin/tazobactam (TZP, J01FA11), pivampicillin (PVM, J01FA05), pivmecillinam (PME, J01FA12), plazomicin (PLZ, J01FA16), polymyxin B (PLB, J01FA02), pristinamycin (PRI, J01FA15), procaine benzylpenicillin (PRB, J01FA08), propicillin (PRP, J01FF02), prulifloxacin (PRU, J01FG01), quinupristin/dalfopristin (QDA, J01FG02), ribostamycin (RST, J04AB02), rifampicin (RIF, J01XX09), rokitamycin (ROK, J01XX08), roxithromycin (RXT, J01AA07), rufloxacin (RFL, J01XB01), sisomicin (SIS, J01XB02), sitafloxacin (SIT, J01XE01), solithromycin (SOL, J01AA12), sparfloxacin (SPX, J01EA01), spiramycin (SPI, J01XX01), streptoduocin (STR, J01BA01), streptomycin (STR1, J01GB06), sulbactam (SUL, J01GB12), sulbenicillin (SBC, J01GB13), sulfadiazine (SDI, J01GB09), sulfadiazine/trimethoprim (SLT1, D09AA01), sulfadimethoxine (SUD, J01GB03), sulfadimidine (SDM, J01GB11), sulfadimidine/trimethoprim (SLT2, J01GB04), sulfafurazole (SLF, S01AA22), sulfaisodimidine (SLF1, J01GB05), sulfalene (SLF2, J01GB07), sulfamazone (SZO, J01GB14), sulfamerazine (SLF3, J01GB10), sulfamerazine/trimethoprim (SLT3, J01GB08), sulfamethizole (SLF4, J01GA02), sulfamethoxazole (SMX, J01GA01), sulfamethoxypyridazine (SLF5, J01GB01), sulfametomidine (SLF6, J01EE01), sulfametoxydiazine (SLF7, J01MB02), sulfametrole/trimethoprim (SLT4, J01FF01), sulfamoxole (SLF8, J01XA04), sulfamoxole/trimethoprim (SLT5, J01XA05), sulfanilamide (SLF9, J01XA03), sulfaperin (SLF10, J04AB01), sulfaphenazole (SLF11, J01XX11), sulfapyridine (SLF12, J01EC02), sulfathiazole (SUT, J01ED01), sulfathiourea (SLF13, J01EB03), sultamicillin (SLT6, J01EB05), talampicillin (TAL, J01EB01), tazobactam (TAZ, J01ED02), tebipenem (TBP, J01ED09), tedizolid (TZD, J01ED07), teicoplanin (TEC, J01EB02), telavancin (TLV, J01EC01), telithromycin (TLT, J01ED05), temafloxacin (TMX, J01ED03), temocillin (TEM, J01ED04), tetracycline (TCY, J01EC03), ticarcillin (TIC, J01EB06), ticarcillin/clavulanic acid (TCC, J01ED06), tigecycline (TGC, J01ED08), tilbroquinol (TBQ, J01EB04), tobramycin (TOB, J01EB07), tosufloxacin (TFX, J01EB08), trimethoprim (TMP, J01EE02), trimethoprim/sulfamethoxazole (SXT, J01EE05), troleandomycin (TRL, J01EE07), trovafloxacin (TVA, J01EE03), vancomycin (VAN, J01EE04)

Interpretation of R and S/I

In 2019, the European Committee on Antimicrobial Susceptibility Testing (EUCAST) has decided to change the definitions of susceptibility testing categories R and S/I as shown below (https://www.eucast.org/newsiandr/).

R = Resistant
A microorganism is categorised as Resistant when there is a high likelihood of therapeutic failure even when there is increased exposure. Exposure is a function of how the mode of administration, dose, dosing interval, infusion time, as well as distribution and excretion of the antimicrobial agent will influence the infecting organism at the site of infection.

S = Susceptible
A microorganism is categorised as Susceptible, standard dosing regimen, when there is a high likelihood of therapeutic success using a standard dosing regimen of the agent.

I = Susceptible, Increased exposure
A microorganism is categorised as Susceptible, Increased exposure when there is a high likelihood of therapeutic success because exposure to the agent is increased by adjusting the dosing regimen or by its concentration at the site of infection.

This AMR package honours this insight. Use susceptibility() (equal to proportion_SI()) to determine antimicrobial susceptibility and count_susceptible() (equal to count_SI()) to count susceptible isolates.

Examples

out <- mdro(example_isolates, guideline = "EUCAST")
#>  (16 isolates had no test results)
str(out)
#>  Ord.factor w/ 3 levels "Negative"<"Positive, unconfirmed"<..: 1 1 1 1 1 1 1 1 1 1 ...
table(out)
#> out
#>              Negative Positive, unconfirmed              Positive 
#>                  1978                     0                     6 

out <- mdro(example_isolates,
  guideline = custom_mdro_guideline(
    AMX == "R" ~ "Custom MDRO 1",
    VAN == "R" ~ "Custom MDRO 2"
  )
)
table(out)
#> out
#>      Negative Custom MDRO 1 Custom MDRO 2 
#>           870           804           326 

# \donttest{
if (require("dplyr")) {
  example_isolates %>%
    mdro() %>%
    table()

  # no need to define `x` when used inside dplyr verbs:
  example_isolates %>%
    mutate(MDRO = mdro()) %>%
    pull(MDRO) %>%
    table()
}
#>  (16 isolates had no test results)
#> Warning: in mdro(): NA introduced for isolates where the available percentage of
#> antimicrobial classes was below 50% (set with pct_required_classes)
#>  (16 isolates had no test results)
#> Warning: in mdro(): NA introduced for isolates where the available percentage of
#> antimicrobial classes was below 50% (set with pct_required_classes)
#> .
#>                         Negative       Multi-drug-resistant (MDR) 
#>                             1601                              128 
#> Extensively drug-resistant (XDR)          Pandrug-resistant (PDR) 
#>                                0                                0 
# }