Mean AMR Distance

This function calculates a normalised mean for antimicrobial resistance between multiple observations.

Usage

mean_amr_distance(x, ...)

# S3 method for default
mean_amr_distance(x, ...)

# S3 method for mic
mean_amr_distance(x, ...)

# S3 method for disk
mean_amr_distance(x, ...)

# S3 method for rsi
mean_amr_distance(x, ..., combine_SI = TRUE)

# S3 method for data.frame
mean_amr_distance(x, ..., combine_SI = TRUE)

amr_distance_from_row(amr_distance, row)

Arguments

x

a vector of class rsi, rsi or rsi, or a data.frame containing columns of any of these classes

...

variables to select (supports tidy selection such as column1:column4 and where(is.mic)), and can thus also be antibiotic selectors

combine_SI

a logical to indicate whether all values of S and I must be merged into one, so the input only consists of S+I vs. R (susceptible vs. resistant), defaults to TRUE

amr_distance

the outcome of mean_amr_distance()

row

an index, such as a row number

Details

The mean AMR distance is a normalised numeric value to compare AMR test results and can help to identify similar isolates, without comparing antibiograms by hand. For common numeric data this distance is equal to Z scores (the number of standard deviations from the mean).

MIC values (see as.mic()) are transformed with log2() first; their distance is calculated as (log2(x) - mean(log2(x))) / sd(log2(x)).

R/SI values (see as.rsi()) are transformed using "S" = 1, "I" = 2, and "R" = 3. If combine_SI is TRUE (default), the "I" will be considered to be 1.

For data sets, the mean AMR distance will be calculated per variable, after which the mean of all columns will returned per row (using rowMeans()), see Examples.

Use amr_distance_from_row() to subtract distances from the distance of one row, see Examples.

Interpretation

Isolates with distances less than 0.01 difference from each other should be considered similar. Differences lower than 0.025 should be considered suspicious.

Examples

x <- random_mic(10)
x
#> Class 'mic'
#>  [1] 64     1      8      0.025  0.002  0.005  0.0625 16     0.01   32    
mean_amr_distance(x)
#>  [1]  1.3091844  0.2558741  0.7825293 -0.6783995 -1.3180848 -1.0860180
#>  [7] -0.4463328  0.9580810 -0.9104663  1.1336327

y <- data.frame(
  id = LETTERS[1:10],
  amox = random_mic(10, ab = "amox", mo = "Escherichia coli"),
  cipr = random_mic(10, ab = "cipr", mo = "Escherichia coli"),
  gent = random_mic(10, ab = "gent", mo = "Escherichia coli"),
  tobr = random_mic(10, ab = "tobr", mo = "Escherichia coli")
)
y
#>    id amox  cipr gent tobr
#> 1   A    8 0.025    1    1
#> 2   B  <=4 0.125  0.5    4
#> 3   C  <=4   >=1    4    4
#> 4   D  <=4 0.125    1  >=8
#> 5   E >=16   >=1    2  0.5
#> 6   F >=16 0.025    1    2
#> 7   G  <=4  0.25    4  0.5
#> 8   H >=16 0.125  0.5  >=8
#> 9   I    8   0.5  0.5  >=8
#> 10  J  <=4  0.25    1    1
mean_amr_distance(y)
#> ℹ Calculating mean AMR distance based on columns "amox", "cipr", "gent",
#>   "id" and "tobr"
#> Warning: NAs introduced by coercion
#>  [1] -0.55426393 -0.42949279  0.62941387 -0.05490015  0.49015830 -0.12782528
#>  [7] -0.09926391  0.26899945  0.26246847 -0.38529402
y$amr_distance <- mean_amr_distance(y, where(is.mic))
#> ℹ Calculating mean AMR distance based on columns "amox", "cipr", "gent" and
#>   "tobr"
y[order(y$amr_distance), ]
#>    id amox  cipr gent tobr amr_distance
#> 1   A    8 0.025    1    1  -0.55426393
#> 2   B  <=4 0.125  0.5    4  -0.42949279
#> 10  J  <=4  0.25    1    1  -0.38529402
#> 6   F >=16 0.025    1    2  -0.12782528
#> 7   G  <=4  0.25    4  0.5  -0.09926391
#> 4   D  <=4 0.125    1  >=8  -0.05490015
#> 9   I    8   0.5  0.5  >=8   0.26246847
#> 8   H >=16 0.125  0.5  >=8   0.26899945
#> 5   E >=16   >=1    2  0.5   0.49015830
#> 3   C  <=4   >=1    4    4   0.62941387

if (require("dplyr")) {
  y %>%
    mutate(
      amr_distance = mean_amr_distance(., where(is.mic)),
      check_id_C = amr_distance_from_row(amr_distance, id == "C")
    ) %>%
    arrange(check_id_C)
}
#> ℹ Calculating mean AMR distance based on columns "amox", "cipr", "gent" and
#>   "tobr"
#>    id amox  cipr gent tobr amr_distance check_id_C
#> 1   C  <=4   >=1    4    4   0.62941387  0.0000000
#> 2   E >=16   >=1    2  0.5   0.49015830  0.1392556
#> 3   H >=16 0.125  0.5  >=8   0.26899945  0.3604144
#> 4   I    8   0.5  0.5  >=8   0.26246847  0.3669454
#> 5   D  <=4 0.125    1  >=8  -0.05490015  0.6843140
#> 6   G  <=4  0.25    4  0.5  -0.09926391  0.7286778
#> 7   F >=16 0.025    1    2  -0.12782528  0.7572391
#> 8   J  <=4  0.25    1    1  -0.38529402  1.0147079
#> 9   B  <=4 0.125  0.5    4  -0.42949279  1.0589067
#> 10  A    8 0.025    1    1  -0.55426393  1.1836778
if (require("dplyr")) {
  # support for groups
  example_isolates %>%
    filter(mo_genus() == "Enterococcus" & mo_species() != "") %>%
    select(mo, TCY, carbapenems()) %>%
    group_by(mo) %>%
    mutate(d = mean_amr_distance(., where(is.rsi))) %>%
    arrange(mo, d)
}
#> ℹ Using column 'mo' as input for mo_genus()
#> ℹ Using column 'mo' as input for mo_species()
#> ℹ For carbapenems() using columns 'IPM' (imipenem) and 'MEM' (meropenem)
#> ℹ Calculating mean AMR distance based on columns "IPM", "MEM" and "TCY"
#> # A tibble: 63 × 5
#> # Groups:   mo [4]
#>    mo           TCY   IPM   MEM         d
#>    <mo>         <rsi> <rsi> <rsi>   <dbl>
#>  1 B_ENTRC_AVIM S     S     NA    NaN    
#>  2 B_ENTRC_AVIM S     S     NA    NaN    
#>  3 B_ENTRC_CSSL NA    S     NA     NA    
#>  4 B_ENTRC_FACM S     S     NA     -2.66 
#>  5 B_ENTRC_FACM S     R     R      -0.423
#>  6 B_ENTRC_FACM S     R     R      -0.423
#>  7 B_ENTRC_FACM NA    R     R       0.224
#>  8 B_ENTRC_FACM NA    R     R       0.224
#>  9 B_ENTRC_FACM NA    R     R       0.224
#> 10 B_ENTRC_FACM NA    R     R       0.224
#> # … with 53 more rows