Data set with 67,151 microorganisms

A data set containing the microbial taxonomy of six kingdoms from the Catalogue of Life. MO codes can be looked up using as.mo().

microorganisms

Format

A data.frame with 67,151 observations and 16 variables:

mo
ID of microorganism as used by this package

fullname
Full name, like "Escherichia coli"

kingdom, phylum, class, order, family, genus, species, subspecies
Taxonomic rank of the microorganism

rank
Text of the taxonomic rank of the microorganism, like "species" or "genus"

ref
Author(s) and year of concerning scientific publication

species_id
ID of the species as used by the Catalogue of Life

source
Either "CoL", "DSMZ" (see Source) or "manually added"

prevalence
Prevalence of the microorganism, see as.mo()

snomed
SNOMED code of the microorganism. Use mo_snomed() to retrieve it quickly, see mo_property().

Source

Catalogue of Life: Annual Checklist (public online taxonomic database), http://www.catalogueoflife.org (check included annual version with catalogue_of_life_version()).

Parte, A.C. (2018). LPSN — List of Prokaryotic names with Standing in Nomenclature (bacterio.net), 20 years on. International Journal of Systematic and Evolutionary Microbiology, 68, 1825-1829; doi: 10.1099/ijsem.0.002786

Leibniz Institute DSMZ-German Collection of Microorganisms and Cell Cultures, Germany, Prokaryotic Nomenclature Up-to-Date, https://www.dsmz.de/services/online-tools/prokaryotic-nomenclature-up-to-date and https://lpsn.dsmz.de (check included version with catalogue_of_life_version()).

Details

Manually added were:

11 entries of Streptococcus (beta-haemolytic: groups A, B, C, D, F, G, H, K and unspecified; other: viridans, milleri)

2 entries of Staphylococcus (coagulase-negative (CoNS) and coagulase-positive (CoPS))

3 entries of Trichomonas (Trichomonas vaginalis, and its family and genus)

1 entry of Candida (Candida krusei), that is not (yet) in the Catalogue of Life

1 entry of Blastocystis (Blastocystis hominis), although it officially does not exist (Noel et al. 2005, PMID 15634993)

5 other 'undefined' entries (unknown, unknown Gram negatives, unknown Gram positives, unknown yeast and unknown fungus)

6 families under the Enterobacterales order, according to Adeolu et al. (2016, PMID 27620848), that are not (yet) in the Catalogue of Life

7,411 species from the DSMZ (Deutsche Sammlung von Mikroorganismen und Zellkulturen) since the DSMZ contain the latest taxonomic information based on recent publications

Direct download

This data set is available as 'flat file' for use even without R - you can find the file here:

https://github.com/msberends/AMR/raw/master/data-raw/microorganisms.txt

The file in R format (with preserved data structure) can be found here:

https://github.com/msberends/AMR/raw/master/data/microorganisms.rda

About the records from DSMZ (see source)

Names of prokaryotes are defined as being validly published by the International Code of Nomenclature of Bacteria. Validly published are all names which are included in the Approved Lists of Bacterial Names and the names subsequently published in the International Journal of Systematic Bacteriology (IJSB) and, from January 2000, in the International Journal of Systematic and Evolutionary Microbiology (IJSEM) as original articles or in the validation lists. (from https://www.dsmz.de/services/online-tools/prokaryotic-nomenclature-up-to-date/complete-list-readme)

In February 2020, the DSMZ records were merged with the List of Prokaryotic names with Standing in Nomenclature (LPSN).

Catalogue of Life

This package contains the complete taxonomic tree of almost all microorganisms (~70,000 species) from the authoritative and comprehensive Catalogue of Life (http://www.catalogueoflife.org). The Catalogue of Life is the most comprehensive and authoritative global index of species currently available.

Click here for more information about the included taxa. Check which version of the Catalogue of Life was included in this package with catalogue_of_life_version().

Reference data publicly available

All reference data sets (about microorganisms, antibiotics, R/SI interpretation, EUCAST rules, etc.) in this AMR package are publicly and freely available. We continually export our data sets to formats for use in R, SPSS, SAS, Stata and Excel. We also supply flat files that are machine-readable and suitable for input in any software program, such as laboratory information systems. Please find all download links on our website, which is automatically updated with every code change.

Read more on our website!

On our website https://msberends.github.io/AMR you can find a comprehensive tutorial about how to conduct AMR analysis, the complete documentation of all functions (which reads a lot easier than here in R) and an example analysis using WHONET data. As we would like to better understand the backgrounds and needs of our users, please participate in our survey!

See also