Data Set with 70,026 Microorganisms

A data set containing the microbial taxonomy, last updated in March 2021, of six kingdoms from the Catalogue of Life (CoL) and the List of Prokaryotic names with Standing in Nomenclature (LPSN). MO codes can be looked up using as.mo().

microorganisms

Format

A data.frame with 70,026 observations and 16 variables:

mo
ID of microorganism as used by this package

fullname
Full name, like "Escherichia coli"

kingdom, phylum, class, order, family, genus, species, subspecies
Taxonomic rank of the microorganism

rank
Text of the taxonomic rank of the microorganism, like "species" or "genus"

ref
Author(s) and year of concerning scientific publication

species_id
ID of the species as used by the Catalogue of Life

source
Either "CoL", "LPSN" or "manually added" (see Source)

prevalence
Prevalence of the microorganism, see as.mo()

snomed
Systematized Nomenclature of Medicine (SNOMED) code of the microorganism, according to the US Edition of SNOMED CT from 1 September 2020 (see Source). Use mo_snomed() to retrieve it quickly, see mo_property().

Source

Catalogue of Life: 2019 Annual Checklist as currently implemented in this AMR package:

Annual Checklist (public online taxonomic database), http://www.catalogueoflife.org

List of Prokaryotic names with Standing in Nomenclature (March 2021) as currently implemented in this AMR package:

Parte, A.C., Sarda Carbasse, J., Meier-Kolthoff, J.P., Reimer, L.C. and Goker, M. (2020). List of Prokaryotic names with Standing in Nomenclature (LPSN) moves to the DSMZ. International Journal of Systematic and Evolutionary Microbiology, 70, 5607-5612; doi: 10.1099/ijsem.0.004332

Parte, A.C. (2018). LPSN — List of Prokaryotic names with Standing in Nomenclature (bacterio.net), 20 years on. International Journal of Systematic and Evolutionary Microbiology, 68, 1825-1829; doi: 10.1099/ijsem.0.002786

Parte, A.C. (2014). LPSN — List of Prokaryotic names with Standing in Nomenclature. Nucleic Acids Research, 42, Issue D1, D613–D616; doi: 10.1093/nar/gkt1111

Euzeby, J.P. (1997). List of Bacterial Names with Standing in Nomenclature: a Folder Available on the Internet. International Journal of Systematic Bacteriology, 47, 590-592; doi: 10.1099/00207713-47-2-590

US Edition of SNOMED CT from 1 September 2020 as currently implemented in this AMR package:

Retrieved from the Public Health Information Network Vocabulary Access and Distribution System (PHIN VADS), OID 2.16.840.1.114222.4.11.1009, version 12; url: https://phinvads.cdc.gov/vads/ViewValueSet.action?oid=2.16.840.1.114222.4.11.1009

Details

Please note that entries are only based on the Catalogue of Life and the LPSN (see below). Since these sources incorporate entries based on (recent) publications in the International Journal of Systematic and Evolutionary Microbiology (IJSEM), it can happen that the year of publication is sometimes later than one might expect.

For example, Staphylococcus pettenkoferi was described for the first time in Diagnostic Microbiology and Infectious Disease in 2002 (doi: 10.1016/s0732-8893(02)00399-1 ), but it was not before 2007 that a publication in IJSEM followed (doi: 10.1099/ijs.0.64381-0 ). Consequently, the AMR package returns 2007 for mo_year("S. pettenkoferi").

Manual additions

For convenience, some entries were added manually:

11 entries of Streptococcus (beta-haemolytic: groups A, B, C, D, F, G, H, K and unspecified; other: viridans, milleri)

2 entries of Staphylococcus (coagulase-negative (CoNS) and coagulase-positive (CoPS))

3 entries of Trichomonas (Trichomonas vaginalis, and its family and genus)

1 entry of Candida (Candida krusei), that is not (yet) in the Catalogue of Life

1 entry of Blastocystis (Blastocystis hominis), although it officially does not exist (Noel et al. 2005, PMID 15634993)

5 other 'undefined' entries (unknown, unknown Gram negatives, unknown Gram positives, unknown yeast and unknown fungus)

6 families under the Enterobacterales order, according to Adeolu et al. (2016, PMID 27620848), that are not (yet) in the Catalogue of Life

Direct download

This data set is available as 'flat file' for use even without R - you can find the file here:

https://github.com/msberends/AMR/raw/master/data-raw/microorganisms.txt

The file in R format (with preserved data structure) can be found here:

https://github.com/msberends/AMR/raw/master/data/microorganisms.rda

About the Records from LPSN (see Source)

The List of Prokaryotic names with Standing in Nomenclature (LPSN) provides comprehensive information on the nomenclature of prokaryotes. LPSN is a free to use service founded by Jean P. Euzeby in 1997 and later on maintained by Aidan C. Parte.

As of February 2020, the regularly augmented LPSN database at DSMZ is the basis of the new LPSN service. The new database was implemented for the Type-Strain Genome Server and augmented in 2018 to store all kinds of nomenclatural information. Data from the previous version of LPSN and from the Prokaryotic Nomenclature Up-to-date (PNU) service were imported into the new system. PNU had been established in 1993 as a service of the Leibniz Institute DSMZ, and was curated by Norbert Weiss, Manfred Kracht and Dorothea Gleim.

Catalogue of Life

This package contains the complete taxonomic tree of almost all microorganisms (~70,000 species) from the authoritative and comprehensive Catalogue of Life (CoL, http://www.catalogueoflife.org). The CoL is the most comprehensive and authoritative global index of species currently available. Nonetheless, we supplemented the CoL data with data from the List of Prokaryotic names with Standing in Nomenclature (LPSN, lpsn.dsmz.de). This supplementation is needed until the CoL+ project is finished, which we await.

Click here for more information about the included taxa. Check which versions of the CoL and LPSN were included in this package with catalogue_of_life_version().

Reference Data Publicly Available

All reference data sets (about microorganisms, antibiotics, R/SI interpretation, EUCAST rules, etc.) in this AMR package are publicly and freely available. We continually export our data sets to formats for use in R, SPSS, SAS, Stata and Excel. We also supply flat files that are machine-readable and suitable for input in any software program, such as laboratory information systems. Please find all download links on our website, which is automatically updated with every code change.

Read more on Our Website!

On our website https://msberends.github.io/AMR/ you can find a comprehensive tutorial about how to conduct AMR data analysis, the complete documentation of all functions and an example analysis using WHONET data. As we would like to better understand the backgrounds and needs of our users, please participate in our survey!

See also