Data Set with 48,787 Microorganisms

A data set containing the full microbial taxonomy (last updated: 12 September, 2022) of five kingdoms from the List of Prokaryotic names with Standing in Nomenclature (LPSN) and the Global Biodiversity Information Facility (GBIF). This data set is the backbone of this AMR package. MO codes can be looked up using as.mo().

Usage

microorganisms

Format

A tibble with 48,787 observations and 22 variables:

mo
ID of microorganism as used by this package

fullname
Full name, like "Escherichia coli". For the taxonomic ranks genus, species and subspecies, this is the 'pasted' text of genus, species, and subspecies. For all taxonomic ranks higher than genus, this is the name of the taxon.

status
Status of the taxon, either "accepted" or "synonym"

kingdom, phylum, class, order, family, genus, species, subspecies
Taxonomic rank of the microorganism

rank
Text of the taxonomic rank of the microorganism, such as "species" or "genus"

ref
Author(s) and year of related scientific publication. This contains only the first surname and year of the latest authors, e.g. "Wallis et al. 2006 emend. Smith and Jones 2018" becomes "Smith et al., 2018". This field is directly retrieved from the source specified in the column source. Moreover, accents were removed to comply with CRAN that only allows ASCII characters, e.g. "Váňová" becomes "Vanova".

lpsn
Identifier ('Record number') of the List of Prokaryotic names with Standing in Nomenclature (LPSN). This will be the first/highest LPSN identifier to keep one identifier per row. For example, Acetobacter ascendens has LPSN Record number 7864 and 11011. Only the first is available in the microorganisms data set.

lpsn_parent
LPSN identifier of the parent taxon

lpsn_renamed_to
LPSN identifier of the currently valid taxon

gbif
Identifier ('taxonID') of the Global Biodiversity Information Facility (GBIF)

gbif_parent
GBIF identifier of the parent taxon

gbif_renamed_to
GBIF identifier of the currently valid taxon

source
Either "GBIF", "LPSN" or "manually added" (see Source)

prevalence
Prevalence of the microorganism, see as.mo()

snomed
Systematized Nomenclature of Medicine (SNOMED) code of the microorganism, version of 1 July, 2021 (see Source). Use mo_snomed() to retrieve it quickly, see mo_property().

Source

Parte, AC et al. (2020). List of Prokaryotic names with Standing in Nomenclature (LPSN) moves to the DSMZ. International Journal of Systematic and Evolutionary Microbiology, 70, 5607-5612; doi:10.1099/ijsem.0.004332 . Accessed from https://lpsn.dsmz.de on 12 September, 2022.

GBIF Secretariat (November 26, 2021). GBIF Backbone Taxonomy. Checklist dataset doi:10.15468/39omei . Accessed from https://www.gbif.org on 12 September, 2022.

Public Health Information Network Vocabulary Access and Distribution System (PHIN VADS). US Edition of SNOMED CT from 1 September 2020. Value Set Name 'Microoganism', OID 2.16.840.1.114222.4.11.1009 (v12). URL: https://phinvads.cdc.gov

Details

Please note that entries are only based on the List of Prokaryotic names with Standing in Nomenclature (LPSN) and the Global Biodiversity Information Facility (GBIF) (see below). Since these sources incorporate entries based on (recent) publications in the International Journal of Systematic and Evolutionary Microbiology (IJSEM), it can happen that the year of publication is sometimes later than one might expect.

For example, Staphylococcus pettenkoferi was described for the first time in Diagnostic Microbiology and Infectious Disease in 2002 (doi:10.1016/s0732-8893(02)00399-1 ), but it was not before 2007 that a publication in IJSEM followed (doi:10.1099/ijs.0.64381-0 ). Consequently, the AMR package returns 2007 for mo_year("S. pettenkoferi").

Included Taxa

Included taxonomic data are:

All ~34,000 (sub)species from the kingdoms of Archaea and Bacteria

~7,400 (sub)species from the kingdom of Fungi. The kingdom of Fungi is a very large taxon with almost 300,000 different (sub)species, of which most are not microbial (but rather macroscopic, like mushrooms). Because of this, not all fungi fit the scope of this package. Only relevant fungi are covered (such as all species of Aspergillus, Candida, Cryptococcus, Histoplasma, Pneumocystis, Saccharomyces and Trichophyton).

~4,900 (sub)species from the kingdom of Protozoa

~1,500 (sub)species from ~50 other relevant genera from the kingdom of Animalia (such as Strongyloides and Taenia)

All ~9,400 previously accepted names of all included (sub)species (these were taxonomically renamed)

The complete taxonomic tree of all included (sub)species: from kingdom to subspecies

The identifier of the parent taxons

The year and first author of the related scientific publication

Manual additions

For convenience, some entries were added manually:

11 entries of Streptococcus (beta-haemolytic: groups A, B, C, D, F, G, H, K and unspecified; other: viridans, milleri)

2 entries of Staphylococcus (coagulase-negative (CoNS) and coagulase-positive (CoPS))

1 entry of Blastocystis (B. hominis), although it officially does not exist (Noel et al. 2005, PMID 15634993)

1 entry of Moraxella (M. catarrhalis), which was formally named Branhamella catarrhalis (Catlin, 1970) though this change was never accepted within the field of clinical microbiology

5 other 'undefined' entries (unknown, unknown Gram negatives, unknown Gram positives, unknown yeast and unknown fungus)

The syntax used to transform the original data to a cleansed R format, can be found here: https://github.com/msberends/AMR/blob/main/data-raw/reproduction_of_microorganisms.R.

Direct download

Like all data sets in this package, this data set is publicly available for download in the following formats: R, MS Excel, Apache Feather, Apache Parquet, SPSS, SAS, and Stata. Please visit our website for the download links. The actual files are of course available on our GitHub repository.

About the Records from LPSN (see Source)

LPSN is the main source for bacteriological taxonomy of this AMR package.

The List of Prokaryotic names with Standing in Nomenclature (LPSN) provides comprehensive information on the nomenclature of prokaryotes. LPSN is a free to use service founded by Jean P. Euzeby in 1997 and later on maintained by Aidan C. Parte.

See also

Examples

microorganisms
#> # A tibble: 48,787 × 22
#>    mo               fullname      status kingdom phylum class order family genus
#>    <mo>             <chr>         <chr>  <chr>   <chr>  <chr> <chr> <chr>  <chr>
#>  1 F_FUNGUS         (unknown fun… accep… Fungi   (unkn… "(un… "(un… "(unk… "(un…
#>  2 B_GRAMN          (unknown Gra… accep… Bacter… (unkn… "(un… "(un… "(unk… "(un…
#>  3 B_GRAMP          (unknown Gra… accep… Bacter… (unkn… "(un… "(un… "(unk… "(un…
#>  4   UNKNOWN        (unknown nam… accep… (unkno… (unkn… "(un… "(un… "(unk… "(un…
#>  5 F_YEAST          (unknown yea… accep… Fungi   (unkn… "(un… "(un… "(unk… "(un…
#>  6 B_[CLS]_ABDTBCTR Abditibacter… accep… Bacter… Armat… "Abd… ""    ""     ""   
#>  7 B_[FAM]_ABDTBCTR Abditibacter… accep… Bacter… Abdit… "Abd… "Abd… "Abdi… ""   
#>  8 B_[ORD]_ABDTBCTR Abditibacter… accep… Bacter… Abdit… "Abd… "Abd… ""     ""   
#>  9 B_[CLS]_ADTBCTRA Abditibacter… accep… Bacter… Abdit… "Abd… ""    ""     ""   
#> 10 B_[PHL]_ABDTBCTR Abditibacter… accep… Bacter… Abdit… ""    ""    ""     ""   
#> # … with 48,777 more rows, and 13 more variables: species <chr>,
#> #   subspecies <chr>, rank <chr>, ref <chr>, source <chr>, lpsn <chr>,
#> #   lpsn_parent <chr>, lpsn_renamed_to <chr>, gbif <chr>, gbif_parent <chr>,
#> #   gbif_renamed_to <chr>, prevalence <dbl>, snomed <list>