Data set with ~70,000 microorganisms

A data set containing the microbial taxonomy of six kingdoms from the Catalogue of Life. MO codes can be looked up using as.mo().

microorganisms

Format

A data.frame with 69,447 observations and 16 variables:

mo
ID of microorganism as used by this package

col_id
Catalogue of Life ID

fullname
Full name, like "Escherichia coli"

kingdom, phylum, class, order, family, genus, species, subspecies
Taxonomic rank of the microorganism

rank
Text of the taxonomic rank of the microorganism, like "species" or "genus"

ref
Author(s) and year of concerning scientific publication

species_id
ID of the species as used by the Catalogue of Life

source
Either "CoL", "DSMZ" (see Source) or "manually added"

prevalence
Prevalence of the microorganism, see as.mo()

Source

Catalogue of Life: Annual Checklist (public online taxonomic database), http://www.catalogueoflife.org (check included annual version with catalogue_of_life_version()).

Leibniz Institute DSMZ-German Collection of Microorganisms and Cell Cultures, Germany, Prokaryotic Nomenclature Up-to-Date, http://www.dsmz.de/bacterial-diversity/prokaryotic-nomenclature-up-to-date (check included version with catalogue_of_life_version()).

Details

Manually added were:

11 entries of Streptococcus (beta-haemolytic: groups A, B, C, D, F, G, H, K and unspecified; other: viridans, milleri)

2 entries of Staphylococcus (coagulase-negative (CoNS) and coagulase-positive (CoPS))

3 entries of Trichomonas (Trichomonas vaginalis, and its family and genus)

1 entry of Blastocystis (Blastocystis hominis), although it officially does not exist (Noel et al. 2005, PMID 15634993)

5 other 'undefined' entries (unknown, unknown Gram negatives, unknown Gram positives, unknown yeast and unknown fungus)

6 families under the Enterobacterales order, according to Adeolu et al. (2016, PMID 27620848), that are not in the Catalogue of Life

12,600 species from the DSMZ (Deutsche Sammlung von Mikroorganismen und Zellkulturen) since the DSMZ contain the latest taxonomic information based on recent publications

About the records from DSMZ (see source)

Names of prokaryotes are defined as being validly published by the International Code of Nomenclature of Bacteria. Validly published are all names which are included in the Approved Lists of Bacterial Names and the names subsequently published in the International Journal of Systematic Bacteriology (IJSB) and, from January 2000, in the International Journal of Systematic and Evolutionary Microbiology (IJSEM) as original articles or in the validation lists.

From: https://www.dsmz.de/support/bacterial-nomenclature-up-to-date-downloads/readme.html

Catalogue of Life

This package contains the complete taxonomic tree of almost all microorganisms (~70,000 species) from the authoritative and comprehensive Catalogue of Life (http://www.catalogueoflife.org). The Catalogue of Life is the most comprehensive and authoritative global index of species currently available.

Click here for more information about the included taxa. Check which version of the Catalogue of Life was included in this package with catalogue_of_life_version().

Read more on our website!

On our website https://msberends.gitlab.io/AMR you can find a tutorial about how to conduct AMR analysis, the complete documentation of all functions (which reads a lot easier than here in R) and an example analysis using WHONET data.

See also