Data Set with 4 957 Common Microorganism Codes

A data set containing commonly used codes for microorganisms, from laboratory systems and WHONET. Define your own with set_mo_source(). They will all be searched when using as.mo() and consequently all the mo_* functions.

Usage

microorganisms.codes

Format

A tibble with 4 957 observations and 2 variables:

code
Commonly used code of a microorganism. This is a unique identifier.

mo
ID of the microorganism in the microorganisms data set

Details

Like all data sets in this package, this data set is publicly available for download in the following formats: R, MS Excel, Apache Feather, Apache Parquet, SPSS, SAS, and Stata. Please visit our website for the download links. The actual files are of course available on our GitHub repository.

See also

Examples

microorganisms.codes
#> Error in .Call(Cchin, x, table): NULL value passed as symbol address

# 'ECO' or 'eco' is the WHONET code for E. coli:
microorganisms.codes[microorganisms.codes$code == "ECO", ]
#> Error in .Call(Cchin, x, table): NULL value passed as symbol address

# and therefore, 'eco' will be understood as E. coli in this package:
mo_info("eco")
#> Error in .Call(Cchin, x, table): NULL value passed as symbol address

# works for all AMR functions:
mo_is_intrinsic_resistant("eco", ab = "vancomycin")
#> Error in .Call(Cchin, x, table): NULL value passed as symbol address