Data Set with 521 Microorganisms In Species Groups

A data set containing species groups and microbiological complexes, which are used in the clinical breakpoints table.

Usage

microorganisms.groups

Format

A tibble with 521 observations and 4 variables:

mo_group
ID of the species group / microbiological complex

mo
ID of the microorganism belonging in the species group / microbiological complex

mo_group_name
Name of the species group / microbiological complex, as retrieved with mo_name()

mo_name
Name of the microorganism belonging in the species group / microbiological complex, as retrieved with mo_name()

Details

Like all data sets in this package, this data set is publicly available for download in the following formats: R, MS Excel, Apache Feather, Apache Parquet, SPSS, SAS, and Stata. Please visit our website for the download links. The actual files are of course available on our GitHub repository.

See also

Examples

microorganisms.groups
#> # A tibble: 521 × 4
#>    mo_group       mo           mo_group_name                   mo_name          
#>    <mo>           <mo>         <chr>                           <chr>            
#>  1 B_ACNTB_BMNN-C B_ACNTB_BMNN Acinetobacter baumannii complex Acinetobacter ba…
#>  2 B_ACNTB_BMNN-C B_ACNTB_CLCC Acinetobacter baumannii complex Acinetobacter ca…
#>  3 B_ACNTB_BMNN-C B_ACNTB_DJKS Acinetobacter baumannii complex Acinetobacter di…
#>  4 B_ACNTB_BMNN-C B_ACNTB_NSCM Acinetobacter baumannii complex Acinetobacter no…
#>  5 B_ACNTB_BMNN-C B_ACNTB_PITT Acinetobacter baumannii complex Acinetobacter pi…
#>  6 B_ACNTB_BMNN-C B_ACNTB_SFRT Acinetobacter baumannii complex Acinetobacter se…
#>  7 B_BCTRD_FRGL-C B_BCTRD_FRGL Bacteroides fragilis complex    Bacteroides frag…
#>  8 B_BCTRD_FRGL-C B_BCTRD_OVTS Bacteroides fragilis complex    Bacteroides ovat…
#>  9 B_BCTRD_FRGL-C B_BCTRD_THTT Bacteroides fragilis complex    Bacteroides thet…
#> 10 B_BCTRD_FRGL-C B_BCTRD_VLGT Bacteroides fragilis complex    Bacteroides vulg…
#> # ℹ 511 more rows

# these are all species in the Bacteroides fragilis group, as per WHONET:
microorganisms.groups[microorganisms.groups$mo_group == "B_BCTRD_FRGL-C", ]
#> # A tibble: 5 × 4
#>   mo_group       mo           mo_group_name                mo_name              
#>   <mo>           <mo>         <chr>                        <chr>                
#> 1 B_BCTRD_FRGL-C B_BCTRD_FRGL Bacteroides fragilis complex Bacteroides fragilis 
#> 2 B_BCTRD_FRGL-C B_BCTRD_OVTS Bacteroides fragilis complex Bacteroides ovatus   
#> 3 B_BCTRD_FRGL-C B_BCTRD_THTT Bacteroides fragilis complex Bacteroides thetaiot…
#> 4 B_BCTRD_FRGL-C B_BCTRD_VLGT Bacteroides fragilis complex Bacteroides vulgatus 
#> 5 B_BCTRD_FRGL-C B_PRBCT_DSTS Bacteroides fragilis complex Parabacteroides dist…