Data set with previously accepted taxonomic names

A data set containing old (previously valid or accepted) taxonomic names according to the Catalogue of Life. This data set is used internally by as.mo.

microorganisms.old

Format

A data.frame with 21,342 observations and 4 variables:

col_id

Catalogue of Life ID that was originally given

col_id_new

New Catalogue of Life ID that responds to an entry in the microorganisms data set

fullname

Old full taxonomic name of the microorganism

ref

Author(s) and year of concerning scientific publication

Source

Catalogue of Life: Annual Checklist (public online taxonomic database), http://www.catalogueoflife.org (check included annual version with catalogue_of_life_version()).

Catalogue of Life

This package contains the complete taxonomic tree of almost all microorganisms (~65,000 species) from the authoritative and comprehensive Catalogue of Life (http://www.catalogueoflife.org). The Catalogue of Life is the most comprehensive and authoritative global index of species currently available.

Click here for more information about the included taxa. The Catalogue of Life releases updates annually; check which version was included in this package with catalogue_of_life_version().

Read more on our website!

On our website https://msberends.gitlab.io/AMR you can find a tutorial about how to conduct AMR analysis, the complete documentation of all functions (which reads a lot easier than here in R) and an example analysis using WHONET data.

See also