Property of a microorganism

Use these functions to return a specific property of a microorganism from the microorganisms data set. All input values will be evaluated internally with as.mo.

mo_name(x, language = get_locale(), ...)

mo_fullname(x, language = get_locale(), ...)

mo_shortname(x, language = get_locale(), ...)

mo_subspecies(x, language = get_locale(), ...)

mo_species(x, language = get_locale(), ...)

mo_genus(x, language = get_locale(), ...)

mo_family(x, language = get_locale(), ...)

mo_order(x, language = get_locale(), ...)

mo_class(x, language = get_locale(), ...)

mo_phylum(x, language = get_locale(), ...)

mo_kingdom(x, language = get_locale(), ...)

mo_type(x, language = get_locale(), ...)

mo_gramstain(x, language = get_locale(), ...)

mo_ref(x, ...)

mo_authors(x, ...)

mo_year(x, ...)

mo_rank(x, ...)

mo_taxonomy(x, language = get_locale(), ...)

mo_url(x, open = FALSE, ...)

mo_property(x, property = "fullname", language = get_locale(), ...)

Arguments

x
any (vector of) text that can be coerced to a valid microorganism code with as.mo

language
language of the returned text, defaults to system language (see get_locale) and can also be set with getOption("AMR_locale"). Use language = NULL or language = "" to prevent translation.

...
other parameters passed on to as.mo

open
browse the URL using browseURL()

property
one of the column names of the microorganisms data set or "shortname"

Value

An integer in case of mo_year

A list in case of mo_taxonomy

A named character in case of mo_url

A character in all other cases

Details

All functions will return the most recently known taxonomic property according to the Catalogue of Life, except for mo_ref, mo_authors and mo_year. This leads to the following results:

mo_name("Chlamydia psittaci") will return "Chlamydophila psittaci" (with a warning about the renaming)

mo_ref("Chlamydia psittaci") will return "Page, 1968" (with a warning about the renaming)

mo_ref("Chlamydophila psittaci") will return "Everett et al., 1999" (without a warning)

The Gram stain - mo_gramstain() - will be determined on the taxonomic kingdom and phylum. According to Cavalier-Smith (2002) who defined subkingdoms Negibacteria and Posibacteria, only these phyla are Posibacteria: Actinobacteria, Chloroflexi, Firmicutes and Tenericutes. These bacteria are considered Gram positive - all other bacteria are considered Gram negative. Species outside the kingdom of Bacteria will return a value NA.

All output will be translated where possible.

The function mo_url() will return the direct URL to the online database entry, which also shows the scientific reference of the concerned species.

Catalogue of Life

This package contains the complete taxonomic tree of almost all microorganisms (~65,000 species) from the authoritative and comprehensive Catalogue of Life (http://www.catalogueoflife.org). The Catalogue of Life is the most comprehensive and authoritative global index of species currently available.

Click here for more information about the included taxa. The Catalogue of Life releases updates annually; check which version was included in this package with catalogue_of_life_version().

Source

[1] Becker K et al. Coagulase-Negative Staphylococci. 2014. Clin Microbiol Rev. 27(4): 870–926. https://dx.doi.org/10.1128/CMR.00109-13

[2] Becker K et al. Implications of identifying the recently defined members of the S. aureus complex, S. argenteus and S. schweitzeri: A position paper of members of the ESCMID Study Group for staphylococci and Staphylococcal Diseases (ESGS). 2019. Clin Microbiol Infect. https://doi.org/10.1016/j.cmi.2019.02.028

[3] Lancefield RC A serological differentiation of human and other groups of hemolytic streptococci. 1933. J Exp Med. 57(4): 571–95. https://dx.doi.org/10.1084/jem.57.4.571

[4] Catalogue of Life: Annual Checklist (public online taxonomic database), www.catalogueoflife.org (check included annual version with catalogue_of_life_version()).

Read more on our website!

On our website https://msberends.gitlab.io/AMR you can find a comprehensive tutorial about how to conduct AMR analysis, the complete documentation of all functions (which reads a lot easier than here in R) and an example analysis using WHONET data.

See also

Examples

# NOT RUN {
## taxonomic tree
mo_kingdom("E. coli")         # "Bacteria"
mo_phylum("E. coli")          # "Proteobacteria"
mo_class("E. coli")           # "Gammaproteobacteria"
mo_order("E. coli")           # "Enterobacteriales"
mo_family("E. coli")          # "Enterobacteriaceae"
mo_genus("E. coli")           # "Escherichia"
mo_species("E. coli")         # "coli"
mo_subspecies("E. coli")      # ""

## colloquial properties
mo_name("E. coli")            # "Escherichia coli"
mo_fullname("E. coli")        # "Escherichia coli", same as mo_name()
mo_shortname("E. coli")       # "E. coli"

## other properties
mo_gramstain("E. coli")       # "Gram negative"
mo_type("E. coli")            # "Bacteria" (equal to kingdom)
mo_rank("E. coli")            # "species"
mo_url("E. coli")             # get the direct url to the online database entry

## scientific reference
mo_ref("E. coli")             # "Castellani et al., 1919"
mo_authors("E. coli")         # "Castellani et al."
mo_year("E. coli")            # 1919

# Abbreviations known in the field
mo_genus("MRSA")              # "Staphylococcus"
mo_species("MRSA")            # "aureus"
mo_shortname("MRSA")          # "S. aureus"
mo_gramstain("MRSA")          # "Gram positive"

mo_genus("VISA")              # "Staphylococcus"
mo_species("VISA")            # "aureus"

# Known subspecies
mo_name("doylei")             # "Campylobacter jejuni doylei"
mo_genus("doylei")            # "Campylobacter"
mo_species("doylei")          # "jejuni"
mo_subspecies("doylei")       # "doylei"

mo_fullname("K. pneu rh")     # "Klebsiella pneumoniae rhinoscleromatis"
mo_shortname("K. pneu rh")    # "K. pneumoniae"

# Becker classification, see ?as.mo
mo_fullname("S. epi")                     # "Staphylococcus epidermidis"
mo_fullname("S. epi", Becker = TRUE)      # "Coagulase-negative Staphylococcus (CoNS)"
mo_shortname("S. epi")                    # "S. epidermidis"
mo_shortname("S. epi", Becker = TRUE)     # "CoNS"

# Lancefield classification, see ?as.mo
mo_fullname("S. pyo")                     # "Streptococcus pyogenes"
mo_fullname("S. pyo", Lancefield = TRUE)  # "Streptococcus group A"
mo_shortname("S. pyo")                    # "S. pyogenes"
mo_shortname("S. pyo", Lancefield = TRUE) # "GAS" ('Group A streptococci')

# language support for German, Dutch, Spanish, Portuguese, Italian and French
mo_gramstain("E. coli", language = "de")  # "Gramnegativ"
mo_gramstain("E. coli", language = "nl")  # "Gram-negatief"
mo_gramstain("E. coli", language = "es")  # "Gram negativo"

# mo_type is equal to mo_kingdom, but mo_kingdom will remain official
mo_kingdom("E. coli")                     # "Bacteria" on a German system
mo_type("E. coli")                        # "Bakterien" on a German system
mo_type("E. coli")                        # "Bacteria" on an English system

mo_fullname("S. pyogenes",
            Lancefield = TRUE,
            language = "de")              # "Streptococcus Gruppe A"
mo_fullname("S. pyogenes",
            Lancefield = TRUE,
            language = "nl")              # "Streptococcus groep A"

# get a list with the complete taxonomy (from kingdom to subspecies)
mo_taxonomy("E. coli")
# }