Random MIC Values/Disk Zones/SIR Generation

These functions can be used for generating random MIC values and disk diffusion diameters, for AMR data analysis practice. By providing a microorganism and antimicrobial drug, the generated results will reflect reality as much as possible.

Usage

random_mic(size = NULL, mo = NULL, ab = NULL, ...)

random_disk(size = NULL, mo = NULL, ab = NULL, ...)

random_sir(size = NULL, prob_SIR = c(0.33, 0.33, 0.33), ...)

Arguments

size

Desired size of the returned vector. If used in a data.frame call or dplyr verb, will get the current (group) size if left blank.

mo

Any character that can be coerced to a valid microorganism code with as.mo().

ab

Any character that can be coerced to a valid antimicrobial drug code with as.ab().

...

Ignored, only in place to allow future extensions.

prob_SIR

A vector of length 3: the probabilities for "S" (1st value), "I" (2nd value) and "R" (3rd value).

Value

class mic for random_mic() (see as.mic()) and class disk for random_disk() (see as.disk())

Details

The base R function sample() is used for generating values.

Generated values are based on the EUCAST 2024 guideline as implemented in the clinical_breakpoints data set. To create specific generated values per bug or drug, set the mo and/or ab argument.

Examples

random_mic(25)
#> Class 'mic'
#>  [1] <=0.005 0.25    1       0.0625  128     0.01    1       4       128    
#> [10] <=0.005 0.25    0.5     0.125   0.01    0.25    2       8       16     
#> [19] 0.125   0.5     >=256   >=256   >=256   0.25    16     
random_disk(25)
#> Class 'disk'
#>  [1] 24 14 46 41  9  8 22 41 50 38 48 48 28 43 47 30 13 13 38 14 48 47 21  6 22
random_sir(25)
#> Class 'sir'
#>  [1] I I I I I I S R R S I S S R I R R R I S S R R I R

# \donttest{
# make the random generation more realistic by setting a bug and/or drug:
random_mic(25, "Klebsiella pneumoniae") # range 0.0625-64
#> Class 'mic'
#>  [1] 0.025  4      1      0.0625 0.005  0.002  2      0.002  >=128  0.0625
#> [11] 0.125  0.5    0.002  0.01   0.001  0.002  0.01   32     64     0.005 
#> [21] 0.01   1      4      1      32    
random_mic(25, "Klebsiella pneumoniae", "meropenem") # range 0.0625-16
#> Class 'mic'
#>  [1] <=1 4   4   2   32  4   32  16  <=1 32  8   8   32  8   4   32  <=1 <=1 16 
#> [20] 8   32  8   16  4   8  
random_mic(25, "Streptococcus pneumoniae", "meropenem") # range 0.0625-4
#> Class 'mic'
#>  [1] 0.125    >=16     0.125    >=16     2        0.5      1        0.5     
#>  [9] 0.125    0.5      <=0.0625 0.25     0.25     0.5      >=16     >=16    
#> [17] 1        >=16     8        0.25     8        0.25     >=16     >=16    
#> [25] 0.125   

random_disk(25, "Klebsiella pneumoniae") # range 8-50
#> Class 'disk'
#>  [1] 42 36 49 13 45 50 37 33 30 14 23 15 30  9 21 47 32 17 48 47 24 23 14 27 48
random_disk(25, "Klebsiella pneumoniae", "ampicillin") # range 11-17
#> Class 'disk'
#>  [1] 13 16 13 16 13 11 17 12 16 14 14 15 11 16 12 13 14 12 13 12 15 12 12 13 17
random_disk(25, "Streptococcus pneumoniae", "ampicillin") # range 12-27
#> Class 'disk'
#>  [1] 16 20 25 24 21 22 21 16 15 18 18 25 19 26 27 26 21 21 19 25 15 22 19 24 26
# }