Random MIC Values/Disk Zones/SIR Generation

These functions can be used for generating random MIC values and disk diffusion diameters, for AMR data analysis practice. By providing a microorganism and antimicrobial drug, the generated results will reflect reality as much as possible.

Usage

random_mic(size = NULL, mo = NULL, ab = NULL, ...)

random_disk(size = NULL, mo = NULL, ab = NULL, ...)

random_sir(size = NULL, prob_SIR = c(0.33, 0.33, 0.33), ...)

Arguments

size

desired size of the returned vector. If used in a data.frame call or dplyr verb, will get the current (group) size if left blank.

mo

any character that can be coerced to a valid microorganism code with as.mo()

ab

any character that can be coerced to a valid antimicrobial drug code with as.ab()

...

ignored, only in place to allow future extensions

prob_SIR

a vector of length 3: the probabilities for "S" (1st value), "I" (2nd value) and "R" (3rd value)

Value

class mic for random_mic() (see as.mic()) and class disk for random_disk() (see as.disk())

Details

The base R function sample() is used for generating values.

Generated values are based on the EUCAST 2022 guideline as implemented in the clinical_breakpoints data set. To create specific generated values per bug or drug, set the mo and/or ab argument.

Examples

random_mic(25)
#> Class 'mic'
#>  [1] 0.005 0.001 0.005 >=256 64    0.25  16    32    32    8     2     0.5  
#> [13] 0.5   >=256 0.005 64    0.01  1     0.005 0.125 64    0.002 0.002 8    
#> [25] 8    
random_disk(25)
#> Class 'disk'
#>  [1] 12  8 42 46 19 36 43 46  8 35  8 50 40 40 12 42 44 42 42 32 41 45 48 44 44
random_sir(25)
#> Class 'sir'
#>  [1] R I S S I R R I I I R S I S R R S I R S R S I R S

# \donttest{
# make the random generation more realistic by setting a bug and/or drug:
random_mic(25, "Klebsiella pneumoniae") # range 0.0625-64
#> Class 'mic'
#>  [1] 0.025   2       <=0.001 0.25    2       <=0.001 0.0625  64      32     
#> [10] 0.0625  2       16      0.25    0.125   16      1       0.01    2      
#> [19] 0.25    2       2       32      0.125   16      4      
random_mic(25, "Klebsiella pneumoniae", "meropenem") # range 0.0625-16
#> Class 'mic'
#>  [1] 16   32   32   0.5  0.25 2    16   1    8    16   8    0.5  16   32   8   
#> [16] 8    8    2    4    32   4    0.5  16   2    16  
random_mic(25, "Streptococcus pneumoniae", "meropenem") # range 0.0625-4
#> Class 'mic'
#>  [1] 0.5      8        <=0.0625 0.125    8        2        0.125    4       
#>  [9] 4        0.25     0.5      0.125    0.125    2        <=0.0625 0.25    
#> [17] 0.25     0.5      8        0.5      0.25     8        4        4       
#> [25] 0.5     

random_disk(25, "Klebsiella pneumoniae") # range 8-50
#> Class 'disk'
#>  [1] 18 14 20 11 18 35 34 48 46 34 47 41 22 14 40 32  8 43 44 10 50 43 23 19 45
random_disk(25, "Klebsiella pneumoniae", "ampicillin") # range 11-17
#> Class 'disk'
#>  [1] 17 14 17 13 17 11 15 14 11 13 17 15 11 12 13 13 17 16 14 13 13 15 17 14 17
random_disk(25, "Streptococcus pneumoniae", "ampicillin") # range 12-27
#> Class 'disk'
#>  [1] 27 26 25 23 19 15 24 18 24 18 25 25 25 26 19 26 18 16 26 23 20 20 16 21 23
# }