Random MIC Values/Disk Zones/RSI Generation

These functions can be used for generating random MIC values and disk diffusion diameters, for AMR data analysis practice. By providing a microorganism and antimicrobial agent, the generated results will reflect reality as much as possible.

Usage

random_mic(size = NULL, mo = NULL, ab = NULL, ...)

random_disk(size = NULL, mo = NULL, ab = NULL, ...)

random_rsi(size = NULL, prob_RSI = c(0.33, 0.33, 0.33), ...)

Arguments

size

desired size of the returned vector. If used in a data.frame call or dplyr verb, will get the current (group) size if left blank.

mo

any character that can be coerced to a valid microorganism code with as.mo()

ab

any character that can be coerced to a valid antimicrobial agent code with as.ab()

...

ignored, only in place to allow future extensions

prob_RSI

a vector of length 3: the probabilities for "R" (1st value), "S" (2nd value) and "I" (3rd value)

Value

class mic for random_mic() (see as.mic()) and class disk for random_disk() (see as.disk())

Details

The base R function sample() is used for generating values.

Generated values are based on the EUCAST 2022 guideline as implemented in the rsi_translation data set. To create specific generated values per bug or drug, set the mo and/or ab argument.

Examples

random_mic(25)
#> Class 'mic'
#>  [1] >=256 0.125 1     0.01  0.01  0.001 32    4     32    0.01  0.125 0.001
#> [13] 0.25  0.002 0.025 0.125 4     0.002 0.01  128   0.5   0.005 0.125 0.001
#> [25] 16   
random_disk(25)
#> Class 'disk'
#>  [1] 29 26 25 49 46 40 38 32 38 49 15 50 16 40 15 27 12  6 22 11 13 17 26 24 47
random_rsi(25)
#> Class 'rsi'
#>  [1] R S S R I S S I R I S R I R I R R S S I R R S R I

# \donttest{
# make the random generation more realistic by setting a bug and/or drug:
random_mic(25, "Klebsiella pneumoniae") # range 0.0625-64
#> Class 'mic'
#>  [1] 0.25   0.125  0.125  32     16     0.5    0.001  8      0.025  4     
#> [11] 0.5    0.001  16     0.01   16     1      >=64   0.025  0.025  0.25  
#> [21] 32     0.25   32     0.005  0.0625
random_mic(25, "Klebsiella pneumoniae", "meropenem") # range 0.0625-16
#> Class 'mic'
#>  [1] <=1 16  16  4   2   <=1 <=1 8   8   2   <=1 16  <=1 2   <=1 4   16  16  8  
#> [20] 8   16  16  4   2   4  
random_mic(25, "Streptococcus pneumoniae", "meropenem") # range 0.0625-4
#> Class 'mic'
#>  [1] 2      0.025  1      2      2      0.5    2      2      0.25   0.25  
#> [11] 0.25   1      2      0.125  0.0625 0.125  0.25   0.025  2      >=4   
#> [21] 0.025  0.025  >=4    0.025  0.25  

random_disk(25, "Klebsiella pneumoniae") # range 8-50
#> Class 'disk'
#>  [1] 22 21 48 15 14 28 48 36 43 24 14 42 41 43  8 10 31  9 47 41 26 41 19 50 35
random_disk(25, "Klebsiella pneumoniae", "ampicillin") # range 11-17
#> Class 'disk'
#>  [1] 13 17 14 16 12 17 15 11 16 13 16 15 17 11 15 14 11 15 17 16 13 11 16 14 11
random_disk(25, "Streptococcus pneumoniae", "ampicillin") # range 12-27
#> Class 'disk'
#>  [1] 15 20 26 20 21 21 24 19 21 20 17 27 21 20 17 20 23 15 21 17 17 24 16 17 16
# }