Random MIC Values/Disk Zones/SIR Generation

These functions can be used for generating random MIC values and disk diffusion diameters, for AMR data analysis practice. By providing a microorganism and antimicrobial drug, the generated results will reflect reality as much as possible.

Usage

random_mic(size = NULL, mo = NULL, ab = NULL, ...)

random_disk(size = NULL, mo = NULL, ab = NULL, ...)

random_sir(size = NULL, prob_SIR = c(0.33, 0.33, 0.33), ...)

Arguments

size

desired size of the returned vector. If used in a data.frame call or dplyr verb, will get the current (group) size if left blank.

mo

any character that can be coerced to a valid microorganism code with as.mo()

ab

any character that can be coerced to a valid antimicrobial drug code with as.ab()

...

ignored, only in place to allow future extensions

prob_SIR

a vector of length 3: the probabilities for "S" (1st value), "I" (2nd value) and "R" (3rd value)

Value

class mic for random_mic() (see as.mic()) and class disk for random_disk() (see as.disk())

Details

The base R function sample() is used for generating values.

Generated values are based on the EUCAST 2024 guideline as implemented in the clinical_breakpoints data set. To create specific generated values per bug or drug, set the mo and/or ab argument.

Examples

random_mic(25)
#> Class 'mic'
#>  [1] 0.025   8       0.002   0.005   0.25    >=256   0.01    0.002   64     
#> [10] 0.5     4       32      0.002   16      >=256   0.01    0.002   0.0625 
#> [19] 0.002   64      8       8       2       <=0.001 0.125  
random_disk(25)
#> Class 'disk'
#>  [1] 43  7 48 40 19 50 50 20 44 14 16 33 31 17 30  6 38  7  6 12 16 11 43 43 45
random_sir(25)
#> Class 'sir'
#>  [1] R S R R S S R I S R S S R I I R S I R R R R I R S

# \donttest{
# make the random generation more realistic by setting a bug and/or drug:
random_mic(25, "Klebsiella pneumoniae") # range 0.0625-64
#> Class 'mic'
#>  [1] 0.025  0.25   16     8      0.001  4      16     0.025  0.25   8     
#> [11] 0.5    0.005  0.0625 2      0.125  0.002  1      16     16     0.005 
#> [21] 0.001  >=32   0.002  0.0625 0.125 
random_mic(25, "Klebsiella pneumoniae", "meropenem") # range 0.0625-16
#> Class 'mic'
#>  [1] 4    8    2    8    8    16   8    8    1    0.5  1    >=32 2    4    1   
#> [16] >=32 4    2    4    4    2    16   4    16   2   
random_mic(25, "Streptococcus pneumoniae", "meropenem") # range 0.0625-4
#> Class 'mic'
#>  [1] <=0.0625 <=0.0625 <=0.0625 0.5      0.125    0.25     0.5      2       
#>  [9] 0.125    0.125    0.5      4        <=0.0625 2        0.25     4       
#> [17] 2        0.25     1        2        2        <=0.0625 2        <=0.0625
#> [25] 0.5     

random_disk(25, "Klebsiella pneumoniae") # range 8-50
#> Class 'disk'
#>  [1] 30 43 22 49 40  8 14 36 49 23 12  9 17 25 43 46 30 49 34 22 50 24 39 38 46
random_disk(25, "Klebsiella pneumoniae", "ampicillin") # range 11-17
#> Class 'disk'
#>  [1] 17 15 16 14 17 14 12 13 13 17 17 12 14 16 15 14 16 14 16 13 11 16 17 13 13
random_disk(25, "Streptococcus pneumoniae", "ampicillin") # range 12-27
#> Class 'disk'
#>  [1] 25 22 21 23 19 20 21 25 20 19 21 19 27 22 26 25 27 21 17 26 18 23 23 26 27
# }