Random MIC Values/Disk Zones/SIR Generation

These functions can be used for generating random MIC values and disk diffusion diameters, for AMR data analysis practice. By providing a microorganism and antimicrobial drug, the generated results will reflect reality as much as possible.

Usage

random_mic(size = NULL, mo = NULL, ab = NULL, ...)

random_disk(size = NULL, mo = NULL, ab = NULL, ...)

random_sir(size = NULL, prob_SIR = c(0.33, 0.33, 0.33), ...)

Arguments

size

desired size of the returned vector. If used in a data.frame call or dplyr verb, will get the current (group) size if left blank.

mo

any character that can be coerced to a valid microorganism code with as.mo()

ab

any character that can be coerced to a valid antimicrobial drug code with as.ab()

...

ignored, only in place to allow future extensions

prob_SIR

a vector of length 3: the probabilities for "S" (1st value), "I" (2nd value) and "R" (3rd value)

Value

class mic for random_mic() (see as.mic()) and class disk for random_disk() (see as.disk())

Details

The base R function sample() is used for generating values.

Generated values are based on the EUCAST 2024 guideline as implemented in the clinical_breakpoints data set. To create specific generated values per bug or drug, set the mo and/or ab argument.

Examples

random_mic(25)
#> Class 'mic'
#>  [1] 16     64     256    0.002  1      0.001  64     0.25   1      16    
#> [11] 0.5    8      8      0.0625 1      0.5    0.025  256    256    1     
#> [21] 0.01   1      64     32     4     
random_disk(25)
#> Class 'disk'
#>  [1] 49 47 14 33 49 43 41 41 43 31 20 43 42 11 18 42  8 42 16 25 42  8 15 30 29
random_sir(25)
#> Class 'sir'
#>  [1] I S S I S I I S R S S I I S S I I S I R S S R I S

# \donttest{
# make the random generation more realistic by setting a bug and/or drug:
random_mic(25, "Klebsiella pneumoniae") # range 0.0625-64
#> Class 'mic'
#>  [1] 64     0.025  2      0.25   32     64     0.002  32     0.01   0.001 
#> [11] 0.25   0.125  0.0625 64     128    0.002  128    128    1      0.002 
#> [21] 1      0.005  0.001  4      0.001 
random_mic(25, "Klebsiella pneumoniae", "meropenem") # range 0.0625-16
#> Class 'mic'
#>  [1] 1    2    1    >=16 8    1    >=16 >=16 2    1    2    1    1    8    4   
#> [16] 1    1    4    2    >=16 >=16 2    4    2    1   
random_mic(25, "Streptococcus pneumoniae", "meropenem") # range 0.0625-4
#> Class 'mic'
#>  [1] 2     0.25  >=8   >=8   0.5   1     1     0.125 0.25  1     2     1    
#> [13] 0.25  >=8   0.25  2     1     1     >=8   0.125 0.5   >=8   0.125 0.25 
#> [25] 0.25 

random_disk(25, "Klebsiella pneumoniae") # range 8-50
#> Class 'disk'
#>  [1] 45 49 24 46 23 18 15 17 17 28 40 20 21 43 39  9 33 17  8 20 30 21 44 48 41
random_disk(25, "Klebsiella pneumoniae", "ampicillin") # range 11-17
#> Class 'disk'
#>  [1] 13 15 11 14 16 16 11 15 11 17 14 15 17 17 16 16 16 12 14 12 15 13 13 13 16
random_disk(25, "Streptococcus pneumoniae", "ampicillin") # range 12-27
#> Class 'disk'
#>  [1] 22 23 19 19 23 15 23 18 21 24 22 25 19 19 16 17 17 16 27 21 27 25 15 21 27
# }