Random MIC Values/Disk Zones/RSI Generation

These functions can be used for generating random MIC values and disk diffusion diameters, for AMR data analysis practice. By providing a microorganism and antimicrobial drug, the generated results will reflect reality as much as possible.

Usage

random_mic(size = NULL, mo = NULL, ab = NULL, ...)

random_disk(size = NULL, mo = NULL, ab = NULL, ...)

random_rsi(size = NULL, prob_RSI = c(0.33, 0.33, 0.33), ...)

Arguments

size

desired size of the returned vector. If used in a data.frame call or dplyr verb, will get the current (group) size if left blank.

mo

any character that can be coerced to a valid microorganism code with as.mo()

ab

any character that can be coerced to a valid antimicrobial drug code with as.ab()

...

ignored, only in place to allow future extensions

prob_RSI

a vector of length 3: the probabilities for "R" (1st value), "S" (2nd value) and "I" (3rd value)

Value

class mic for random_mic() (see as.mic()) and class disk for random_disk() (see as.disk())

Details

The base R function sample() is used for generating values.

Generated values are based on the EUCAST 2022 guideline as implemented in the rsi_translation data set. To create specific generated values per bug or drug, set the mo and/or ab argument.

Examples

random_mic(25)
#> Class 'mic'
#>  [1] 4      4      4      128    8      4      0.0625 0.25   0.5    2     
#> [11] 4      64     0.0625 32     0.25   0.001  0.001  0.025  0.001  >=256 
#> [21] 4      0.025  0.5    4      0.025 
random_disk(25)
#> Class 'disk'
#>  [1] 32 46 50 35 15  7 42 27 21 15  8 36 30 46 47 43 42 50 32 34 21 43 21 28 43
random_rsi(25)
#> Class 'rsi'
#>  [1] R I I S S R S I S R R S I S S I R I R I S S S S R

# \donttest{
# make the random generation more realistic by setting a bug and/or drug:
random_mic(25, "Klebsiella pneumoniae") # range 0.0625-64
#> Class 'mic'
#>  [1] 128     32      0.025   0.25    0.125   0.01    <=0.001 1       0.25   
#> [10] 1       0.005   0.025   0.0625  128     0.5     0.005   0.0625  0.005  
#> [19] <=0.001 128     4       0.5     0.002   0.25    0.01   
random_mic(25, "Klebsiella pneumoniae", "meropenem") # range 0.0625-16
#> Class 'mic'
#>  [1] 0.5  8    8    0.5  0.5  2    4    >=16 0.5  4    8    >=16 0.5  >=16 2   
#> [16] >=16 1    8    1    4    4    >=16 >=16 >=16 0.5 
random_mic(25, "Streptococcus pneumoniae", "meropenem") # range 0.0625-4
#> Class 'mic'
#>  [1] 1      2      0.125  0.5    0.5    0.25   2      0.125  2      1     
#> [11] 0.125  0.125  4      8      0.25   0.125  4      0.5    2      0.0625
#> [21] 4      0.5    1      >=16   1     

random_disk(25, "Klebsiella pneumoniae") # range 8-50
#> Class 'disk'
#>  [1] 11 41 18 50 12 39 48 47 25 23 19 14 14 36 46 21 46 18 17 50 26 20 33 12 44
random_disk(25, "Klebsiella pneumoniae", "ampicillin") # range 11-17
#> Class 'disk'
#>  [1] 11 16 14 13 11 17 16 12 14 12 13 13 17 13 15 15 16 13 11 15 11 16 11 16 11
random_disk(25, "Streptococcus pneumoniae", "ampicillin") # range 12-27
#> Class 'disk'
#>  [1] 24 22 18 26 22 20 27 27 20 18 24 26 25 20 20 25 19 19 25 20 18 25 16 20 17
# }