Predict antimicrobial resistance

Create a prediction model to predict antimicrobial resistance for the next years on statistical solid ground. Standard errors (SE) will be returned as columns se_min and se_max. See Examples for a real live example.

resistance_predict(
  x,
  col_ab,
  col_date = NULL,
  year_min = NULL,
  year_max = NULL,
  year_every = 1,
  minimum = 30,
  model = NULL,
  I_as_S = TRUE,
  preserve_measurements = TRUE,
  info = interactive(),
  ...
)

rsi_predict(
  x,
  col_ab,
  col_date = NULL,
  year_min = NULL,
  year_max = NULL,
  year_every = 1,
  minimum = 30,
  model = NULL,
  I_as_S = TRUE,
  preserve_measurements = TRUE,
  info = interactive(),
  ...
)

# S3 method for resistance_predict
plot(x, main = paste("Resistance Prediction of", x_name), ...)

ggplot_rsi_predict(
  x,
  main = paste("Resistance Prediction of", x_name),
  ribbon = TRUE,
  ...
)

Arguments

x
a data.frame containing isolates.

col_ab
column name of x containing antimicrobial interpretations ("R", "I" and "S")

col_date
column name of the date, will be used to calculate years if this column doesn't consist of years already, defaults to the first column of with a date class

year_min
lowest year to use in the prediction model, dafaults to the lowest year in col_date

year_max
highest year to use in the prediction model, defaults to 10 years after today

year_every
unit of sequence between lowest year found in the data and year_max

minimum
minimal amount of available isolates per year to include. Years containing less observations will be estimated by the model.

model
the statistical model of choice. This could be a generalised linear regression model with binomial distribution (i.e. using `glm(..., family = binomial)``, assuming that a period of zero resistance was followed by a period of increasing resistance leading slowly to more and more resistance. See Details for all valid options.

I_as_S
a logical to indicate whether values I should be treated as S (will otherwise be treated as R). The default, TRUE, follows the redefinition by EUCAST about the interpretion of I (increased exposure) in 2019, see section Interpretation of S, I and R below.

preserve_measurements
a logical to indicate whether predictions of years that are actually available in the data should be overwritten by the original data. The standard errors of those years will be NA.

info
a logical to indicate whether textual analysis should be printed with the name and summary() of the statistical model.

...
parameters passed on to functions

main
title of the plot

ribbon
a logical to indicate whether a ribbon should be shown (default) or error bars

Value

A data.frame with extra class resistance_predict with columns:

year

value, the same as estimated when preserve_measurements = FALSE, and a combination of observed and estimated otherwise

se_min, the lower bound of the standard error with a minimum of 0 (so the standard error will never go below 0%)

se_max the upper bound of the standard error with a maximum of 1 (so the standard error will never go above 100%)

observations, the total number of available observations in that year, i.e. \(S + I + R\)

observed, the original observed resistant percentages

estimated, the estimated resistant percentages, calculated by the model

Furthermore, the model itself is available as an attribute: attributes(x)$model, please see Examples.

Details

Valid options for the statistical model (parameter model) are:

"binomial" or "binom" or "logit": a generalised linear regression model with binomial distribution

"loglin" or "poisson": a generalised log-linear regression model with poisson distribution

"lin" or "linear": a linear regression model

Maturing lifecycle

The lifecycle of this function is maturing. The unlying code of a maturing function has been roughed out, but finer details might still change. Since this function needs wider usage and more extensive testing, you are very welcome to suggest changes at our repository or write us an email (see section 'Contact Us').

Interpretation of R and S/I

In 2019, the European Committee on Antimicrobial Susceptibility Testing (EUCAST) has decided to change the definitions of susceptibility testing categories R and S/I as shown below (https://www.eucast.org/newsiandr/).

R = Resistant
A microorganism is categorised as Resistant when there is a high likelihood of therapeutic failure even when there is increased exposure. Exposure is a function of how the mode of administration, dose, dosing interval, infusion time, as well as distribution and excretion of the antimicrobial agent will influence the infecting organism at the site of infection.

S = Susceptible
A microorganism is categorised as Susceptible, standard dosing regimen, when there is a high likelihood of therapeutic success using a standard dosing regimen of the agent.

I = Increased exposure, but still susceptible
A microorganism is categorised as Susceptible, Increased exposure when there is a high likelihood of therapeutic success because exposure to the agent is increased by adjusting the dosing regimen or by its concentration at the site of infection.

This AMR package honours this new insight. Use susceptibility() (equal to proportion_SI()) to determine antimicrobial susceptibility and count_susceptible() (equal to count_SI()) to count susceptible isolates.

Read more on our website!

On our website https://msberends.github.io/AMR/ you can find a comprehensive tutorial about how to conduct AMR analysis, the complete documentation of all functions and an example analysis using WHONET data. As we would like to better understand the backgrounds and needs of our users, please participate in our survey!

See also

Examples

x <- resistance_predict(example_isolates, 
                        col_ab = "AMX",
                        year_min = 2010,
                        model = "binomial")
plot(x)
if (require("ggplot2")) {
  ggplot_rsi_predict(x)
}

# using dplyr:
if (require("dplyr")) {
  x <- example_isolates %>%
    filter_first_isolate() %>%
    filter(mo_genus(mo) == "Staphylococcus") %>%
    resistance_predict("PEN", model = "binomial")
  plot(x)

  # get the model from the object
  mymodel <- attributes(x)$model
  summary(mymodel)
}

# create nice plots with ggplot2 yourself
if (require("dplyr") & require("ggplot2")) {

  data <- example_isolates %>%
    filter(mo == as.mo("E. coli")) %>%
    resistance_predict(col_ab = "AMX",
                       col_date = "date",
                       model = "binomial",
                       info = FALSE,
                       minimum = 15)

  ggplot(data,
         aes(x = year)) +
    geom_col(aes(y = value),
             fill = "grey75") +
    geom_errorbar(aes(ymin = se_min,
                      ymax = se_max),
                  colour = "grey50") +
    scale_y_continuous(limits = c(0, 1),
                       breaks = seq(0, 1, 0.1),
                       labels = paste0(seq(0, 100, 10), "%")) +
    labs(title = expression(paste("Forecast of Amoxicillin Resistance in ",
                                  italic("E. coli"))),
         y = "%R",
         x = "Year") +
    theme_minimal(base_size = 13)
}