Data Set for R/SI Interpretation

Data set containing reference data to interpret MIC and disk diffusion to R/SI values, according to international guidelines. Currently implemented guidelines are EUCAST (2013-2022) and CLSI (2013-2022). Use as.rsi() to transform MICs or disks measurements to R/SI values.

Usage

rsi_translation

Format

A tibble with 18,308 observations and 11 variables:

guideline
Name of the guideline

method
Either "DISK" or "MIC"

site
Body site, e.g. "Oral" or "Respiratory"

mo
Microbial ID, see as.mo()

rank_index
Taxonomic rank index of mo from 1 (subspecies/infraspecies) to 5 (unknown microorganism)

ab
Antibiotic ID, see as.ab()

ref_tbl
Info about where the guideline rule can be found

disk_dose
Dose of the used disk diffusion method

breakpoint_S
Lowest MIC value or highest number of millimetres that leads to "S"

breakpoint_R
Highest MIC value or lowest number of millimetres that leads to "R"

uti
A logical value (TRUE/FALSE) to indicate whether the rule applies to a urinary tract infection (UTI)

Details

Like all data sets in this package, this data set is publicly available for download in the following formats: R, MS Excel, Apache Feather, Apache Parquet, SPSS, SAS, and Stata. Please visit our website for the download links. The actual files are of course available on our GitHub repository.

They allow for machine reading EUCAST and CLSI guidelines, which is almost impossible with the MS Excel and PDF files distributed by EUCAST and CLSI.

See also

Examples

rsi_translation
#> # A tibble: 18,308 × 11
#>    guideline method site  mo               rank_…¹ ab   ref_tbl  disk_…² break…³
#>    <chr>     <chr>  <chr> <mo>               <dbl> <ab> <chr>    <chr>     <dbl>
#>  1 EUCAST 2… MIC    NA    F_ASPRG_MGTS           2 AMB  Aspergi… NA            1
#>  2 EUCAST 2… MIC    NA    F_ASPRG_NIGR           2 AMB  Aspergi… NA            1
#>  3 EUCAST 2… MIC    NA    F_CANDD_ALBC           2 AMB  Candida  NA            1
#>  4 EUCAST 2… MIC    NA    F_CANDD_DBLN           2 AMB  Candida  NA            1
#>  5 EUCAST 2… MIC    NA    F_CANDD_GLBR           2 AMB  Candida  NA            1
#>  6 EUCAST 2… MIC    NA    F_CANDD_KRUS           2 AMB  Candida  NA            1
#>  7 EUCAST 2… MIC    NA    F_CANDD_PRPS           2 AMB  Candida  NA            1
#>  8 EUCAST 2… MIC    NA    F_CANDD_TRPC           2 AMB  Candida  NA            1
#>  9 EUCAST 2… MIC    NA    F_CRYPT_NFRM           2 AMB  Candida  NA            1
#> 10 EUCAST 2… DISK   NA    B_[ORD]_ENTRBCTR       5 AMC  Enterob… 20ug/1…      19
#> # … with 18,298 more rows, 2 more variables: breakpoint_R <dbl>, uti <lgl>, and
#> #   abbreviated variable names ¹​rank_index, ²​disk_dose, ³​breakpoint_S