Data set with 2000 blood culture isolates of septic patients

An anonymised data set containing 2,000 microbial blood culture isolates with their full antibiograms found in septic patients in 4 different hospitals in the Netherlands, between 2001 and 2017. It is true, genuine data. This data.frame can be used to practice AMR analysis. For examples, please read the tutorial on our website.

septic_patients

Format

A data.frame with 2,000 observations and 49 variables:

date

date of receipt at the laboratory

hospital_id

ID of the hospital, from A to D

ward_icu

logical to determine if ward is an intensive care unit

ward_clinical

logical to determine if ward is a regular clinical ward

ward_outpatient

logical to determine if ward is an outpatient clinic

age

age of the patient

gender

gender of the patient

patient_id

ID of the patient, first 10 characters of an SHA hash containing irretrievable information

mo

ID of microorganism created with as.mo, see also microorganisms

peni:rifa

40 different antibiotics with class rsi (see as.rsi); these column names occur in antibiotics data set and can be translated with abname

Read more on our website!

On our website https://msberends.gitlab.io/AMR you can find a omprehensive tutorial about how to conduct AMR analysis and find the complete documentation of all functions, which reads a lot easier than in R.